% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anderson.darling.normality.test.R
\name{anderson.darling.normality.test}
\alias{anderson.darling.normality.test}
\title{Anderson-Darling Normality Test}
\usage{
anderson.darling.normality.test(x)
}
\arguments{
\item{x}{A vector of values to be tested against the normal distribution}
}
\value{
An htest object including results from Anderson Darling normality test.
}
\description{
Performs the Anderson Darling normality test on a vector of data. 
Null hypothesis is that normal distribution can beused to model sample. 
Low p-value (rejection) indicates non-normality.
}
