% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.independent.kruskal.wallis.R
\name{anova.independent.kruskal.wallis}
\alias{anova.independent.kruskal.wallis}
\title{Kruskal-Wallis One-Way Analysis of Variance By Ranks}
\usage{
\method{anova}{independent.kruskal.wallis}(
  fx,
  data = NULL,
  tie.correct = T,
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{fx}{A formula defining groups and a dependent variable}

\item{data}{A data frame that corresponds to the formulas in fx.}

\item{tie.correct}{Tie correction (T/F)}

\item{alternative}{Alternative hypothesis to be tested}

\item{conf.level}{Confidence level for test}
}
\value{
htest object containing results of the test.
}
\description{
Perform Kruskal-Wallis One-Way Analysis of Variance By Ranks (sometimes called Kruskal-Wallis H test).
}
