% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.mean.R
\name{center.mean}
\alias{center.mean}
\title{Center Data Using Arithmetic Mean}
\usage{
center.mean(x)
}
\arguments{
\item{x}{A vector of values to be centered using the mean value}
}
\value{
A vector containing centered data (mean of zero).
}
\description{
Subtract a measure of central tendency from data causing data's center to become zero.
}
