% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.median.R
\name{center.median}
\alias{center.median}
\title{Center Data Using Median}
\usage{
center.median(x)
}
\arguments{
\item{x}{A vector of values to be centered using the median value}
}
\value{
A vector containing centered data (median of zero).
}
\description{
Subtract a measure of central tendency from data causing data's center to become zero.
}
