% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Center Data}
\usage{
center(x, center.function = mean, center.function.na.rm = T)
}
\arguments{
\item{x}{A vector of values to be centered}

\item{center.function}{A function to use as a measure of center}

\item{center.function.na.rm}{logial (T/F) identifying whether NA should be dropped from input}
}
\value{
A vector containing centered data (central tendency of zero).
}
\description{
Subtract a measure of central tendency from data causing data's center to become zero.
}
