% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.square.2d.expected.frequencies.R
\name{chi.square.2d.expected.frequencies}
\alias{chi.square.2d.expected.frequencies}
\title{Expected Cell Frequencies}
\usage{
chi.square.2d.expected.frequencies(observed.frequencies)
}
\arguments{
\item{observed.frequencies}{A matrix of values to use to calculate expected cell frequencies.}
}
\value{
A matrix with expected cell frequencies matching the dimensions of observed.frequencies.
}
\description{
Calculates expected cell frequencies for a given matrix.
}
