% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.square.goodness.of.fit.simple.R
\name{chi.square.goodness.of.fit.simple}
\alias{chi.square.goodness.of.fit.simple}
\title{Chi Square Goodness of Fit Test}
\usage{
chi.square.goodness.of.fit.simple(
  observed.frequencies,
  expected.frequencies = chi.square.2d.expected.frequencies(observed.frequencies),
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{observed.frequencies}{A matrix of values to test.}

\item{expected.frequencies}{A matrix of values to compare with observed.frequencies.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}
}
\value{
The results of the statistical test.
}
\description{
Calculates chi square goodness of fit test for given contingency table.
}
