% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.group.cell.codes.R
\name{compute.group.cell.codes}
\alias{compute.group.cell.codes}
\title{Calculate Unique Subgroup Codes}
\usage{
compute.group.cell.codes(fx, data = NULL, start.at = 1)
}
\arguments{
\item{fx}{Formula - defines formula for dependent variable and subgroups ( dv ~ v1 + v2 + ...)}

\item{data}{Data Frame - data frame with columns corresponding to variables in fx.}

\item{start.at}{Scalar - A number to start numbering at.}
}
\value{
A vector the same length as rows in data containing the results the subgroup code.
}
\description{
Calculate subgroup codes. Useful for input into other grouping functions and for performing specific tests (ex. one-way ANOVA on multivariate data).
}
