% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.group.dispersion.ADA.R
\name{compute.group.dispersion.ADA}
\alias{compute.group.dispersion.ADA}
\title{Calculate Subgroup Dispersion - Mean Absolute Deviation}
\usage{
compute.group.dispersion.ADA(fx, data = NULL)
}
\arguments{
\item{fx}{Formula - defines formula for dependent variable and subgroups ( dv ~ v1 + v2 + ...)}

\item{data}{Data Frame - data frame with columns corresponding to variables in fx.}
}
\value{
A vector containing absolute deviations from subgroup means.
}
\description{
Calculate subgroup dispersion using absolute deviation from subgroup average.
}
