% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.group.dispersion.ADM.R
\name{compute.group.dispersion.ADM}
\alias{compute.group.dispersion.ADM}
\title{Calculate Subgroup Dispersion - Median Absolute Deviation}
\usage{
compute.group.dispersion.ADM(fx, data = NULL)
}
\arguments{
\item{fx}{Formula - defines formula for dependent variable and subgroups ( dv ~ v1 + v2 + ...)}

\item{data}{Data Frame - data frame with columns corresponding to variables in fx.}
}
\value{
A vector containing absolute deviations from subgroup medians.
}
\description{
Calculate subgroup dispersion using absolute deviation from subgroup median.
}
