% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.group.variable.R
\name{compute.group.variable}
\alias{compute.group.variable}
\title{Calculate Subgroup Variable}
\usage{
compute.group.variable(fx, data = NULL, FUN = function(x) {
    x
})
}
\arguments{
\item{fx}{Formula - defines formula for dependent variable and subgroups ( dv ~ v1 + v2 + ...)}

\item{data}{Data Frame - data frame with columns corresponding to variables in fx.}

\item{FUN}{Function - A function(x) where x is subgroup values that returns a vector that is the same length as x.}
}
\value{
A vector the same length as rows in data containing the results of the subgroup calculation.
}
\description{
Calculate subgroup variable using an arbitrary function.
}
