% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.accuracy.R
\name{confusion.matrix.accuracy}
\alias{confusion.matrix.accuracy}
\title{Confusion Matrix Measures - Accuracy}
\usage{
confusion.matrix.accuracy(confusion.matrix)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}
}
\value{
A scalar with computed accuracy value.
}
\description{
Calculate Accuracy, defined as (TP+TN)/(P+N), for a given confusion matrix.
}
