% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.accuracy.R,
%   R/confusion.matrix.accuracy.simple.R
\name{confusion.matrix.accuracy}
\alias{confusion.matrix.accuracy}
\alias{confusion.matrix.accuracy.simple}
\title{Confusion Matrix Measures - Accuracy}
\usage{
confusion.matrix.accuracy(confusion.matrix)

confusion.matrix.accuracy.simple(
  true.positive = 0,
  false.positive = 0,
  true.negative = 1,
  false.negative = 0,
  count.positive = true.positive + false.negative,
  count.negative = true.negative + false.positive
)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{true.positive}{Scalar - Cases identified as true positive}

\item{false.positive}{Scalar - Cases identified as false positive - optional if count.positive is used}

\item{true.negative}{Scalar - Cases identified as true negative}

\item{false.negative}{Scalar - Cases identified as false negative - optional if count.negative is used}

\item{count.positive}{Scalar - Total cases identified as positive - optional if first four parameters are used.}

\item{count.negative}{Scalar - Total cases identified as negative - optional if first four parameters are used.}
}
\value{
A scalar with computed accuracy value.
}
\description{
Calculate Accuracy, defined as (TP+TN)/(P+N), for a given confusion matrix.
}
