% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.bookmaker.informedness.R,
%   R/confusion.matrix.bookmaker.informedness.simple.R
\name{confusion.matrix.bookmaker.informedness}
\alias{confusion.matrix.bookmaker.informedness}
\alias{confusion.matrix.bookmaker.informedness.simple}
\title{Confusion Matrix Measures - Bookmaker Informedness}
\usage{
confusion.matrix.bookmaker.informedness(confusion.matrix)

confusion.matrix.bookmaker.informedness.simple(
  true.positive = 0,
  false.negative = 1,
  count.positive = true.positive + false.negative,
  false.positive = 0,
  true.negative = 1,
  count.negative = false.positive + true.negative,
  true.positive.rate = confusion.matrix.true.positive.rate.simple(true.positive =
    true.positive, count.positive = count.positive),
  true.negative.rate = confusion.matrix.true.negative.rate.simple(true.negative =
    true.negative, count.negative = count.negative)
)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{true.positive}{Scalar - Cases identified as true positive}

\item{false.negative}{Scalar - Cases identified as false negative}

\item{count.positive}{Scalar - Total cases identified as positive - optional if first four parameters are used.}

\item{false.positive}{Scalar - Cases identified as false positive}

\item{true.negative}{Scalar - Cases identified as true negative}

\item{count.negative}{Scalar - Total cases identified as negative - optional if first four parameters are used.}

\item{true.positive.rate}{Scalar - True positive rate, overrides other parameters used to calculate TPR.}

\item{true.negative.rate}{Scalar - True negative rate, overrides other parameters used to calculate TNR.}
}
\value{
A scalar with computed value.
}
\description{
Calculate Bookmaker Informedness, defined as TPR+TNR-1, for a given confusion matrix.
}
