% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.f1.score.R,
%   R/confusion.matrix.f1.score.simple.R
\name{confusion.matrix.f1.score}
\alias{confusion.matrix.f1.score}
\alias{confusion.matrix.f1.score.simple}
\title{Confusion Matrix Measures - F1 Score}
\usage{
confusion.matrix.f1.score(confusion.matrix)

confusion.matrix.f1.score.simple(
  true.positive = 1,
  false.positive = 0,
  false.negative = 0
)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{true.positive}{Scalar - Cases identified as true positive}

\item{false.positive}{Scalar - Cases identified as false positive}

\item{false.negative}{Scalar - Cases identified as false negative}
}
\value{
A scalar with computed value.
}
\description{
Calculate F1 Score, defined as 2*TP / (2*TP+FP+FN), for a given confusion matrix.
}
