% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.false.discovery.rate.R,
%   R/confusion.matrix.false.discovery.rate.simple.R
\name{confusion.matrix.false.discovery.rate}
\alias{confusion.matrix.false.discovery.rate}
\alias{confusion.matrix.false.discovery.rate.simple}
\title{Confusion Matrix Measures - False Discovery Rate}
\usage{
confusion.matrix.false.discovery.rate(confusion.matrix)

confusion.matrix.false.discovery.rate.simple(
  false.positive = 0,
  true.positive = 1
)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{false.positive}{Scalar - Cases identified as false positive}

\item{true.positive}{Scalar - Cases identified as true positive}
}
\value{
A scalar with computed value.
}
\description{
Calculate False Discovery Rate, defined as FP / (FP + TP), for a given confusion matrix.
}
