% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.false.negative.rate.R,
%   R/confusion.matrix.false.negative.rate.simple.R
\name{confusion.matrix.false.negative.rate}
\alias{confusion.matrix.false.negative.rate}
\alias{confusion.matrix.miss.rate}
\alias{confusion.matrix.false.negative.rate.simple}
\alias{confusion.matrix.miss.rate.simple}
\title{Confusion Matrix Measures - False Negative/Miss Rate}
\usage{
confusion.matrix.false.negative.rate(confusion.matrix)

confusion.matrix.miss.rate(confusion.matrix)

confusion.matrix.false.negative.rate.simple(
  false.negative = 0,
  true.positive = 1,
  count.positive = false.negative + true.positive
)

confusion.matrix.miss.rate.simple(
  false.negative = 0,
  true.positive = 1,
  count.positive = false.negative + true.positive
)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{false.negative}{Scalar - Cases identified as false negative}

\item{true.positive}{Scalar - Cases identified as true positive}

\item{count.positive}{Scalar - Total cases identified as positive - optional if first two parameters are used.}
}
\value{
A scalar with computed value.
}
\description{
Calculate False Negative Rate (also called Miss Rate), defined as FN/P, for a given confusion matrix.
}
