% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.false.negative.R
\name{confusion.matrix.false.negative}
\alias{confusion.matrix.false.negative}
\title{Confusion Matrix Measures - False Negative}
\usage{
confusion.matrix.false.negative(confusion.matrix)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}
}
\value{
A scalar with extracted value.
}
\description{
Extract false negative for a given confusion matrix.
}
