% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.false.omission.rate.simple.R
\name{confusion.matrix.false.omission.rate.simple}
\alias{confusion.matrix.false.omission.rate.simple}
\alias{confusion.matrix.false.omission.rate}
\title{Confusion Matrix Measures - False Omission Rate}
\usage{
confusion.matrix.false.omission.rate.simple(
  false.negative = 0,
  true.negative = 1
)
}
\arguments{
\item{false.negative}{Scalar - Cases identified as false negative}

\item{true.negative}{Scalar - Cases identified as true negative}

\item{confusion.matrix}{Matrix - confusion matrix.}
}
\value{
A scalar with computed value.
}
\description{
Calculate False Omission Rate, defined as FN/(FN+TN), for a given confusion matrix.
}
