% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.false.positive.rate.R,
%   R/confusion.matrix.false.positive.rate.simple.R
\name{confusion.matrix.false.positive.rate}
\alias{confusion.matrix.false.positive.rate}
\alias{confusion.matrix.fall.out}
\alias{confusion.matrix.false.positive.rate.simple}
\alias{confusion.matrix.fall.out.simple}
\title{Confusion Matrix Measures - False Positive/Fallout Rate}
\usage{
confusion.matrix.false.positive.rate(confusion.matrix)

confusion.matrix.fall.out(confusion.matrix)

confusion.matrix.false.positive.rate.simple(
  false.positive = 0,
  true.negative = 1,
  count.negative = false.positive + true.negative
)

confusion.matrix.fall.out.simple(
  false.positive = 0,
  true.negative = 1,
  count.negative = false.positive + true.negative
)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{false.positive}{Scalar - Cases identified as false positive}

\item{true.negative}{Scalar - Cases identified as true negative}

\item{count.negative}{Scalar - Total cases identified as negative - optional if first two parameters are used.}
}
\value{
A scalar with computed value.
}
\description{
Calculate False Positive Rate (sometimes called Fallout rate), defined as FP/N, for a given confusion matrix.
}
