% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.false.positive.R
\name{confusion.matrix.false.positive}
\alias{confusion.matrix.false.positive}
\title{Confusion Matrix Measures - False Positive}
\usage{
confusion.matrix.false.positive(confusion.matrix)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}
}
\value{
A scalar with extracted value.
}
\description{
Extract false positive for a given confusion matrix.
}
