% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.fn.multiple.values.R
\name{confusion.matrix.fn.multiple.values}
\alias{confusion.matrix.fn.multiple.values}
\title{Confusion Matrix Generation - Multiple Values Function}
\usage{
confusion.matrix.fn.multiple.values(values)
}
\arguments{
\item{values}{Vector - Values to compare actual and predicted with.}
}
\value{
A function for use with a call to confusion.matrix().
}
\description{
Returns a function for use with confusion.matrix based on actual or predicted being in a set of values.
}
