% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.fn.threshold.R
\name{confusion.matrix.fn.threshold}
\alias{confusion.matrix.fn.threshold}
\title{Confusion Matrix Generation - Threshold Function}
\usage{
confusion.matrix.fn.threshold(threshold)
}
\arguments{
\item{threshold}{Scalar - Value to compare actual and predicted with.}
}
\value{
A function for use with a call to confusion.matrix().
}
\description{
Returns a function for use with confusion.matrix based on actual or predicted being larger than a particular value.
}
