% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.fn.value.R
\name{confusion.matrix.fn.value}
\alias{confusion.matrix.fn.value}
\title{Confusion Matrix Generation - Individual Value Function}
\usage{
confusion.matrix.fn.value(value)
}
\arguments{
\item{value}{Scalar - Value to compare actual and predicted with.}
}
\value{
A function for use with a call to confusion.matrix().
}
\description{
Returns a function for use with confusion.matrix based on actual or predicted being a particular value.
}
