% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.markedness.R,
%   R/confusion.matrix.markedness.simple.R
\name{confusion.matrix.markedness}
\alias{confusion.matrix.markedness}
\alias{confusion.matrix.markedness.simple}
\title{Confusion Matrix Measures - Markedness}
\usage{
confusion.matrix.markedness(confusion.matrix)

confusion.matrix.markedness.simple(
  true.positive = 0,
  false.positive = 1,
  true.negative = 0,
  false.negative = 1,
  positive.predictive.value =
    confusion.matrix.positive.predictive.value.simple(true.positive = true.positive,
    false.positive = false.positive),
  negative.predictive.value =
    confusion.matrix.negative.predictive.value.simple(true.negative = true.negative,
    false.negative = false.negative)
)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{true.positive}{Scalar - Cases identified as true positive - optional if PPV specified}

\item{false.positive}{Scalar - Cases identified as false positive - optional if PPV specified}

\item{true.negative}{Scalar - Cases identified as true negative - optional if NPV specified}

\item{false.negative}{Scalar - Cases identified as false negative - optional if NPV specified}

\item{positive.predictive.value}{Scalar - Positive Predictive Value (PPV) - optional if true.positive and false.positive parameters are used.}

\item{negative.predictive.value}{Scalar - Negative Predictive Value (NPV) - optional if true.negative and false.negative parameters are used.}
}
\value{
A scalar with computed value.
}
\description{
Calculate Markedness, defined as PPV+NPV-1, for a given confusion matrix.
}
