% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/confusion.matrix.matthews.correlation.coefficient.R,
%   R/confusion.matrix.matthews.correlation.coefficient.simple.R
\name{confusion.matrix.matthews.correlation.coefficient}
\alias{confusion.matrix.matthews.correlation.coefficient}
\alias{cor.matthews}
\alias{confusion.matrix.matthews.correlation.coefficient.simple}
\alias{cor.matthews.simple}
\title{Confusion Matrix Measures - Matthews' Correlation Coefficient}
\usage{
confusion.matrix.matthews.correlation.coefficient(confusion.matrix)

cor.matthews(confusion.matrix)

confusion.matrix.matthews.correlation.coefficient.simple(
  true.positive = 1,
  false.positive = 1,
  true.negative = 1,
  false.negative = 1
)

cor.matthews.simple(
  true.positive = 1,
  false.positive = 1,
  true.negative = 1,
  false.negative = 1
)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{true.positive}{Scalar - Cases identified as true positive}

\item{false.positive}{Scalar - Cases identified as false positive}

\item{true.negative}{Scalar - Cases identified as true negative}

\item{false.negative}{Scalar - Cases identified as false negative}
}
\value{
A scalar with computed value.
}
\description{
Calculate Matthews' Correlation Coefficient, defined as (TP*TN - FP*FN)/sqrt((TP+FP)(TP+FN)(TN+FP)(TN+FN)), for a given confusion matrix.
}
