% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.negative.predictive.value.R,
%   R/confusion.matrix.negative.predictive.value.simple.R
\name{confusion.matrix.negative.predictive.value}
\alias{confusion.matrix.negative.predictive.value}
\alias{confusion.matrix.negative.predictive.value.simple}
\title{Confusion Matrix Measures - Negative Predictive Value}
\usage{
confusion.matrix.negative.predictive.value(confusion.matrix)

confusion.matrix.negative.predictive.value.simple(
  true.negative = 0,
  false.negative = 1
)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{true.negative}{Scalar - Cases identified as true negative}

\item{false.negative}{Scalar - Cases identified as false negative}
}
\value{
A scalar with computed value.
}
\description{
Calculate Negative Predictive Value, defined as TN/(TN+FN), for a given confusion matrix.
}
