% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.positive.likelihood.ratio.R,
%   R/confusion.matrix.positive.likelihood.ratio.simple.R
\name{confusion.matrix.positive.likelihood.ratio}
\alias{confusion.matrix.positive.likelihood.ratio}
\alias{confusion.matrix.positive.likelihood.ratio.simple}
\title{Confusion Matrix Measures - Positive Likelihood Ratio}
\usage{
confusion.matrix.positive.likelihood.ratio(confusion.matrix)

confusion.matrix.positive.likelihood.ratio.simple(
  true.positive = 0,
  false.negative = 1,
  count.positive = true.positive + false.negative,
  false.positive = 0,
  true.negative = 1,
  count.negative = false.positive + true.negative,
  true.positive.rate = confusion.matrix.true.positive.rate.simple(true.positive =
    true.positive, count.positive = count.positive),
  false.positive.rate = confusion.matrix.false.positive.rate.simple(false.positive =
    false.positive, count.negative = count.negative)
)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{true.positive}{Scalar - Cases identified as true positive - optional if TPR specified}

\item{false.negative}{Scalar - Cases identified as false negative - optional if TPR specified}

\item{count.positive}{Scalar - Total cases identified as positive - optional if TPR specified.}

\item{false.positive}{Scalar - Cases identified as false positive - optional if FPR specified}

\item{true.negative}{Scalar - Cases identified as true negative - optional if FPR specified}

\item{count.negative}{Scalar - Total cases identified as negative - optional if FPR specified.}

\item{true.positive.rate}{Scalar - True Positive Rate (TPR) - optional if TPR parameters specified.}

\item{false.positive.rate}{Scalar - False Positive Rate (FPR) - optional if FPR parameters specified.}
}
\value{
A scalar with computed value.
}
\description{
Calculate Positive Likelihood Ratio, defined as TPR/FPR, for a given confusion matrix.
}
