% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.positive.predictive.value.R,
%   R/confusion.matrix.positive.predictive.value.simple.R
\name{confusion.matrix.positive.predictive.value}
\alias{confusion.matrix.positive.predictive.value}
\alias{confusion.matrix.precision}
\alias{confusion.matrix.positive.predictive.value.simple}
\alias{confusion.matrix.precision.simple}
\title{Confusion Matrix Measures - Positive Predictive Value/Precision}
\usage{
confusion.matrix.positive.predictive.value(confusion.matrix)

confusion.matrix.precision(confusion.matrix)

confusion.matrix.positive.predictive.value.simple(
  true.positive = 0,
  false.positive = 1
)

confusion.matrix.precision.simple(true.positive = 0, false.positive = 1)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{true.positive}{Scalar - Cases identified as true positive}

\item{false.positive}{Scalar - Cases identified as false positive}
}
\value{
A scalar with computed value.
}
\description{
Calculate Positive Predictive Value (also called Precision), defined as TP/(TP+FP), for a given confusion matrix.
}
