% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.total.negative.R
\name{confusion.matrix.total.negative}
\alias{confusion.matrix.total.negative}
\title{Confusion Matrix Measures - Total Negative Cases}
\usage{
confusion.matrix.total.negative(confusion.matrix)
}
\arguments{
\item{confusion.matrix}{Matrix - Confusion matrix}
}
\value{
A scalar with computed value.
}
\description{
Calculate Total Negative cases, defined as TN+FP, for a given confusion matrix.
}
