% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.total.negative.simple.R
\name{confusion.matrix.total.negative.simple}
\alias{confusion.matrix.total.negative.simple}
\title{Confusion Matrix Measures - Total Negative Cases}
\usage{
confusion.matrix.total.negative.simple(true.negative, false.positive)
}
\arguments{
\item{true.negative}{Scalar - Cases identified as true negative}

\item{false.positive}{Scalar - Cases identified as false positive}

\item{confusion.matrix}{Matrix - confusion matrix.}
}
\value{
A scalar with computed value.
}
\description{
Calculate Total Negative cases, defined as TN+FP, for a given confusion matrix.
}
