% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.total.positive.R
\name{confusion.matrix.total.positive}
\alias{confusion.matrix.total.positive}
\title{Confusion Matrix Measures - Total Positive Cases}
\usage{
confusion.matrix.total.positive(confusion.matrix)
}
\arguments{
\item{confusion.matrix}{Matrix - Confusion matrix}
}
\value{
A scalar with computed value.
}
\description{
Calculate Total Positive Cases, defined as FP+TN, for a given confusion matrix.
}
