% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.total.positive.simple.R
\name{confusion.matrix.total.positive.simple}
\alias{confusion.matrix.total.positive.simple}
\title{Confusion Matrix Measures - Total Positive Cases}
\usage{
confusion.matrix.total.positive.simple(false.positive, true.negative)
}
\arguments{
\item{false.positive}{Scalar - Cases identified as false positive}

\item{true.negative}{Scalar - Cases identified as true negative}

\item{confusion.matrix}{Matrix - confusion matrix.}
}
\value{
A scalar with computed value.
}
\description{
Calculate Total Positive Cases, defined as FP+TN, for a given confusion matrix.
}
