% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.true.negative.rate.R,
%   R/confusion.matrix.true.negative.rate.simple.R
\name{confusion.matrix.true.negative.rate}
\alias{confusion.matrix.true.negative.rate}
\alias{confusion.matrix.specificity}
\alias{confusion.matrix.true.negative.rate.simple}
\alias{confusion.matrix.specificity.simple}
\title{Confusion Matrix Measures - True Negative Rate/Specificity}
\usage{
confusion.matrix.true.negative.rate(confusion.matrix)

confusion.matrix.specificity(confusion.matrix)

confusion.matrix.true.negative.rate.simple(
  true.negative = 0,
  false.positive = 1,
  count.negative = true.negative + false.positive
)

confusion.matrix.specificity.simple(
  true.negative = 0,
  false.positive = 1,
  count.negative = true.negative + false.positive
)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{false.positive}{Scalar - Cases identified as false positive - optional if count.negative specified}

\item{count.negative}{Scalar - Total cases identified as negative - optional if FP specified.}

\item{true.positive}{Scalar - Cases identified as true positive}
}
\value{
A scalar with computed value.
}
\description{
Calculate True Negative Rate (also called Specificity), defined as TN/N, for a given confusion matrix.
}
