% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.true.positive.rate.R,
%   R/confusion.matrix.true.positive.rate.simple.R
\name{confusion.matrix.true.positive.rate}
\alias{confusion.matrix.true.positive.rate}
\alias{confusion.matrix.sensitivity}
\alias{confusion.matrix.recall}
\alias{confusion.matrix.hit.rate}
\alias{confusion.matrix.true.positive.rate.simple}
\alias{confusion.matrix.sensitivity.simple}
\alias{confusion.matrix.recall.simple}
\alias{confusion.matrix.hit.rate.simple}
\title{Confusion Matrix Measures - True Positive Rate}
\usage{
confusion.matrix.true.positive.rate(confusion.matrix)

confusion.matrix.sensitivity(confusion.matrix)

confusion.matrix.recall(confusion.matrix)

confusion.matrix.hit.rate(confusion.matrix)

confusion.matrix.true.positive.rate.simple(
  true.positive = 0,
  false.negative = 1,
  count.positive = true.positive + false.negative
)

confusion.matrix.sensitivity.simple(
  true.positive = 0,
  false.negative = 1,
  count.positive = true.positive + false.negative
)

confusion.matrix.recall.simple(
  true.positive = 0,
  false.negative = 1,
  count.positive = true.positive + false.negative
)

confusion.matrix.hit.rate.simple(
  true.positive = 0,
  false.negative = 1,
  count.positive = true.positive + false.negative
)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}

\item{true.positive}{Scalar - Cases identified as true positive}

\item{false.negative}{Scalar - Cases identified as false negative - optional if count.positive specified}

\item{count.positive}{Scalar - Total cases identified as positive - optional if false.negative specified.}
}
\value{
A scalar with computed value.
}
\description{
Calculate True Positive Rate, defined as TP/P, for a given confusion matrix. True positive rate is also sometimes called sensitivity, recall, and hit rate.
}
