% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.true.positive.R
\name{confusion.matrix.true.positive}
\alias{confusion.matrix.true.positive}
\title{Confusion Matrix Measures - True Positive}
\usage{
confusion.matrix.true.positive(confusion.matrix)
}
\arguments{
\item{confusion.matrix}{Matrix - confusion matrix.}
}
\value{
A scalar with extracted value.
}
\description{
Extract true positive for a given confusion matrix.
}
