% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.e.R
\docType{data}
\name{constants.e}
\alias{constants.e}
\alias{e}
\title{Constant - e - Euler's number}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
constants.e

e
}
\description{
Euler's number is the base for natural logarithms, among many other uses and appearances in mathematics.
}
\keyword{datasets}
