% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.g.R
\docType{data}
\name{constants.g}
\alias{constants.g}
\title{Constant - g - Acceleration Free Fall Constant (Earth)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
constants.g
}
\description{
The free fall acceleration constant in units of meters per seconds squared (m/s^2).
}
\references{
Thompson, Ambler and Taylor, Barry. NIST Special Publication 811 - Guide for theUse of the International System of Units (SI). 2008.  
https://nvlpubs.nist.gov/nistpubs/Legacy/SP/nistspecialpublication811e2008.pdf
}
\keyword{datasets}
