% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.pi.R
\docType{data}
\name{constants.pi}
\alias{constants.pi}
\title{Constant - pi - Archimedes' Constant}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
constants.pi
}
\description{
The ratio of the circumference and diameter of a circle.
This formulation of pi is based on the formulation used by NASA/JPL. 
It provides more digits of resolution than R's built-in version (6 digits after decimal point vs 15 digits).
}
\examples{
# options(digits=22)
# pi #Default R version
##[1] 3.141593
# constants.pi #NASA/JPL Version
##[1] 3.1415926535897931

}
\references{
NASA/JPL Edu. How Many Decimals of Pi Do We Really Need?. March 16, 2016. Accessed March 15, 2022. https://www.jpl.nasa.gov/edu/news/2016/3/16/how-many-decimals-of-pi-do-we-really-need/
}
\keyword{datasets}
