% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.kurtosis.R
\name{convert.kurtosis}
\alias{convert.kurtosis}
\title{Convert Kurtosis}
\usage{
convert.kurtosis(
  x,
  sample.size,
  from = c("fisher", "pearson"),
  to = c("pearson", "fisher")
)
}
\arguments{
\item{x}{Scalar/Vector - One or more kurtosis values to convert.}

\item{sample.size}{Scalar/Vector - One or more sample sizes used in calculation of kurtosis.}

\item{from}{Scalar/Character - Identify which kurtosis calculation is used for input.}

\item{to}{Scalar/Character - Identify which kurtosis calculation is used for output.}
}
\value{
Scalar/Vector - Converted kurtosis values.
}
\description{
Convert between Fisher's and Pearson's calculations for kurtosis.
}
