% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.skewness.R
\name{convert.skewness}
\alias{convert.skewness}
\title{Convert Skewness}
\usage{
convert.skewness(
  x,
  sample.size,
  from = c("fisher", "pearson"),
  to = c("pearson", "fisher")
)
}
\arguments{
\item{x}{Scalar/Vector - One or more skewness values to convert.}

\item{sample.size}{Scalar/Vector - One or more sample sizes used in calculation of skewness.}

\item{from}{Scalar/Character - Identify which skewness calculation is used for input.}

\item{to}{Scalar/Character - Identify which skewness calculation is used for output.}
}
\value{
Scalar/Vector - Converted skewness values.
}
\description{
Convert between Fisher's and Pearson's calculations for skewness.
}
