% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.biserial.R
\name{cor.biserial}
\alias{cor.biserial}
\title{Biserial Correlation Coefficient}
\usage{
cor.biserial(
  discrete_var,
  continuous_var,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{discrete_var}{A vector of values corresponding to the discrete/nominal variable}

\item{continuous_var}{A vector of values corresponding to the continuous variable}

\item{alternative}{Alternative hypothesis directionality}

\item{conf.level}{Confidence level for calculating confidence intervals}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate the biserial correlation coefficient, a correlation between a discrete/nominal variable and a continuous measure.
}
