% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.bowker.symmetry.1948.R
\name{cor.bowker.symmetry.1948}
\alias{cor.bowker.symmetry.1948}
\title{Bowker's Symmetry Test}
\usage{
cor.bowker.symmetry.1948(
  observed.frequencies,
  alternative = c("greater", "two.sided", "less")
)
}
\arguments{
\item{observed.frequencies}{A matrix of observed frequency values to test.}

\item{alternative}{The alternative hypothesis to use for the test computation.}
}
\value{
The results of the statistical test.
}
\description{
Calculates Bowker's Symmetry Test test for given contingency table.
}
