% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.cohen.kappa.onesample.1960.cohen.R
\name{cor.cohen.kappa.onesample.1960.cohen}
\alias{cor.cohen.kappa.onesample.1960.cohen}
\title{Cohen's Kappa (1960)}
\usage{
cor.cohen.kappa.onesample.1960.cohen(
  observed.frequencies,
  expected.frequencies = chi.square.2d.expected.frequencies(observed.frequencies),
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{observed.frequencies}{A matrix of values to test.}

\item{expected.frequencies}{A matrix of values to compare with observed.frequencies.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}
}
\value{
The results of the statistical test.
}
\description{
Calculate Cohen's Kappa.
}
