% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.cohen.kappa.onesample.1971.fleiss.R
\name{cor.cohen.kappa.onesample.1971.fleiss}
\alias{cor.cohen.kappa.onesample.1971.fleiss}
\title{Cohen's Kappa}
\usage{
cor.cohen.kappa.onesample.1971.fleiss(
  subject,
  rater,
  rating,
  weight = rep(1, length(rating)),
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{subject}{A vector identifying subjects.}

\item{rater}{A vector identifying raters. Not used in this calculation.}

\item{rating}{A vector identifying ratings.}

\item{weight}{A vector identifying weights.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}
}
\value{
The results of the statistical test.
}
\description{
Calculate Fleiss' extension of Cohen's Kappa based on Fleiss' 1971 paper.
}
