% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.cohen.kappa.twosample.independent.simple.R
\name{cor.cohen.kappa.twosample.independent.simple}
\alias{cor.cohen.kappa.twosample.independent.simple}
\title{Two-Sample Test for Differences in Cohen's Kappa}
\usage{
cor.cohen.kappa.twosample.independent.simple(
  kappa.g1,
  se.kappa.g1,
  kappa.g2,
  se.kappa.g2,
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{kappa.g1}{Scalar - Kappa from the first calculation.}

\item{se.kappa.g1}{Scalar - Standard error from first calculation of kappa.}

\item{kappa.g2}{Scalar - Kappa from the second calculation.}

\item{se.kappa.g2}{Scalar - Standard error from second calculation of kappa.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}
}
\value{
The results of the statistical test.
}
\description{
Calculate test to determine if two calculations of Cohen's kappa differ statistically.
}
