% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cor.cohen.kappa.weighted.onesample.1969.fleiss.R
\name{cor.cohen.kappa.weighted.onesample.1969.fleiss}
\alias{cor.cohen.kappa.weighted.onesample.1969.fleiss}
\alias{cor.cohen.kappa.onesample}
\title{Cohen's Weighted Kappa (Fleiss 1969)}
\usage{
cor.cohen.kappa.weighted.onesample.1969.fleiss(
  observed.frequencies,
  weight = matrix(rep(1, length(observed.frequencies)), nrow =
    nrow(observed.frequencies)),
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{observed.frequencies}{A matrix of values to test.}

\item{weight}{Matrix - A matrix of weights to use for cells in observed.frequencies, values between 0 and 1.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}
}
\value{
The results of the statistical test.
}
\description{
Calculate Cohen's Kappa based on Fleiss' 1969 paper.
}
