% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.concordance.helper.R
\name{cor.concordance.helper}
\alias{cor.concordance.helper}
\title{Concordance Helper Function}
\usage{
cor.concordance.helper(x1, x2, call.rank = T)
}
\arguments{
\item{x1}{Vector - group 1 - scores}

\item{x2}{Vector - group 2 - scores}

\item{call.rank}{- Scalar - Logical - all rank() on groups if it has not been called prior to running function.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculates concordant and discordant pairs for Kendall's Tau.
}
