% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.cramer.v.R
\name{cor.cramer.phi}
\alias{cor.cramer.phi}
\alias{cor.cramer.v}
\title{Cramer's V/Phi}
\usage{
cor.cramer.phi(
  observed.frequencies,
  expected.frequencies = chi.square.2d.expected.frequencies(observed.frequencies)
)

cor.cramer.v(
  observed.frequencies,
  expected.frequencies = chi.square.2d.expected.frequencies(observed.frequencies)
)
}
\arguments{
\item{observed.frequencies}{A matrix of values to test.}

\item{expected.frequencies}{A matrix of values to compare with observed.frequencies.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate the Cramer's V (sometimes referred to as Cramer's phi).
}
