% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.goodman.kruskal.gamma.onesample.R
\name{cor.goodman.kruskal.gamma.onesample}
\alias{cor.goodman.kruskal.gamma.onesample}
\title{Goodman and Kruskal's Gamma}
\usage{
cor.goodman.kruskal.gamma.onesample(
  x,
  null.hypothesis.gamma = 0,
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{x}{A matrix that is an ordered contingency table.}

\item{null.hypothesis.gamma}{Scalar - null hypothesis gamma.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate Goodman and Kruskal's Gamma for a contingency table.
}
