% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cor.goodman.kruskal.gamma.twosample.independent.simple.R
\name{cor.goodman.kruskal.gamma.twosample.independent.simple}
\alias{cor.goodman.kruskal.gamma.twosample.independent.simple}
\title{Two-Sample Goodman and Kruskal's Gamma Test}
\usage{
cor.goodman.kruskal.gamma.twosample.independent.simple(
  gamma.g1,
  se.est.gamma.g1,
  gamma.g2,
  se.est.gamma.g2,
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{gamma.g1}{Scalar - Group 1 gamma.}

\item{se.est.gamma.g1}{Scalar - Group 1 gamma estimated standard error.}

\item{gamma.g2}{Scalar - Group 2 gamma.}

\item{se.est.gamma.g2}{Scalar - Group 2 gamma estimated standard error.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate test for two different independent Goodman and Kruskal's Gamma values.
}
