% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.j.index.onesample.R
\name{cor.j.index.onesample}
\alias{cor.j.index.onesample}
\title{J Index of Predictive Efficiency}
\usage{
cor.j.index.onesample(
  x,
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{x}{A 2x2 matrix}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{alternative}{The alternative hypothesis to use for the test computation.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate J Index of Predictive Efficiency for a contingency table.
}
