% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.j.index.twosample.R
\name{cor.j.index.twosample}
\alias{cor.j.index.twosample}
\title{Two-Sample J Index of Predictive Efficiency}
\usage{
cor.j.index.twosample(
  x1,
  x2,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{x1}{A 2x2 matrix for first sample}

\item{x2}{A 2x2 matrix for second sample}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate test for differences between two samples' J Index of Predictive Efficiency calculations.
}
