% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.kendall.coefficient.concordance.R
\name{cor.kendall.coefficient.concordance}
\alias{cor.kendall.coefficient.concordance}
\title{Kendall's Coefficient of Concordance}
\usage{
cor.kendall.coefficient.concordance(
  x,
  raters.in.column.variable = T,
  call.rank = T,
  tie.correct = T,
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{x}{Matrix - matrix of ratings}

\item{raters.in.column.variable}{Logical - The raters are expected to be in the columns, if this is not the case, then set to FALSE.}

\item{call.rank}{Logical - if true, rank() is called. If FALSE, then values are used as-is}

\item{tie.correct}{Logical - if true, a correction for ties is applied to the calculation.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate Kendall's Coefficient of Concordance (sometimes called Kendall's W).
}
