% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.kendall.tau.R, R/cor.kendall.tau.simple.R
\name{cor.kendall.tau}
\alias{cor.kendall.tau}
\alias{cor.kendall.tau.simple}
\title{Kendall's Tau}
\usage{
cor.kendall.tau(
  x1,
  x2,
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)

cor.kendall.tau.simple(
  count.concordant = 0,
  count.discordant = 0,
  ties.x1 = NA,
  ties.x2 = NA,
  sample.size = 0.5 * (1 + sqrt(8 * (count.concordant + count.discordant) + 1)),
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{x1}{Vector - group 1 - scores}

\item{x2}{Vector - group 2 - scores}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{count.concordant}{Scalar - counts of concordant pairs between groups.}

\item{count.discordant}{Scalar - counts of discordant pairs between groups.}

\item{ties.x1}{Vector - group 1 - score counts tied for a particular score.}

\item{ties.x2}{Vector - group 2 - score counts tied for a particular score.}

\item{sample.size}{Scalar - sample size to use for the calculation.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate Kendall's Rank Correlation Coefficient (also called Kendall's Tau).
}
