% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.light.g.onesample.1971.R
\name{cor.light.g.onesample}
\alias{cor.light.g.onesample}
\title{Light's G}
\usage{
cor.light.g.onesample(
  subject,
  rater,
  rating,
  rater.standard = "Standard",
  alternative = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{subject}{A vector identifying subjects.}

\item{rater}{A vector identifying raters.}

\item{rating}{A vector identifying ratings.}

\item{rater.standard}{A scalar identifying the name of the rater considered "correct".}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}
}
\value{
The results of the statistical test.
}
\description{
Calculate Light's G based on Light's 1971 paper. Used to compare multiple raters with standard.
}
