% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.pearson.c.R
\name{cor.pearson.c}
\alias{cor.pearson.c}
\title{Pearson's Contingency Coefficient}
\usage{
cor.pearson.c(x)
}
\arguments{
\item{x}{Matrix - contingency table}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate Pearson's C (sometimes called Pearson's Contingency Coefficient).
}
