% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.pearson.phi.R
\name{cor.pearson.phi}
\alias{cor.pearson.phi}
\title{Pearson's Phi Coefficient}
\usage{
cor.pearson.phi(x)
}
\arguments{
\item{x}{Matrix - contingency table}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate Pearson's Phi. If x is a 2x2 contingency table, Fisher's exact test is reported for significance. 
For any other dimension of contingency table, a chi square test is performed and reported.
}
