% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.pearson.r.onesample.R,
%   R/cor.pearson.r.onesample.simple.R
\name{cor.pearson.r.onesample}
\alias{cor.pearson.r.onesample}
\alias{cor.pearson.r.onesample.simple}
\title{One Sample Test of Pearson's Correlation Coefficient}
\usage{
cor.pearson.r.onesample(
  x,
  y,
  null.hypothesis.rho = 0,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  na.rm = T
)

cor.pearson.r.onesample.simple(
  sample.r,
  sample.size,
  null.hypothesis.rho = 0,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{x}{Vector - Variable 1 values}

\item{y}{Vector - Variable 2 values}

\item{null.hypothesis.rho}{Scalar - The assumed value of rho to test the sample correlation coefficient against.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{sample.r}{Scalar - Sample correlation coefficient.}

\item{sample.size}{Scalar - Sample size to use for the calculation.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate test of significance for Pearson's Correlation Coefficient.
}
