% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.pearson.r.twosample.independent.R,
%   R/cor.pearson.r.twosample.independent.simple.R
\name{cor.pearson.r.twosample.independent}
\alias{cor.pearson.r.twosample.independent}
\alias{cor.pearson.r.twosample.independent.simple}
\title{Two Independent Sample Test of Pearson's Correlation Coefficient}
\usage{
cor.pearson.r.twosample.independent(
  x1,
  x2,
  x3,
  x4,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)

cor.pearson.r.twosample.independent.simple(
  sample.r.g1.g2,
  sample.size.g1.g2,
  sample.r.g3.g4,
  sample.size.g3.g4,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{x1}{Vector - Variable 1 values}

\item{x2}{Vector - Variable 2 values}

\item{x3}{Vector - Variable 4 values}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{sample.r.g1.g2}{Scalar - Sample correlation coefficient between x1 and x2.}

\item{sample.size.g1.g2}{Scalar - Sample size for correlation between x1 and x2.}

\item{sample.r.g3.g4}{Scalar - Sample correlation coefficient between x3 and x4.}

\item{sample.size.g3.g4}{Scalar - Sample size for correlation between x3 and x4.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate test of significance difference for Pearson's Correlation Coefficient between four samples.
Null hypothesis: No significant difference between correlation coefficient between x1 and x2 vs. 
correlation coefficient between x3 and x4.
Significant result: Low p value indicates that a statistically significant difference 
exists between correlation coefficient between x1 and x2 vs. correlation coefficient between x3 and x4.
}
