% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.sakoda.p.R
\name{cor.sakoda.p}
\alias{cor.sakoda.p}
\title{Sakoda's p* Coefficient}
\usage{
cor.sakoda.p(x)
}
\arguments{
\item{x}{Matrix - contingency table}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate Sakoda's p* for a contingency table.
}
