% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.spearman.rank.R,
%   R/cor.spearman.rank.simple.R
\name{cor.spearman.rank}
\alias{cor.spearman.rank}
\alias{cor.spearman.rank.simple}
\title{Spearman's Rank Correlation Coefficient}
\usage{
cor.spearman.rank(
  x1,
  x2,
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater")
)

cor.spearman.rank.simple(
  r_sp,
  sample.size,
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{x1}{Vector - Variable 1 values}

\item{x2}{Vector - Variable 2 values}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{r_sp}{Scalar - Sample rank correlation coefficient.}

\item{sample.size}{Scalar - Sample size to use for the calculation.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate test of significance for Spearman's Rank Correlation Coefficient.
}
