% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.tetrachoric.R
\name{cor.tetrachoric}
\alias{cor.tetrachoric}
\title{Tetrachoric Correlation Coefficient}
\usage{
cor.tetrachoric(
  x,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  method = c("cosine", "sine")
)
}
\arguments{
\item{x}{Matrix - 2x2 contingency table}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{method}{Method to use for calculation}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate Tetrachoric Correlation Coefficient on a 2x2 contingency table.
}
