% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.tschuprow.tc.R
\name{cor.tschuprow.tc}
\alias{cor.tschuprow.tc}
\alias{cor.chuprov.tc}
\title{Tschuprow's T}
\usage{
cor.tschuprow.tc(x)

cor.chuprov.tc(x)
}
\arguments{
\item{x}{Matrix - contingency table}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate Tschuprow's T on a contingency table.
}
