% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.yule.q.R
\name{cor.yule.q}
\alias{cor.yule.q}
\title{Yule's Q Coefficient}
\usage{
cor.yule.q(x, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{Matrix - 2x2 contingency table}

\item{alternative}{The alternative hypothesis to use for the test computation.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculate Yule's Q Coefficient.
}
