% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagostino.normality.omnibus.test.R,
%   R/dagostino.normality.omnibus.test.simple.R
\name{dagostino.normality.omnibus.test}
\alias{dagostino.normality.omnibus.test}
\alias{dagostino.normality.omnibus.test.simple}
\title{D'Agostino's Normality Test - Omnibus}
\usage{
dagostino.normality.omnibus.test(x, conf.level = 0.95, alternative = "greater")

dagostino.normality.omnibus.test.simple(
  skewness,
  kurtosis,
  sample.size,
  input = c("fisher"),
  conf.level = 0.95,
  alternative = "greater"
)
}
\arguments{
\item{x}{Vector/numeric - sample data.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{skewness}{Scalar/numeric - sample skewness.}

\item{kurtosis}{Scalar/numeric - sample kurtosis.}

\item{sample.size}{Scalar/numeric - sample size.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculates a statistical test to determine if normality assumption for the sample is rejected.
}
