% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleted.case.statistic.R
\name{deleted.case.statistic}
\alias{deleted.case.statistic}
\title{Deleted Case Statistic Calculation}
\usage{
deleted.case.statistic(x, FUN = mean)
}
\arguments{
\item{x}{Vector to compute deleted case statistics values for.}

\item{FUN}{Function used to compute summary statistic.}
}
\value{
Vector with same length as x. Each value i is the result of FUN(x) where x[i] is removed.
}
\description{
Calculates a summary statistic for each value of an input vector by calculating 
the value of the statistic when a particular case is deleted.
}
