% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexp.low.R
\name{dexp.low}
\alias{dexp.low}
\alias{pexp.low}
\alias{qexp.low}
\alias{rexp.low}
\title{Shifted Exponential Distribution}
\usage{
dexp.low(x, rate = 1, low = 0, log = FALSE)

pexp.low(q, rate = 1, low = 0, mean = NA, lower.tail = TRUE, log.p = FALSE)

qexp.low(p, rate = 1, low = 0, mean = NA, lower.tail = TRUE, log.p = FALSE)

rexp.low(n, rate = 1, low = 0, mean = NA)
}
\arguments{
\item{x}{Scalar - numeric - quantile.}

\item{rate}{Scalar - numeric - distribution rate.}

\item{low}{Scalar - numeric - Minimum distribution value.}

\item{log}{Scalar - logical - if true, probabilities reported as log(p).}

\item{q}{Scalar - numeric - quantile.}

\item{mean}{Scalar - numeric - mean of distribution. If specified, rate is 1/(mean-low)}

\item{lower.tail}{Scalar - logical - if true, probabilities are lower tail probabilities.}

\item{log.p}{Scalar - logical - if true, probabilities reported as log(p).}

\item{p}{Scalar - numeric - probability.}

\item{n}{Scalar - integer - number of observations.}
}
\value{
dexp.low returns density function, pexp.low returns distribution function, qexp.low returns quantiles, and rexp.low returns random value(s).
}
\description{
Calculations for shifted exponential distribution, sometimes called the exponential(low) distribution. 
This is equivalent to a regular exponential distribution, but shifted on the X axis left or right so that 
it has a non-zero origin. It can be specified in two ways, using a minimum and rate parameter or using 
a minimum and mean parameter.
}
