% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.ADA.R
\name{dispersion.ADA}
\alias{dispersion.ADA}
\title{Absolute Deviation From Average}
\usage{
dispersion.ADA(x)
}
\arguments{
\item{x}{Vector to compute ADA values for.}
}
\value{
Vector with results of ADA calculation.
}
\description{
Calculates absolute deviation from average (ADA). This is also sometimes called mean absolute deviation (MAD)
}
