% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.ADM.R
\name{dispersion.ADM}
\alias{dispersion.ADM}
\title{Absolute Deviation From Median}
\usage{
dispersion.ADM(x)
}
\arguments{
\item{x}{Vector to compute ADM values for.}
}
\value{
Vector with results of ADM calculation.
}
\description{
Calculates absolute deviation from median (ADM).
}
