% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.ADMn1.R
\name{dispersion.ADMn1}
\alias{dispersion.ADMn1}
\title{Absolute Deviation From Median (Midpoint Removed)}
\usage{
dispersion.ADMn1(x)
}
\arguments{
\item{x}{Vector to compute \eqn{ADM_{n-1}}{ADMn-1} values for.}
}
\value{
Vector with results of \eqn{ADM_{n-1}}{ADMn-1} calculation.
}
\description{
Calculates absolute deviation from median (ADM) while removing the midpoint.
}
