% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/du.dist.R
\name{du.dist}
\alias{du.dist}
\alias{pu.dist}
\alias{qu.dist}
\alias{ru.dist}
\title{U Shaped Distribution}
\usage{
du.dist(x, a = 1)

pu.dist(q, a = 1, lower.tail = T)

qu.dist(p, a = 1, lower.tail = T)

ru.dist(n, a = 1)
}
\arguments{
\item{x}{Scalar - numeric - quantile.}

\item{a}{Scalar - numeric - support for distribution (-a,a).}

\item{q}{Scalar - numeric - quantile.}

\item{lower.tail}{Scalar - logical - if true, probabilities are lower tail probabilities.}

\item{p}{Scalar - numeric - probability.}

\item{n}{Scalar - integer - number of observations.}
}
\value{
du.dist returns density function, pu.dist returns distribution function, qu.dist returns quantiles, and ru.dist returns random value(s).
}
\description{
Calculate a U Shaped Distribution with support from -a to a. This is also the distribution of a harmonic oscillator.
}
\examples{
## Example 1 - plot the distribution function 
# x <- seq(-1.9,1.9,.001)
# y <- du.dist(x, a=2)
# plot(x,y, type="l", col="blue", main="U Shaped Distribution", xlab="quantile", ylab="density)

}
\references{
Robinett, R.W. Quantum and classical probability distributions for position and momentum. Am. J. Phys. 63 (9) September 1995. pp 823-832.
}
