% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis}
\usage{
kurtosis(x, method = c("fisher", "pearson"))
}
\arguments{
\item{x}{A vector of values}

\item{method}{The method to use to compute kurtosis. Methods due to Pearson and Fisher are available.}
}
\value{
A scalar containing the kurtosis estimate.
}
\description{
Calculate kurtosis for a variable.
}
