% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.test.R, R/kurtosis.test.simple.R,
%   R/kurtosis.test.zar.simple.R
\name{kurtosis.test}
\alias{kurtosis.test}
\alias{kurtosis.test.simple}
\alias{kurtosis.test.zar.simple}
\title{D'Agostino's Normality Test - Kurtosis}
\usage{
kurtosis.test(
  x,
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater")
)

kurtosis.test.simple(
  kurtosis,
  sample.size,
  input = c("fisher"),
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater")
)

kurtosis.test.zar.simple(
  kurtosis,
  sample.size,
  input = c("fisher"),
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{x}{Vector/numeric - sample data.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{kurtosis}{Scalar/numeric - sample kurtosis.}

\item{sample.size}{Scalar/numeric - sample size.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculates a statistical test to determine if sample kurtosis differs from zero.
}
