\name{lolcat-package}
\alias{lolcat-package}
\alias{lolcat}
\docType{package}
\title{
\packageTitle{lolcat}
}
\description{
\packageDescription{lolcat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lolcat}
\packageIndices{lolcat}
Miscellaneous stuff written by Mike for reuse purposes.
}
\author{
\packageAuthor{lolcat}

Maintainer: \packageMaintainer{lolcat}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# No examples at the package level
}