% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.t.test.onesample.R,
%   R/mean.t.test.onesample.simple.R
\name{mean.t.test.onesample}
\alias{mean.t.test.onesample}
\alias{mean.t.test.onesample.simple}
\title{One Sample t Test for Mean}
\usage{
\method{mean}{t.test.onesample}(
  x,
  null.hypothesis.mean = 0,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  finite.population.N = NA,
  na.rm = T
)

\method{mean}{t.test.onesample.simple}(
  sample.mean,
  sample.variance,
  sample.size,
  null.hypothesis.mean = 0,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  finite.population.N = NA
)
}
\arguments{
\item{x}{Vector - sample values to be used for mean calculation.}

\item{null.hypothesis.mean}{Scalar - null hypothesis mean value, sample.mean is tested against this value.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{finite.population.N}{Scalar - the population size if finite population adjustment needs to be made.}

\item{na.rm}{Logical - remove NA values prior to running calculation.}

\item{sample.mean}{Scalar - sample mean to be tested.}

\item{sample.variance}{Scalar - sample variance.}

\item{sample.size}{Scalar - sample size.}
}
\value{
The results of the statistical test.
}
\description{
Calculate one sample t test for mean.
}
