% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.continuous.repeatability.reproducibility.R
\name{msa.continuous.repeatability.reproducibility}
\alias{msa.continuous.repeatability.reproducibility}
\title{Perform Gage Repeatability and Reproducibility Study}
\usage{
msa.continuous.repeatability.reproducibility(
  measurement,
  part = rep(1, length(measurement)),
  appraiser = rep(1, length(measurement)),
  conf.level.interaction = 0.95,
  stat.lsl = NA,
  stat.usl = NA
)
}
\arguments{
\item{measurement}{A vector of values used as measurements}

\item{part}{A vector/factor of values used as part identifiers}

\item{appraiser}{A vector/factor of values used as appraiser identifiers}

\item{conf.level.interaction}{Confidence level to drop interaction and use combined error term for ANOVA}

\item{stat.lsl}{Lower specification limit (if applicable)}

\item{stat.usl}{Upper specification limit (if applicable)}
}
\value{
A list including all results. Recommended pieces to use are summary.aov.reduced, 
        ev.reduced, summary.reduced, ev.reduced.number.distinct.categories, and ev.reduced.discrimination.ratio
}
\description{
Uses either 1 way (single appraiser) analysis of variance or 2-way (multi appraiser) random effects analysis of variance. 
Calculates components of variance and study variation.
}
