% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.nominal.concordance.R
\name{msa.nominal.concordance}
\alias{msa.nominal.concordance}
\title{Perform Study of Nominal Scale Agreement/Concordance With or Without Standard}
\usage{
msa.nominal.concordance(..., standard = NA, conf.level = 0.95)
}
\arguments{
\item{...}{Vectors - character or factor - One or more rater's evaluations of subjects}

\item{standard}{Vector - character or factor - the "standard" or "true" measurement}

\item{conf.level}{Confidence level to use for the statistical tests}
}
\value{
A data structure including all of the details of the tests between operators/standard. Use summary() to summarize results.
}
\description{
Evaluate one or more rater's agreement for a set of subjects with or without a "standard" or "true" value.
This is sometimes called a nominal gage study or a study of interrater agreement. 
Note that all input vectors should be ordered by parts/subjects (i.e 1st value in rater 1's vector is first part, 
2nd value in rater 2's vector is 2nd part, 3rd value in standard is 3rd part, etc.)
}
