% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.nominal.concordance.simple.R
\name{msa.nominal.concordance.simple}
\alias{msa.nominal.concordance.simple}
\title{Perform Comparison of Two Nominal Measurements}
\usage{
msa.nominal.concordance.simple(measurement1, measurement2, conf.level = 0.95)
}
\arguments{
\item{measurement1}{Vector - First operator's assessments ordered by part}

\item{measurement2}{Vector - Second operator's assessments ordered by part}

\item{conf.level}{Confidence level to use for the statistical tests}
}
\value{
A data structure including all of the details of the tests. Use summary() to summarize results.
}
\description{
Evaluate statistical tests of agreement and symmetry for two measurements.
}
