% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural.tolerance.chisquare.simple.R
\name{natural.tolerance.chisquare.simple}
\alias{natural.tolerance.chisquare.simple}
\title{Natural Tolerance - Chi Square Distribution}
\usage{
natural.tolerance.chisquare.simple(df = 1, ncp = 0, ...)
}
\arguments{
\item{df}{Scalar/numeric - degrees of freedom.}

\item{ncp}{Scalar/numeric - non-centrality parameter in case of noncentral chi square distribution.}

\item{...}{Additional parameters for natural.tolerance}
}
\value{
Scalar or data.frame - scalar if details = F, otherwise data frame with details from calculation.
}
\description{
Calculates the natural tolerance of a given probability distribution, defined as F(p.upper)-F(p.lower) where 
F is a quantile function for a given univariate distribution. Default is to calculate the natural tolerance 
with a total tail area of .0027, or .00135 on the lower tail and .00135 on the upper tail.
}
