% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural.tolerance.exp.low.R,
%   R/natural.tolerance.exp.low.simple.R
\name{natural.tolerance.exp.low}
\alias{natural.tolerance.exp.low}
\alias{natural.tolerance.exp.low.simple}
\title{Natural Tolerance - Exponential(low) Distribution}
\usage{
natural.tolerance.exp.low(x, low = min(x), ...)

natural.tolerance.exp.low.simple(rate = 1, low = 0, ...)
}
\arguments{
\item{x}{Vector/numeric - data to fit parameters}

\item{low}{Scalar/numeric - minimum value for distribution}

\item{...}{Additional parameters for natural.tolerance}

\item{rate}{Scalar/numeric - rate parameter for exponential distribution.}
}
\value{
Scalar or data.frame - scalar if details = F, otherwise data frame with details from calculation.
}
\description{
Calculates the natural tolerance of a given probability distribution, defined as F(p.upper)-F(p.lower) where 
F is a quantile function for a given univariate distribution. Default is to calculate the natural tolerance 
with a total tail area of .0027, or .00135 on the lower tail and .00135 on the upper tail.
}
