% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural.tolerance.normal.R,
%   R/natural.tolerance.normal.simple.R
\name{natural.tolerance.normal}
\alias{natural.tolerance.normal}
\alias{natural.tolerance.normal.simple}
\title{Natural Tolerance - Normal/Gaussian Distribution}
\usage{
natural.tolerance.normal(x, ...)

natural.tolerance.normal.simple(mean = 1, variance = 1, ...)
}
\arguments{
\item{x}{Vector/numeric - data to fit parameters}

\item{...}{Additional parameters for natural.tolerance}

\item{mean}{Scalar/numeric - mean parameter for normal distribution.}

\item{variance}{Scalar/numeric - variance parameter for normal distribution, expressed as variance.}
}
\value{
Scalar or data.frame - scalar if details = F, otherwise data frame with details from calculation.
}
\description{
Calculates the natural tolerance of a given probability distribution, defined as F(p.upper)-F(p.lower) where 
F is a quantile function for a given univariate distribution. Default is to calculate the natural tolerance 
with a total tail area of .0027, or .00135 on the lower tail and .00135 on the upper tail.
}
