% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.dist.test.R
\name{poisson.dist.test}
\alias{poisson.dist.test}
\title{Shape Tests - Poisson Distribution Test}
\usage{
poisson.dist.test(x)
}
\arguments{
\item{x}{Vector/numeric - Sample Values}
}
\value{
Hypothesis test result showing results of test. Low p value rejects assumption of data fitting Poisson distribution.
}
\description{
Calculates a test to see if the sample data cannot be assumed to be Poisson distributed..
}
