% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.count.poisson.onesample.exact.R
\name{power.count.poisson.onesample.exact}
\alias{power.count.poisson.onesample.exact}
\title{Power - Single Sample Poisson Test}
\usage{
power.count.poisson.onesample.exact(
  sample.size,
  lambda.null.hypothesis,
  lambda.alternative.hypothesis,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater"),
  details = T
)
}
\arguments{
\item{sample.size}{Scalar - sample size}

\item{lambda.null.hypothesis}{Scalar - null hypothesis lambda parameter}

\item{lambda.alternative.hypothesis}{Scalar - alternative hypothesis lambda parameter}

\item{alpha}{Scalar - Type I error rate}

\item{alternative}{Scalar (character) - alternative hypothesis}

\item{details}{Logical - Return calculation details (default) or return only power (details = FALSE)}
}
\value{
A data frame with details about the calculation or a single value with power (details = F).
}
\description{
Power calculation utilizes the Poisson distribution ("exact"" sample size/power).
}
