% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.group.plot.R
\name{process.group.plot}
\alias{process.group.plot}
\title{One-Way and Two-Way Histograms/Plots}
\usage{
process.group.plot(
  fx,
  data = NULL,
  FUN = hist.grouped,
  interval.size = NA,
  anchor.value = NA,
  width.consider = lolcat.default.width.consider,
  right = F,
  hist.correct.label = isTRUE(all.equal(FUN, hist.grouped)),
  constant.x.axis = T,
  constant.y.axis = T,
  ...
)
}
\arguments{
\item{fx}{Formula - grouping formula of form dv ~ iv or dv ~ iv1 + iv2}

\item{data}{Data Frame - The data for the plots}

\item{FUN}{Function - Plotting function compatible with parameters for hist.grouped}

\item{interval.size}{Numeric - Bin size}

\item{anchor.value}{Numeric - One of the breakpoints}

\item{width.consider}{Numeric - Bin sizes to consider when plotting histograms}

\item{right}{Logical - If TRUE, the histogram cells are right-closed (left open) intervals.}

\item{hist.correct.label}{Logical - Provide better placement for labels compared with the labeled cells/bins.}

\item{constant.x.axis}{Logical - If TRUE, ensure all plots have the same domains.}

\item{constant.y.axis}{Logical - If TRUE, ensure all plots have the same ranges.}

\item{...}{Additional parameters - Additional parameters passed to FUN}
}
\value{
No return value
}
\description{
Prepares data and plots a dependent variable across one or two independent variables,
yielding a matrix of plots.
}
