% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.test.onesample.approximate.R,
%   R/proportion.test.onesample.approximate.simple.R
\name{proportion.test.onesample.approximate}
\alias{proportion.test.onesample.approximate}
\alias{proportion.test.onesample.approximate.simple}
\title{One Sample Proportion Test (Approximate)}
\usage{
proportion.test.onesample.approximate(
  x,
  success.value = 1,
  null.hypothesis.proportion = 0.5,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  continuity.correction = T
)

proportion.test.onesample.approximate.simple(
  sample.proportion,
  sample.size,
  null.hypothesis.proportion = 0.5,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  continuity.correction = T
)
}
\arguments{
\item{x}{Vector - Sample Values}

\item{success.value}{Scalar - Value compared with x using == operator to determine if a trial is a "success"}

\item{null.hypothesis.proportion}{Scalar/numeric - assumed population proportion.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{continuity.correction}{Scalar/logical - if TRUE, apply the continuity correction to the calculation}

\item{sample.proportion}{Scalar/numeric - sample proportion between 0 and 1}

\item{sample.size}{Scalar/numeric - sample size.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculates a one-sample proportion test to determine if a sample 
proportion is statistically different from an assumed population proportion.
}
