% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.test.onesample.exact.R,
%   R/proportion.test.onesample.exact.simple.R
\name{proportion.test.onesample.exact}
\alias{proportion.test.onesample.exact}
\alias{proportion.test.onesample.exact.simple}
\title{One Sample Proportion Test (Exact)}
\usage{
proportion.test.onesample.exact(
  x,
  success.value = 1,
  null.hypothesis.proportion = 0.5,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)

proportion.test.onesample.exact.simple(
  sample.proportion = NA,
  sample.size = NA,
  np = sample.size * sample.proportion,
  null.hypothesis.proportion = 0.5,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{x}{Vector - Sample Values}

\item{success.value}{Scalar - Value compared with x using == operator to determine if a trial is a "success"}

\item{null.hypothesis.proportion}{Scalar/numeric - assumed population proportion.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{sample.proportion}{Scalar/numeric - sample proportion between 0 and 1. Required if np not specified.}

\item{sample.size}{Scalar/numeric - sample size. Required.}

\item{np}{Scalar/numeric - sample size multipled by sample proportion. Required if sample.proportion not specified.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculates a one-sample proportion test to determine if a sample 
proportion is statistically different from an assumed population proportion.
}
