% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.test.twosample.approximate.simple.R
\name{proportion.test.twosample.approximate.simple}
\alias{proportion.test.twosample.approximate.simple}
\title{Two Sample Proportion Test (Approximate)}
\usage{
proportion.test.twosample.approximate.simple(
  sample.proportion.g1,
  sample.size.g1,
  sample.proportion.g2,
  sample.size.g2,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  continuity.correction = T
)
}
\arguments{
\item{sample.proportion.g1}{Scalar/numeric - Group 1 - sample proportion between 0 and 1}

\item{sample.size.g1}{Scalar/numeric - Group 1 - sample size.}

\item{sample.proportion.g2}{Scalar/numeric - Group 2 - sample proportion between 0 and 1}

\item{sample.size.g2}{Scalar/numeric - Group 2 - sample size.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{continuity.correction}{Scalar/logical - if TRUE, use continuity correction for the test.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculates a two-sample proportion test to determine if a samples  
come from different populations. Z test is used for test computation.
}
