% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.reverse.R
\name{rank.reverse}
\alias{rank.reverse}
\title{Calculate Reverse Ranks}
\usage{
rank.reverse(x, ...)
}
\arguments{
\item{x}{Vector - The object to rank}

\item{...}{Additional parameters - Additional parameters for rank(...) (ex. na.last or ties.method)}
}
\value{
Rank values for x from largest to smallest
}
\description{
Calculates ranks where rank 1 is largest value and rank k is smallest value.
}
\seealso{
\code{\link{rank}}
}
