% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.siegel.tukey.R
\name{rank.siegel.tukey}
\alias{rank.siegel.tukey}
\title{Calculate Ranks for Siegel-Tukey Variability Test}
\usage{
rank.siegel.tukey(
  g1,
  g2,
  details = T,
  ties.method = c("average", "max", "min", "none", "first", "random")
)
}
\arguments{
\item{g1}{Vector - Group 1 data values}

\item{g2}{Vector - Group 2 data values}

\item{details}{Scalar/logical - Whether to return full rank details or just ranks}

\item{ties.method}{Scalar/character - The method used to break ties during ranking}
}
\value{
Rank values for g1 and g2
}
\description{
Calculate ranks for the Siegel-Tukey test for variability.
}
