% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend.box.cox.R
\name{recommend.box.cox}
\alias{recommend.box.cox}
\title{Normalizing Transformations - Box-Cox/Ladder of Powers - Parameter Recommendation}
\usage{
recommend.box.cox(
  x,
  lambda.min = -5,
  lambda.max = 5,
  step = 0.1,
  correct.min = T,
  target = c("normal"),
  metric = c("skewness+kurtosis", "skewness", "kurtosis")
)
}
\arguments{
\item{x}{Vector/numeric - sample data.}

\item{lambda.min}{Scalar/numeric - Minimum exponent.}

\item{lambda.max}{Scalar/numeric - Maximum exponent.}

\item{step}{- Scalar/numeric - The step size between min and max exponent to try.}

\item{correct.min}{Scalar/logical - Shift data by minimum in sample to avoid transformation/reverse transformation issues.}

\item{target}{- Scalar/character - The target distribution.}

\item{metric}{- Scalar/character - Which metric to use for evaluating transformations.}
}
\value{
Data frame with best parameters found in search.
}
\description{
Recommend parameters to optimally transform sample data to normality based on the Box Cox transformation.
}
