% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmnames.R
\name{rmnames}
\alias{rmnames}
\title{Remove Names From Object}
\usage{
rmnames(x)
}
\arguments{
\item{x}{Vector - The object to remove names from}
}
\value{
x, but without names
}
\description{
Returns the input object, but without names.
}
