% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.object.R
\name{round.object}
\alias{round.object}
\title{Round Numeric Values in Object}
\usage{
\method{round}{object}(x, digits = 4)
}
\arguments{
\item{x}{Vector - The object to remove names from}

\item{digits}{Scalar - The number of digits to round to}
}
\value{
x, but rounded
}
\description{
Returns the input object, but rounded.
}
