% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.mode.R
\name{sample.mode}
\alias{sample.mode}
\title{Calculate Sample Mode}
\usage{
sample.mode(x)
}
\arguments{
\item{x}{Vector - The object to remove names from}
}
\value{
Calculated mode value
}
\description{
Calculate the mode of a data set, defined as the value(s) that occur most often.
}
