% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetest.exp.epps.pulley.1986.R
\name{shapetest.exp.epps.pulley.1986}
\alias{shapetest.exp.epps.pulley.1986}
\title{Epps-Pulley Exponentiality Test}
\usage{
shapetest.exp.epps.pulley.1986(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
An htest object with results of test.
}
\description{
Calculate Epps-Pulley exponentiality test for a variable. 
Low p-values indicate rejection of assumption of exponential distribution.
}
