% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro.wilk.normality.test.R
\name{shapiro.wilk.normality.test}
\alias{shapiro.wilk.normality.test}
\title{Shapiro-Wilk Normality Test}
\usage{
shapiro.wilk.normality.test(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
An htest object with results of test.
}
\description{
Calculate Shapiro-Wilk normality test for a variable. 
Low p-values indicate rejection of assumption of normality.
}
