% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x, method = c("fisher", "pearson"))
}
\arguments{
\item{x}{A vector of values}

\item{method}{The method to use to compute skewness. Methods due to Pearson and Fisher are available.}
}
\value{
A scalar containing the skewness estimate.
}
\description{
Calculate skewness for a variable.
}
