% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.test.R, R/skewness.test.simple.R,
%   R/skewness.test.zar.simple.R
\name{skewness.test}
\alias{skewness.test}
\alias{skewness.test.simple}
\alias{skewness.test.zar.simple}
\title{D'Agostino's Normality Test - Skewness}
\usage{
skewness.test(
  x,
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater")
)

skewness.test.simple(
  skewness,
  sample.size,
  input = c("fisher"),
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater")
)

skewness.test.zar.simple(
  skewness,
  sample.size,
  input = c("fisher"),
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{x}{Vector/numeric - sample data.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{skewness}{Scalar/numeric - sample skewness.}

\item{sample.size}{Scalar/numeric - sample size.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculates a statistical test to determine if sample skewness differs from zero.
}
