% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.capability.cp.simple.R
\name{spc.capability.cp.simple}
\alias{spc.capability.cp.simple}
\title{Calculate Capability Measures - Cp}
\usage{
spc.capability.cp.simple(
  lower.specification,
  upper.specification,
  process.center = NA,
  process.natural.tolerance
)
}
\arguments{
\item{lower.specification}{Lower specification limit (if applicable)}

\item{upper.specification}{Upper specification limit (if applicable)}

\item{process.center}{Estimate of process center, used if specification limits are one-sided}

\item{process.natural.tolerance}{Estimate of process natural tolerance}
}
\value{
A scalar with computed Cp.
}
\description{
Calculate Cp, a measure that compares specification spread with observed process process spread. 
Higher values indicate that more process 'widths' can fit into the specification limits.
This measure does not identify whether a process is 'targeted' within the specification limits.
}
