% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.capability.cpk.simple.R
\name{spc.capability.cpk.simple}
\alias{spc.capability.cpk.simple}
\title{Calculate Capability Measures - Cpk}
\usage{
spc.capability.cpk.simple(
  lower.specification,
  upper.specification,
  process.variability,
  process.center,
  n.sigma = 6
)
}
\arguments{
\item{lower.specification}{Lower specification limit (if applicable)}

\item{upper.specification}{Upper specification limit (if applicable)}

\item{process.variability}{Estimate of process variability, expressed as variance}

\item{process.center}{Estimate of process center}

\item{n.sigma}{The number of standard deviations to use in the denominator of the calculation. 6 is recommended, but 5.15 has also been historically used by Automotive Industry Action Group (AIAG).}
}
\value{
A scalar with computed Cp.
}
\description{
Calculate Cpk, a measure that compares the distance between process center and nearest specification limit to process variability. 
Lower process variability and larger distance between process center and nearest specification limit lead to higher values, which 
are generally considered better.
}
