% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.capability.cpm.simple.R
\name{spc.capability.cpm.simple}
\alias{spc.capability.cpm.simple}
\title{Calculate Capability Measures - Cpm}
\usage{
spc.capability.cpm.simple(
  lower.specification,
  upper.specification,
  process.variability,
  process.center,
  nominal.center,
  n.sigma = 6
)
}
\arguments{
\item{lower.specification}{Lower specification limit (if applicable)}

\item{upper.specification}{Upper specification limit (if applicable)}

\item{process.variability}{Estimate of process variability, expressed as variance}

\item{process.center}{Estimate of process center}

\item{nominal.center}{Nominal target for the process}

\item{n.sigma}{The number of standard deviations to use in the denominator of the calculation. 6 is recommended, but 5.15 has also been historically used by Automotive Industry Action Group (AIAG).}
}
\value{
A scalar with computed Cpm.
}
\description{
Calculate Cpm, a measure that combines the measure of variability and targeting relative to nominal specifications. Higher is better.
}
