% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.capability.cpU.simple.R
\name{spc.capability.cpU.simple}
\alias{spc.capability.cpU.simple}
\title{Calculate Capability Measures - CpU}
\usage{
spc.capability.cpU.simple(
  upper.specification,
  process.center,
  process.variability,
  n.sigma = 6
)
}
\arguments{
\item{upper.specification}{Upper specification limit (if applicable)}

\item{process.center}{Estimate of process center}

\item{process.variability}{Estimate of process variability, expressed as variance}

\item{n.sigma}{The number of standard deviations to use in the denominator of the calculation. 6 is recommended, but 5.15 has also been historically used by Automotive Industry Action Group (AIAG).}
}
\value{
A scalar with computed CpU.
}
\description{
Calculate CpU, a measure that compares the distance between process center and upper specification limit to process variability. 
Lower process variability and larger distance between process center and upper specification limit lead to higher values, which 
are generally considered better.
}
