% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.capability.summary.normal.simple.R
\name{spc.capability.summary.normal.simple}
\alias{spc.capability.summary.normal.simple}
\title{Calculate Capability Measures - Normally Distributed Data}
\usage{
spc.capability.summary.normal.simple(
  stat.lsl = NA,
  stat.target = NA,
  stat.usl = NA,
  process.center = NA,
  process.variability.estimate = NA,
  process.variability.overall = NA,
  process.n.upper = NA,
  process.n.lower = NA,
  process.n = NA,
  process.ppm.upper = round(1e+06 * process.n.upper/process.n),
  process.ppm.lower = round(1e+06 * process.n.lower/process.n)
)
}
\arguments{
\item{stat.lsl}{Lower specification limit (if applicable)}

\item{stat.target}{Target for process}

\item{stat.usl}{Upper specification limit (if applicable)}

\item{process.center}{Estimate of process center}

\item{process.variability.estimate}{Estimate of process variability, expressed as variance, usually estimated based on within-group variability.}

\item{process.variability.overall}{Estimate of process variability for performance measures, expressed as variance, usually estimated based on overall variability.}

\item{process.n.upper}{Observed count above upper specification limit, used to calculate PPM (parts per million).}

\item{process.n.lower}{Observed count below lower specification limit, used to calculate PPM (parts per million).}

\item{process.n}{Total count of subjects/objects, used to calculate PPM (parts per million).}

\item{process.ppm.upper}{Observed parts per million (PPM) above upper specification limit.}

\item{process.ppm.lower}{Observed parts per million (PPM) below lower specification limit.}
}
\value{
A data frame with computed measures.
}
\description{
Calculate capability measures and potential out of specification parts per million.
}
