% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/spc.capability.summary.ungrouped.nonnormal.simple.R
\name{spc.capability.summary.ungrouped.nonnormal.simple.R}
\alias{spc.capability.summary.ungrouped.nonnormal.simple.R}
\title{Calculate Capability Measures - Non-Normally Distributed Data}
\usage{
spc.capability.summary.ungrouped.nonnormal.simple.R(
  stat.lsl = NA,
  stat.target = NA,
  stat.usl = NA,
  stat.lsl.capability = stat.lsl,
  stat.target.capability = stat.target,
  stat.usl.capability = stat.usl,
  stat.lsl.performance = stat.lsl,
  stat.target.performance = stat.target,
  stat.usl.performance = stat.usl,
  process.center.capability = NA,
  process.center.performance = NA,
  process.variability = NA,
  process.n.upper = NA,
  process.n.lower = NA,
  process.n = NA,
  process.ppm.upper = round(1e+06 * process.n.upper/process.n),
  process.ppm.lower = round(1e+06 * process.n.lower/process.n),
  natural.tolerance = NA,
  p.lower = NA,
  p.upper = NA
)
}
\arguments{
\item{stat.lsl}{Lower specification limit (if applicable)}

\item{stat.target}{Target for process}

\item{stat.usl}{Upper specification limit (if applicable)}

\item{stat.lsl.capability}{Lower specification limit (if applicable) for capability measures}

\item{stat.target.capability}{Target for process for capability measures}

\item{stat.usl.capability}{Upper specification limit (if applicable) for capability measures}

\item{stat.lsl.performance}{Lower specification limit (if applicable) for performance measures}

\item{stat.target.performance}{Target for process for performance measures}

\item{stat.usl.performance}{Upper specification limit (if applicable) for performance measures}

\item{process.center.capability}{Estimate of process center for capability measures}

\item{process.center.performance}{Estimate of process center for performance measures}

\item{process.variability}{Estimate of process variability, expressed as variance. Used for performance measures only.}

\item{process.n.upper}{Observed count above upper specification limit, used to calculate PPM (parts per million).}

\item{process.n.lower}{Observed count below lower specification limit, used to calculate PPM (parts per million).}

\item{process.n}{Total count of subjects/objects, used to calculate PPM (parts per million).}

\item{process.ppm.upper}{Observed parts per million (PPM) above upper specification limit.}

\item{process.ppm.lower}{Observed parts per million (PPM) below lower specification limit.}

\item{natural.tolerance}{Natural tolerance estimate. Used for capability measures.}

\item{p.lower}{Fit distribution area below LSL.}

\item{p.upper}{Fit distribution area above USL.}
}
\value{
A data frame with computed measures.
}
\description{
Calculate capability measures and potential out of specification parts per million.
}
