% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.constant.calculation.A2.R
\name{spc.constant.calculation.A2}
\alias{spc.constant.calculation.A2}
\title{Control Chart Constant A2}
\usage{
spc.constant.calculation.A2(sample.size, n.sigma = 3)
}
\arguments{
\item{sample.size}{A scalar or vector of sample sizes}

\item{n.sigma}{How many standard errors to use (default 3 for most control charts)}
}
\value{
A scalar or vector with computed constants.
}
\description{
Calculate A2 control chart constant from Wheeler's SPC books.
}
