% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.constant.calculation.B3.R
\name{spc.constant.calculation.B3}
\alias{spc.constant.calculation.B3}
\title{Control Chart Constant B3}
\usage{
spc.constant.calculation.B3(
  sample.size,
  n.sigma = 3,
  negative.as.NA = T,
  exact = T
)
}
\arguments{
\item{sample.size}{A scalar or vector of sample sizes}

\item{n.sigma}{How many standard errors to use (default 3 for most control charts)}

\item{negative.as.NA}{Logical - return NA in the case a negative value is calculated}

\item{exact}{Logical - Use exact calculation or approximate calculation (if FALSE)?}
}
\value{
A scalar or vector with computed constants.
}
\description{
Calculate B3 control chart constant from Wheeler's SPC books.
}
