% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.constant.calculation.B4.R
\name{spc.constant.calculation.B4}
\alias{spc.constant.calculation.B4}
\title{Control Chart Constant B4}
\usage{
spc.constant.calculation.B4(sample.size, n.sigma = 3, exact = T)
}
\arguments{
\item{sample.size}{A scalar or vector of sample sizes}

\item{n.sigma}{How many standard errors to use (default 3 for most control charts)}

\item{exact}{Logical - Use exact calculation or approximate calculation (if FALSE)?}
}
\value{
A scalar or vector with computed constants.
}
\description{
Calculate B4 control chart constant from Wheeler's SPC books.
}
