% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.constant.calculation.c2.R
\name{spc.constant.calculation.c2}
\alias{spc.constant.calculation.c2}
\title{Control Chart Constant c2}
\usage{
spc.constant.calculation.c2(sample.size)
}
\arguments{
\item{sample.size}{A scalar or vector of sample sizes}
}
\value{
A scalar or vector with computed constants.
}
\description{
Calculate c2 control chart constant from Wheeler's SPC books.
}
