% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.constant.calculation.c4.R
\name{spc.constant.calculation.c4}
\alias{spc.constant.calculation.c4}
\title{Control Chart Constant c4}
\usage{
spc.constant.calculation.c4(sample.size)
}
\arguments{
\item{sample.size}{A scalar or vector of sample sizes}
}
\value{
A scalar or vector with computed constants.
}
\description{
Calculate c4 control chart constant from Wheeler's SPC books.
}
