% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.constant.calculation.d3.lowercase.R
\name{spc.constant.calculation.d3}
\alias{spc.constant.calculation.d3}
\title{Control Chart Constant d3}
\usage{
spc.constant.calculation.d3(sample.size)
}
\arguments{
\item{sample.size}{A scalar or vector of sample sizes}
}
\value{
A scalar or vector with computed constants.
}
\description{
Calculate d3 control chart constant from Wheeler's SPC books.
}
