% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.constant.calculation.d4.lowercase.R
\name{spc.constant.calculation.d4}
\alias{spc.constant.calculation.d4}
\title{Control Chart Constant d4}
\usage{
spc.constant.calculation.d4(sample.size)
}
\arguments{
\item{sample.size}{A scalar or vector of sample sizes}
}
\value{
A scalar or vector with computed constants.
}
\description{
Calculate d4 control chart constant from Wheeler's SPC books.
}
