% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.constant.calculation.D5.R
\name{spc.constant.calculation.D5}
\alias{spc.constant.calculation.D5}
\title{Control Chart Constant D5}
\usage{
spc.constant.calculation.D5(sample.size, n.sigma = 3, negative.as.NA = T)
}
\arguments{
\item{sample.size}{A scalar or vector of sample sizes}

\item{n.sigma}{Scalar - How many standard errors to use (default 3 for most control charts)}

\item{negative.as.NA}{Logical - negative values are possible from the calculation, but not useful. Set to false to obtain negative values instead of NA.}
}
\value{
A scalar or vector with computed constants.
}
\description{
Calculate D5 control chart constant from Wheeler's SPC books.
}
