% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.constant.calculation.D6.R
\name{spc.constant.calculation.D6}
\alias{spc.constant.calculation.D6}
\title{Control Chart Constant D6}
\usage{
spc.constant.calculation.D6(sample.size, n.sigma = 3)
}
\arguments{
\item{sample.size}{A scalar or vector of sample sizes}

\item{n.sigma}{Scalar - How many standard errors to use (default 3 for most control charts)}
}
\value{
A scalar or vector with computed constants.
}
\description{
Calculate D6 control chart constant from Wheeler's SPC books.
}
