% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.controlviolation.evaluate.rules.R
\name{spc.controlviolation.evaluate.rules}
\alias{spc.controlviolation.evaluate.rules}
\title{Control Violation Rule Evaluation}
\usage{
spc.controlviolation.evaluate.rules(
  control.rules = list(),
  chart.series = NA,
  center.line = NA,
  control.limits.ucl = NA,
  zone.a.upper = NA,
  zone.ab.upper = NA,
  zone.bc.upper = NA,
  control.limits.lcl = NA,
  zone.a.lower = NA,
  zone.ab.lower = NA,
  zone.bc.lower = NA,
  ...
)
}
\arguments{
\item{control.rules}{List - list of functions following conventions in spc.controlviolation.template}

\item{chart.series}{Vector - Chart series}

\item{center.line}{Vector - Chart center line}

\item{control.limits.ucl}{Vector - Chart upper control limit (UCL)}

\item{zone.a.upper}{Vector - Chart Zone A boundary above center line.}

\item{zone.ab.upper}{Vector - Chart boundary between Zone A and Zone B above center line.}

\item{zone.bc.upper}{Vector - Chart boundary between Zone B and Zone C above center line.}

\item{control.limits.lcl}{Vector - Chart lower control limit (LCL)}

\item{zone.a.lower}{Vector - Chart Zone A boundary below center line.}

\item{zone.ab.lower}{Vector - Chart boundary between Zone A and Zone B below center line.}

\item{zone.bc.lower}{Vector - Chart boundary between Zone B and Zone C below center line.}

\item{...}{Additional parameters}
}
\value{
A data structure with individual rule violations (sharing names with control.rules) 
and whether each point is subject to one or more control violations.
}
\description{
This function is used to evaluate control rules in the SPC charting functions.
}
