% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/spc.controlviolation.nelson.1984.test1.outside.zone.a.upper.R
\name{spc.controlviolation.nelson.1984.test1.outside.zone.a.upper}
\alias{spc.controlviolation.nelson.1984.test1.outside.zone.a.upper}
\title{Control Violations - Nelson Rule 1 - Points Above Zone A}
\usage{
spc.controlviolation.nelson.1984.test1.outside.zone.a.upper(
  chart.series = NA,
  center.line = NA,
  control.limits.ucl = NA,
  zone.a.upper = NA,
  zone.ab.upper = NA,
  zone.bc.upper = NA,
  control.limits.lcl = NA,
  zone.a.lower = NA,
  zone.ab.lower = NA,
  zone.bc.lower = NA,
  ...
)
}
\arguments{
\item{chart.series}{Vector - Chart series}

\item{center.line}{Vector - Chart center line}

\item{control.limits.ucl}{Vector - Chart upper control limit (UCL)}

\item{zone.a.upper}{Vector - Chart Zone A boundary above center line.}

\item{zone.ab.upper}{Vector - Chart boundary between Zone A and Zone B above center line.}

\item{zone.bc.upper}{Vector - Chart boundary between Zone B and Zone C above center line.}

\item{control.limits.lcl}{Vector - Chart lower control limit (LCL)}

\item{zone.a.lower}{Vector - Chart Zone A boundary below center line.}

\item{zone.ab.lower}{Vector - Chart boundary between Zone A and Zone B below center line.}

\item{zone.bc.lower}{Vector - Chart boundary between Zone B and Zone C below center line.}

\item{...}{Additional parameters}
}
\value{
Vector of logical values corresponding to whether points in chart.series violate this rule.
}
\description{
Zone A (above and below the center line) typically corresponds with the upper and 
lower control limits for the chart.
This rule identifies points above the upper Zone A limit. 
Points in chart.series that are above the upper control limit may indicate that 
a special cause of variability is present that is causing the process to rise higher 
than would be expected for a process exhibiting a state of control.
}
