% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.controlviolation.nelson.1984.test2.runs.R
\name{spc.controlviolation.nelson.1984.test2.runs}
\alias{spc.controlviolation.nelson.1984.test2.runs}
\title{Control Violations - Nelson Rule 2 - Runs Above or Below Center Line}
\usage{
spc.controlviolation.nelson.1984.test2.runs(
  chart.series = NA,
  center.line = NA,
  control.limits.ucl = NA,
  zone.a.upper = NA,
  zone.ab.upper = NA,
  zone.bc.upper = NA,
  control.limits.lcl = NA,
  zone.a.lower = NA,
  zone.ab.lower = NA,
  zone.bc.lower = NA,
  point.count = 9,
  ...
)
}
\arguments{
\item{chart.series}{Vector - Chart series}

\item{center.line}{Vector - Chart center line}

\item{control.limits.ucl}{Vector - Chart upper control limit (UCL)}

\item{zone.a.upper}{Vector - Chart Zone A boundary above center line.}

\item{zone.ab.upper}{Vector - Chart boundary between Zone A and Zone B above center line.}

\item{zone.bc.upper}{Vector - Chart boundary between Zone B and Zone C above center line.}

\item{control.limits.lcl}{Vector - Chart lower control limit (LCL)}

\item{zone.a.lower}{Vector - Chart Zone A boundary below center line.}

\item{zone.ab.lower}{Vector - Chart boundary between Zone A and Zone B below center line.}

\item{zone.bc.lower}{Vector - Chart boundary between Zone B and Zone C below center line.}

\item{point.count}{Scalar - Number of points to use for detecting violation}

\item{...}{Additional parameters}
}
\value{
Vector of logical values corresponding to whether points in chart.series violate this rule.
}
\description{
Runs are sequential points above or below the center line on a control chart.
Too many sequential points above or below the center line indicates that the process
may be behaving non-randomly and has one or more special causes of variation present.
}
