% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.controlviolation.template.R
\name{spc.controlviolation.template}
\alias{spc.controlviolation.template}
\title{Control Violation Rule Template}
\usage{
spc.controlviolation.template(
  chart.series = NA,
  center.line = NA,
  control.limits.ucl = NA,
  zone.a.upper = NA,
  zone.ab.upper = NA,
  zone.bc.upper = NA,
  control.limits.lcl = NA,
  zone.a.lower = NA,
  zone.ab.lower = NA,
  zone.bc.lower = NA,
  ...
)
}
\arguments{
\item{chart.series}{Vector - Chart series}

\item{center.line}{Vector - Chart center line}

\item{control.limits.ucl}{Vector - Chart upper control limit (UCL)}

\item{zone.a.upper}{Vector - Chart Zone A boundary above center line.}

\item{zone.ab.upper}{Vector - Chart boundary between Zone A and Zone B above center line.}

\item{zone.bc.upper}{Vector - Chart boundary between Zone B and Zone C above center line.}

\item{control.limits.lcl}{Vector - Chart lower control limit (LCL)}

\item{zone.a.lower}{Vector - Chart Zone A boundary below center line.}

\item{zone.ab.lower}{Vector - Chart boundary between Zone A and Zone B below center line.}

\item{zone.bc.lower}{Vector - Chart boundary between Zone B and Zone C below center line.}
}
\value{
A logical vector with same length as chart.series identifying whether each point has a control violation or not.
}
\description{
Use this template to create your own chart rules for use in custom rule sets.
}
