% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.controlviolation.zones.classify.R
\name{spc.controlviolation.zones.classify}
\alias{spc.controlviolation.zones.classify}
\title{Classify SPC Chart Series By Zone}
\usage{
spc.controlviolation.zones.classify(
  chart.series = NA,
  center.line = NA,
  control.limits.ucl = NA,
  zone.a.upper = NA,
  zone.ab.upper = NA,
  zone.bc.upper = NA,
  control.limits.lcl = NA,
  zone.a.lower = NA,
  zone.ab.lower = NA,
  zone.bc.lower = NA,
  ...
)
}
\arguments{
\item{chart.series}{Vector - Chart series}

\item{center.line}{Vector - Chart center line}

\item{control.limits.ucl}{Vector - Chart upper control limit (UCL)}

\item{zone.a.upper}{Vector - Chart Zone A boundary above center line.}

\item{zone.ab.upper}{Vector - Chart boundary between Zone A and Zone B above center line.}

\item{zone.bc.upper}{Vector - Chart boundary between Zone B and Zone C above center line.}

\item{control.limits.lcl}{Vector - Chart lower control limit (LCL)}

\item{zone.a.lower}{Vector - Chart Zone A boundary below center line.}

\item{zone.ab.lower}{Vector - Chart boundary between Zone A and Zone B below center line.}

\item{zone.bc.lower}{Vector - Chart boundary between Zone B and Zone C below center line.}
}
\value{
A character vector with same length as chart.series identifying which zone a point is part of. Possible values are:
UU - beyond zone A/UCL
UA - Zone A above center line
UB - Zone B above center line
UC - Zone C above center line
CC - On Center Line
LC - Zone C below center line
LB - Zone B below center line
LA - Zone A below center line
LL - beyond zone A/LCL
}
\description{
Identify which zone points in a chart series are part of.
}
