% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.rulesets.nelson.1984.test.1.2.3.4.5.6.7.8.R
\name{spc.rulesets.nelson.1984.test.1.2.3.4.5.6.7.8}
\alias{spc.rulesets.nelson.1984.test.1.2.3.4.5.6.7.8}
\title{Control Chart Rule Set - Nelson Rules 1-8}
\usage{
spc.rulesets.nelson.1984.test.1.2.3.4.5.6.7.8()
}
\value{
A list with control violation evaluation functions to be used by spc.chart functions.
}
\description{
Provides an SPC chart rule set using eight Nelson rules: Points outside of limits, runs, 
trends, and alternating points, 2/3 points in zone A, 4/5 points in zones A/B, 
consecutive points in zone C, and consecutive points in zones A/B.
}
