% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.rulesets.nelson.1984.test.1.2.3.4.R
\name{spc.rulesets.nelson.1984.test.1.2.3.4}
\alias{spc.rulesets.nelson.1984.test.1.2.3.4}
\title{Control Chart Rule Set - Nelson Rules 1,2,3,4}
\usage{
spc.rulesets.nelson.1984.test.1.2.3.4()
}
\value{
A list with control violation evaluation functions to be used by spc.chart functions.
}
\description{
Provides an SPC chart rule set using the first four Nelson rules: Points outside of limits, runs, trends, and alternating points.
}
