% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.rulesets.outside.limits.R
\name{spc.rulesets.outside.limits}
\alias{spc.rulesets.outside.limits}
\title{Control Chart Rule Set - Outside Limits}
\usage{
spc.rulesets.outside.limits()
}
\value{
A list with control violation evaluation functions to be used by spc.chart functions.
}
\description{
Provides an SPC chart rule set that reports control violations for points outside of control limits.
}
