% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.all.variables.R, R/summary.impl.R
\name{summary.all.variables}
\alias{summary.all.variables}
\alias{summary.impl}
\alias{summary.continuous}
\title{Summarize a Data Frame}
\usage{
\method{summary}{all.variables}(data, summary.default = summary.continuous, ...)

\method{summary}{impl}(
  fx,
  data = NULL,
  stat.n = F,
  stat.total.n = F,
  stat.miss = F,
  stat.sum = F,
  stat.mean = F,
  stat.var = F,
  stat.sd = F,
  stat.mean.ADA = F,
  stat.mean.ADM = F,
  stat.mean.ADMn1 = F,
  stat.quantiles = NULL,
  stat.five.number = F,
  stat.min = F,
  stat.q1 = F,
  stat.median = F,
  stat.q3 = F,
  stat.max = F,
  stat.range = F,
  stat.iqr = F,
  stat.psd = F,
  stat.sir = F,
  stat.coefvar = F,
  stat.distinct = F,
  stat.distinct.withna = F,
  stat.true.mode = F,
  stat.shape.rejection.conf.level = NA,
  stat.shape.text.rej = "Reject",
  stat.shape.text.ftr = "",
  stat.ad.test = 0,
  stat.sw.test = 0,
  stat.skew.test = 0,
  stat.kurt.test = 0,
  stat.dago.test = 0,
  stat.pois.dist.test = F,
  stat.sw.exp.test = F,
  stat.sd.report = NULL,
  stat.lsl = NA,
  stat.target = NA,
  stat.usl = NA,
  stat.nonconform.nbelow = F,
  stat.nonconform.nabove = F,
  stat.nonconform.nout = F,
  stat.nonconform.pbelow = F,
  stat.nonconform.pabove = F,
  stat.nonconform.pout = F,
  format.generate.cellcodes = F,
  ...
)
}
\arguments{
\item{data}{Data Frame - Data frame to be summarized using names specified in fx}

\item{...}{Additional Parameters - additional parameters for summary.impl}

\item{fx}{Formula - variable to be summarized and grouping variables, usually like "dependent.var ~ group1 + group2 + group3 + ...""}

\item{stat.n}{Logical - return non-NA counts}

\item{stat.total.n}{Logical - return subgroup counts including NAs}

\item{stat.miss}{Logical - return NA counts in subgoups}

\item{stat.sum}{Logical - return subgroup sum}

\item{stat.mean}{Logical - return subgroup mean}

\item{stat.var}{Logical - return subgroup sample variance}

\item{stat.sd}{Logical - return subgroup sample standard deviation}

\item{stat.mean.ADA}{Logical - return subgroup mean absolute deviation}

\item{stat.mean.ADM}{Logical - return subgroup median absolute deviation}

\item{stat.mean.ADMn1}{Logical - return subgroup median absolute deviation with midpoint removed, see also dispersion.ADMn1}

\item{stat.quantiles}{Vector - return quantiles (input is values between 0 and 1 indicating quantiles, ex .25 for first quartile)}

\item{stat.five.number}{Logical - return subgroup five number summary (min, Q1, median, Q3, max)}

\item{stat.min}{Logical - return subgroup minimum}

\item{stat.q1}{Logical - return subgroup 1st quartile (25th percentile)}

\item{stat.median}{Logical - return subgroup median (50th percentile)}

\item{stat.q3}{Logical - return subgroup third quartile (75th percentile)}

\item{stat.max}{Logical - return subgroup maximum}

\item{stat.range}{Logical - return subgroup range (max - min)}

\item{stat.iqr}{Logical - return subgroup interquartile range, abbreviated IQR, defined as (q3 - q1)}

\item{stat.psd}{Logical - return subgroup pseudo standard deviation (IQR / 1.35)}

\item{stat.sir}{Logical - return subgroup semi-interquartile range (IQR / 2)}

\item{stat.coefvar}{Logical - return subgroup coefficient of variation (standard deviation / mean)}

\item{stat.distinct}{Logical - return subgroup count of distinct values excluding NA}

\item{stat.distinct.withna}{Logical - return subgroup count of distinct values including NA as a distinct value}

\item{stat.true.mode}{Logical - return subgroup "true mode", defined as  3*median-2*mean}

\item{stat.shape.rejection.conf.level}{Numeric - confidence level for rejection for shape tests}

\item{stat.shape.text.rej}{Character - Text to identify rejected null hypothesis for shape test}

\item{stat.shape.text.ftr}{Character - Text to identify null hypothesis not rejected for shape test}

\item{stat.ad.test}{Numeric - Return Anderson Darling test for normality (0=off, 1=if n<25, 2=on)}

\item{stat.sw.test}{Numeric - Return Shapiro Wilk test for normality (0=off, 1=if n<25, 2=on)}

\item{stat.skew.test}{Numeric - Return D'Agostino test for skewness (normality) (0=off, 1=if n>=20, 2=on)}

\item{stat.kurt.test}{Numeric - Return D'Agostino test for kurtosis (normality) (0=off, 1=if n>=20, 2=on)}

\item{stat.dago.test}{Numeric - Return D'Agostino omnibus test for normality (0=off, 1=if n>=20, 2=on)}

\item{stat.pois.dist.test}{Logical - Return test for Poisson distribution}

\item{stat.sw.exp.test}{Logical - Return Shapiro-Wilk exponentiality test}

\item{stat.sd.report}{Vector - return multiples of subgroup standard deviation}

\item{stat.lsl}{Numeric - Lower specification limit}

\item{stat.target}{Numeric - Target value for distribution}

\item{stat.usl}{Numeric - Upper specification limit}

\item{stat.nonconform.nbelow}{Logical - Return count of subgroup values below lower specification limit}

\item{stat.nonconform.nabove}{Logical - Return count of subgroup values above upper specification limit}

\item{stat.nonconform.nout}{Logical - Return count of subgroup values outside of specification limits}

\item{stat.nonconform.pbelow}{Logical - Return percentage of subgroup values below lower specification limit}

\item{stat.nonconform.pabove}{Logical - Return percentage of subgroup values above upper specification limit}

\item{stat.nonconform.pout}{Logical - Return percentage of subgroup values outside of specification limits}

\item{format.generate.cellcodes}{Logical - Calculate/return group cell codes}
}
\value{
A data frame with subgroups and selected measures
}
\description{
Calculate various summary measures for subgroups defined in a data frame by defining a dependent variable and one or more grouping variables.
}
