% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.dist.binomial.R
\name{table.dist.binomial}
\alias{table.dist.binomial}
\title{Table of Binomial Distribution}
\usage{
table.dist.binomial(n = 10, p = 0.5)
}
\arguments{
\item{n}{Scalar - sample size}

\item{p}{Scalar - probability of "success"}
}
\value{
A data frame with x, probability at x, and cumulative probability from each tail to x.
}
\description{
Calculates probabilities for the binomial distribution and includes cumulative probabilities from each tail.
}
