% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.dist.geometric.R
\name{table.dist.geometric}
\alias{table.dist.geometric}
\title{Table of Geometric Distribution}
\usage{
table.dist.geometric(p, max.x = ceiling(4 * (1 - p)/p^2))
}
\arguments{
\item{p}{Scalar - probability of "success"}

\item{max.x}{Scalar - maximum value of x to include (this is an infinite distribution)}
}
\value{
A data frame with x, probability at x, and cumulative probability from each tail to x.
}
\description{
Calculates probabilities for the geometric distribution and includes cumulative probabilities from each tail.
}
