% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.dist.hypergeometric.R
\name{table.dist.hypergeometric}
\alias{table.dist.hypergeometric}
\title{Table of Hypergeometric Distribution}
\usage{
table.dist.hypergeometric(pop.success.count, total.count, sample.size)
}
\arguments{
\item{pop.success.count}{Scalar - Total "successes" in population}

\item{total.count}{Scalar - Total population size}

\item{sample.size}{Scaler - How many items are sampled}
}
\value{
A data frame with x, probability at x, and cumulative probability from each tail to x.
}
\description{
Calculates probabilities for the hypergeometric distribution and includes cumulative probabilities from each tail.
}
