% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.dist.poisson.R
\name{table.dist.poisson}
\alias{table.dist.poisson}
\title{Table of Poisson Distribution}
\usage{
table.dist.poisson(
  lambda = 5,
  include.lambda = ifelse(abs(1 - lambda) < 0.3, max(10, 10/lambda), max(5, 5/lambda)),
  include.x = ceiling(include.lambda * lambda)
)
}
\arguments{
\item{lambda}{Scalar - mean of poisson distribution}

\item{include.lambda}{Scalar - how many lambdas of x to include in table}

\item{include.x}{Scalar - maximum value of x to include}
}
\value{
A data frame with x, probability at x, and cumulative probability from each tail to x.
}
\description{
Calculates probabilities for the poisson distribution and includes cumulative probabilities from each tail.
}
