% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true.mode.R
\name{true.mode}
\alias{true.mode}
\title{True Mode}
\usage{
true.mode(x)
}
\arguments{
\item{x}{Vector - data to calculate "true mode" for}
}
\value{
Scalar result of true mode calculation.
}
\description{
Calculates the "true mode" for the data, defined as \eqn{3\tilde{x} - 2\bar{x}}{3*median(x)-2*mean(x)}
}
