% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.acre.to.square.mile.R
\name{unitconversion.acre.to.square.mile}
\alias{unitconversion.acre.to.square.mile}
\alias{unitconversion.ac.to.sq.mi}
\title{Unit Conversion - Area - Square Mile to Acre}
\usage{
unitconversion.acre.to.square.mile(x = 1)

unitconversion.ac.to.sq.mi(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of square miles}
}
\value{
x, but converted to acres
}
\description{
Performs a conversion of areas from square miles (U.S. Survey) to acres.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
