% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.astronomical.unit.to.meter.R
\name{unitconversion.astronomical.unit.to.meter}
\alias{unitconversion.astronomical.unit.to.meter}
\alias{unitconversion.au.to.m}
\title{Unit Conversion - Length - Astronomical Unit to Meter}
\usage{
unitconversion.astronomical.unit.to.meter(x = 1)

unitconversion.au.to.m(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of astronomical units}
}
\value{
x, but converted to meters
}
\description{
Performs a conversion of length from astronomical units to meters.
}
\references{
Thompson, Ambler and Taylor, Barry. NIST Special Publication 811 - Guide for theUse of the International System of Units (SI). 2008.  
https://nvlpubs.nist.gov/nistpubs/Legacy/SP/nistspecialpublication811e2008.pdf
}
