% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.bushel.to.peck.R
\name{unitconversion.bushel.to.peck}
\alias{unitconversion.bushel.to.peck}
\alias{unitconversion.bu.to.pk}
\title{Unit Conversion - Dry Volume - Bushel to Peck}
\usage{
unitconversion.bushel.to.peck(x = 1)

unitconversion.bu.to.pk(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of bushels}
}
\value{
x, but converted to pecks
}
\description{
Performs a conversion of volumes from bushels to pecks.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
