% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.carat.to.milligram.R
\name{unitconversion.carat.to.milligram}
\alias{unitconversion.carat.to.milligram}
\alias{unitconversion.ct.to.mg}
\title{Unit Conversion - Mass/Weight - Carat to Milligram}
\usage{
unitconversion.carat.to.milligram(x = 1)

unitconversion.ct.to.mg(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of carats}
}
\value{
x, but converted to milligrams
}
\description{
Performs a conversion of weights from carats to milligrams.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
