% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.carat.to.ounce.R
\name{unitconversion.carat.to.ounce}
\alias{unitconversion.carat.to.ounce}
\alias{unitconversion.ct.to.oz}
\title{Unit Conversion - Mass/Weight - Carat to Avoirdupois Ounce}
\usage{
unitconversion.carat.to.ounce(x = 1)

unitconversion.ct.to.oz(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of carats}
}
\value{
x, but converted to ounces
}
\description{
Performs a conversion of weights from carats to ounces (avoirdupois).
}
