% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.chain.to.link.R
\name{unitconversion.chain.to.link}
\alias{unitconversion.chain.to.link}
\alias{unitconversion.ch.to.li}
\title{Unit Conversion - Length - Chain to Link}
\usage{
unitconversion.chain.to.link(x = 1)

unitconversion.ch.to.li(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of chains}
}
\value{
x, but converted to links
}
\description{
Performs a conversion of lengths from chains to links (surveyor's chain).
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
