% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.cubic.inch.to.bushel.R
\name{unitconversion.cubic.inch.to.bushel}
\alias{unitconversion.cubic.inch.to.bushel}
\alias{unitconversion.cu.in.to.bu}
\title{Unit Conversion - Dry Volume - Cubic Inch to Bushel}
\usage{
unitconversion.cubic.inch.to.bushel(x = 1)

unitconversion.cu.in.to.bu(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of cubic inches}
}
\value{
x, but converted to bushels
}
\description{
Performs a conversion of volumes from cubic inches to bushels.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
