% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.cubic.inch.to.gallon.R
\name{unitconversion.cubic.inch.to.gallon}
\alias{unitconversion.cubic.inch.to.gallon}
\alias{unitconversion.cu.in.to.gal}
\title{Unit Conversion - Liquid Volume -  Cubic Inch to Gallon}
\usage{
unitconversion.cubic.inch.to.gallon(x = 1)

unitconversion.cu.in.to.gal(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of cubic inches}
}
\value{
x, but converted to gallons
}
\description{
Performs a conversion of volumes from cubic inches to gallons.
}
