% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/unitconversion.degree.celsius.to.degree.kelvin.R
\name{unitconversion.degree.celsius.to.degree.kelvin}
\alias{unitconversion.degree.celsius.to.degree.kelvin}
\alias{unitconversion.deg.C.to.deg.K}
\title{Unit Conversion - Temperature - Celsius to Kelvin}
\usage{
unitconversion.degree.celsius.to.degree.kelvin(x = 1)

unitconversion.deg.C.to.deg.K(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of degrees Celsius.}
}
\value{
x, but converted to degrees Kelvin.
}
\description{
Performs a conversion of temperature from Celsius to Kelvin.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
