% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/unitconversion.degree.fahrenheit.to.degree.kelvin.R
\name{unitconversion.degree.fahrenheit.to.degree.kelvin}
\alias{unitconversion.degree.fahrenheit.to.degree.kelvin}
\alias{unitconversion.deg.F.to.deg.K}
\title{Unit Conversion - Temperature - Fahrenheit to Kelvin}
\usage{
unitconversion.degree.fahrenheit.to.degree.kelvin(x = 1)

unitconversion.deg.F.to.deg.K(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of degrees Fahrenheit.}
}
\value{
x, but converted to degrees Kelvin
}
\description{
Performs a conversion of temperature from Fahrenheit to Kelvin.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
