% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/unitconversion.degree.kelvin.to.degree.fahrenheit.R
\name{unitconversion.degree.kelvin.to.degree.fahrenheit}
\alias{unitconversion.degree.kelvin.to.degree.fahrenheit}
\alias{unitconversion.deg.K.to.deg.F}
\title{Unit Conversion - Temperature - Kelvin to Fahrenheit}
\usage{
unitconversion.degree.kelvin.to.degree.fahrenheit(x = 1)

unitconversion.deg.K.to.deg.F(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of degrees Kelvin.}
}
\value{
x, but converted to degrees Fahrenheit.
}
\description{
Performs a conversion of temperature from Kelvin to Fahrenheit.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
