% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.degree.to.degree.minute.R
\name{unitconversion.degree.to.degree.minute}
\alias{unitconversion.degree.to.degree.minute}
\title{Unit Conversion - Angles - Degrees to Degrees and Minutes}
\usage{
unitconversion.degree.to.degree.minute(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of degrees}
}
\value{
if x is length 1, then a single list, otherwise a nested list for each value of x, names for 
degrees and minutes. Sign of resulting values matches sign of input.
}
\description{
Performs a conversion of angle measures from degrees to degrees and minutes. 
Minutes (also called minutes of arc) are 1/60th of a degree.
Any remaining decimal will be included in the minutes portion of the resulting data structure.
}
