% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.dry.pint.to.cubic.inch.R
\name{unitconversion.dry.pint.to.cubic.inch}
\alias{unitconversion.dry.pint.to.cubic.inch}
\alias{unitconversion.dry.pt.to.cu.in}
\title{Unit Conversion - Dry Volume - Dry Pint to Cubic Inch}
\usage{
unitconversion.dry.pint.to.cubic.inch(x = 1)

unitconversion.dry.pt.to.cu.in(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pints}
}
\value{
x, but converted to cubic inches
}
\description{
Performs a conversion of volumes from dry pints to cubic inches.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
