% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.dry.pint.to.dry.quart.R
\name{unitconversion.dry.pint.to.dry.quart}
\alias{unitconversion.dry.pint.to.dry.quart}
\alias{unitconversion.dry.pt.to.dry.qt}
\title{Unit Conversion - Dry Volume - Dry Pint to Dry Quart}
\usage{
unitconversion.dry.pint.to.dry.quart(x = 1)

unitconversion.dry.pt.to.dry.qt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pints}
}
\value{
x, but converted to quarts
}
\description{
Performs a conversion of volumes from dry pints to dry quarts.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
