% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.dry.quart.to.dry.pint.R
\name{unitconversion.dry.quart.to.dry.pint}
\alias{unitconversion.dry.quart.to.dry.pint}
\alias{unitconversion.dry.qt.to.dry.pt}
\title{Unit Conversion - Dry Volume - Dry Quart to Dry Pint}
\usage{
unitconversion.dry.quart.to.dry.pint(x = 1)

unitconversion.dry.qt.to.dry.pt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of quarts}
}
\value{
x, but converted to pints
}
\description{
Performs a conversion of volumes from dry quarts to dry pints.
}
