% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.fluid.ounce.to.cubic.inch.R
\name{unitconversion.fluid.ounce.to.cubic.inch}
\alias{unitconversion.fluid.ounce.to.cubic.inch}
\alias{unitconversion.fl.oz.to.cu.in}
\title{Unit Conversion - Liquid Volume - Fluid Ounce to Cubic Inch}
\usage{
unitconversion.fluid.ounce.to.cubic.inch(x = 1)

unitconversion.fl.oz.to.cu.in(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of fluid ounces}
}
\value{
x, but converted to cubic inches
}
\description{
Performs a conversion of volumes from fluid ounces to cubic inches.
}
