% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.fluid.ounce.to.fluid.dram.R
\name{unitconversion.fluid.ounce.to.fluid.dram}
\alias{unitconversion.fluid.ounce.to.fluid.dram}
\alias{unitconversion.fl.oz.to.fl.dr}
\title{Unit Conversion - Liquid Volume - Fluid Ounce to Fluid Dram}
\usage{
unitconversion.fluid.ounce.to.fluid.dram(x = 1)

unitconversion.fl.oz.to.fl.dr(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of fluid ounces}
}
\value{
x, but converted to fluid drams
}
\description{
Performs a conversion of volumes from fluid ounces to fluid drams.
}
