% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.fluid.ounce.to.pint.R
\name{unitconversion.fluid.ounce.to.pint}
\alias{unitconversion.fluid.ounce.to.pint}
\alias{unitconversion.fl.oz.to.pt}
\title{Unit Conversion - Liquid Volume - Fluid Ounce to Pint}
\usage{
unitconversion.fluid.ounce.to.pint(x = 1)

unitconversion.fl.oz.to.pt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of fluid ounces}
}
\value{
x, but converted to pints
}
\description{
Performs a conversion of volumes from fluid ounces to pints.
}
