% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.foot.to.rod.R
\name{unitconversion.foot.to.rod}
\alias{unitconversion.foot.to.rod}
\alias{unitconversion.ft.to.rd}
\alias{unitconversion.foot.to.pole}
\alias{unitconversion.foot.to.perch}
\title{Unit Conversion - Length - Foot to Rod}
\usage{
unitconversion.foot.to.rod(x = 1)

unitconversion.ft.to.rd(x = 1)

unitconversion.foot.to.pole(x = 1)

unitconversion.foot.to.perch(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of feet (survey feet)}
}
\value{
x, but converted to rods
}
\description{
Performs a conversion of lengths from feet to rods (also called poles and perches).
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
