% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.foot.to.yard.R
\name{unitconversion.foot.to.yard}
\alias{unitconversion.foot.to.yard}
\alias{unitconversion.ft.to.yd}
\title{Unit Conversion - Length - Foot to Yard}
\usage{
unitconversion.foot.to.yard(x = 1)

unitconversion.ft.to.yd(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of feet}
}
\value{
x, but converted to yards
}
\description{
Performs a conversion of lengths from feet to yards.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
