% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.furlong.to.mile.R
\name{unitconversion.furlong.to.mile}
\alias{unitconversion.furlong.to.mile}
\alias{unitconversion.fur.to.mi}
\title{Unit Conversion - Length - Furlong to Mile}
\usage{
unitconversion.furlong.to.mile(x = 1)

unitconversion.fur.to.mi(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of furlongs}
}
\value{
x, but converted to miles
}
\description{
Performs a conversion of lengths from furlongs to miles (U.S. statute/survey miles).
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
