% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.furlong.to.rod.R
\name{unitconversion.furlong.to.rod}
\alias{unitconversion.furlong.to.rod}
\alias{unitconversion.fur.to.rd}
\alias{unitconversion.furlong.to.pole}
\alias{unitconversion.furlong.to.perch}
\title{Unit Conversion - Length - Furlong to Rod}
\usage{
unitconversion.furlong.to.rod(x = 1)

unitconversion.fur.to.rd(x = 1)

unitconversion.furlong.to.pole(x = 1)

unitconversion.furlong.to.perch(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of furlongs}
}
\value{
x, but converted to rods
}
\description{
Performs a conversion of lengths from furlongs to rods (also called poles and perches).
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
