% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.gallon.to.cubic.inch.R
\name{unitconversion.gallon.to.cubic.inch}
\alias{unitconversion.gallon.to.cubic.inch}
\alias{unitconversion.gal.to.cu.in}
\title{Unit Conversion - Liquid Volume -  Gallon to Cubic Inch}
\usage{
unitconversion.gallon.to.cubic.inch(x = 1)

unitconversion.gal.to.cu.in(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of gallons}
}
\value{
x, but converted to cubic inches
}
\description{
Performs a conversion of volumes from gallons to cubic inches.
}
