% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.gallon.to.liter.R
\name{unitconversion.gallon.to.liter}
\alias{unitconversion.gallon.to.liter}
\alias{unitconversion.gal.to.L}
\title{Unit Conversion - Liquid Volume -  Gallon to Liter}
\usage{
unitconversion.gallon.to.liter(x = 1)

unitconversion.gal.to.L(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of gallons}
}
\value{
x, but converted to liters
}
\description{
Performs a conversion of volumes from gallons to liters.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
