% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.gallon.to.quart.R
\name{unitconversion.gallon.to.quart}
\alias{unitconversion.gallon.to.quart}
\alias{unitconversion.gal.to.qt}
\title{Unit Conversion - Liquid Volume -  Gallon to Quart}
\usage{
unitconversion.gallon.to.quart(x = 1)

unitconversion.gal.to.qt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of gallons}
}
\value{
x, but converted to quarts
}
\description{
Performs a conversion of volumes from gallons to quarts.
}
