% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.gill.to.pint.R
\name{unitconversion.gill.to.pint}
\alias{unitconversion.gill.to.pint}
\alias{unitconversion.gi.to.pt}
\title{Unit Conversion - Liquid Volume - Gill to Pint}
\usage{
unitconversion.gill.to.pint(x = 1)

unitconversion.gi.to.pt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of gills}
}
\value{
x, but converted to pints
}
\description{
Performs a conversion of volumes from gills to pints.
}
