% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.grain.to.ounce.R
\name{unitconversion.grain.to.ounce}
\alias{unitconversion.grain.to.ounce}
\alias{unitconversion.gr.to.oz}
\title{Unit Conversion - Mass/Weight - Grain to Avoirdupois Ounce}
\usage{
unitconversion.grain.to.ounce(x = 1)

unitconversion.gr.to.oz(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of grains}
}
\value{
x, but converted to ounces
}
\description{
Performs a conversion of weights from grains to ounces (avoirdupois).
}
