% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.grain.to.pennyweight.R
\name{unitconversion.grain.to.pennyweight}
\alias{unitconversion.grain.to.pennyweight}
\alias{unitconversion.gr.to.dwt}
\title{Unit Conversion - Mass/Weight - Grain to Troy Pennyweight}
\usage{
unitconversion.grain.to.pennyweight(x = 1)

unitconversion.gr.to.dwt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of grains}
}
\value{
x, but converted to pennyweight
}
\description{
Performs a conversion of weights from grains to pennyweights.
}
