% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.grain.to.pound.R
\name{unitconversion.grain.to.pound}
\alias{unitconversion.grain.to.pound}
\alias{unitconversion.gr.to.lb}
\title{Unit Conversion - Mass/Weight - Grain to U.S. Pound}
\usage{
unitconversion.grain.to.pound(x = 1)

unitconversion.gr.to.lb(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of grains}
}
\value{
x, but converted to pounds
}
\description{
Performs a conversion of weights from grains to pounds (avoirdupois).
}
