% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.grain.to.troy.ounce.R
\name{unitconversion.grain.to.troy.ounce}
\alias{unitconversion.grain.to.troy.ounce}
\alias{unitconversion.gr.to.ozt}
\title{Unit Conversion - Mass/Weight - Grain to Troy Ounce}
\usage{
unitconversion.grain.to.troy.ounce(x = 1)

unitconversion.gr.to.ozt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of grains}
}
\value{
x, but converted to Troy ounces
}
\description{
Performs a conversion of weights from grains to troy ounces.
}
