% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.grain.to.troy.pound.R
\name{unitconversion.grain.to.troy.pound}
\alias{unitconversion.grain.to.troy.pound}
\alias{unitconversion.gr.to.lbt}
\title{Unit Conversion - Mass/Weight - Grain to Troy Pound}
\usage{
unitconversion.grain.to.troy.pound(x = 1)

unitconversion.gr.to.lbt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of grains}
}
\value{
x, but converted to Troy pounds
}
\description{
Performs a conversion of weights from grains to pounds (Troy).
}
