% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.hand.to.inch.R
\name{unitconversion.hand.to.inch}
\alias{unitconversion.hand.to.inch}
\title{Unit Conversion - Length - Hand to Inch}
\usage{
unitconversion.hand.to.inch(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of hands}
}
\value{
x, but converted to inches
}
\description{
Performs a conversion of lengths from hands to inches.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
