% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.hectare.to.square.meter.R
\name{unitconversion.hectare.to.square.meter}
\alias{unitconversion.hectare.to.square.meter}
\alias{unitconversion.ha.to.sq.m}
\title{Unit Conversion - Area - Hectare to Square Meter}
\usage{
unitconversion.hectare.to.square.meter(x = 1)

unitconversion.ha.to.sq.m(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of hectares}
}
\value{
x, but converted to square meters
}
\description{
Performs a conversion of areas from hectares to square meters.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
