% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.inch.to.centimeter.R
\name{unitconversion.inch.to.centimeter}
\alias{unitconversion.inch.to.centimeter}
\alias{unitconversion.in.to.cm}
\title{Unit Conversion - Length - Inch to Centimeter}
\usage{
unitconversion.inch.to.centimeter(x = 1)

unitconversion.in.to.cm(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of inches}
}
\value{
x, but converted to centimeters
}
\description{
Performs a conversion of lengths from inches to centimeters.
}
\references{
NIST. Approximate Conversions from U.S. Customary Measures to Metric. 2022. Accessed 3/16/2022. 
https://www.nist.gov/pml/weights-and-measures/approximate-conversions-us-customary-measures-metric
}
