% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.inch.to.foot.R
\name{unitconversion.inch.to.foot}
\alias{unitconversion.inch.to.foot}
\alias{unitconversion.in.to.ft}
\title{Unit Conversion - Length - Inch to Foot}
\usage{
unitconversion.inch.to.foot(x = 1)

unitconversion.in.to.ft(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of inches}
}
\value{
x, but converted to feet
}
\description{
Performs a conversion of lengths from inches to feet.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
