% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.inch.to.mil.R
\name{unitconversion.inch.to.mil}
\alias{unitconversion.inch.to.mil}
\alias{unitconversion.in.to.mil}
\title{Unit Conversion - Thickness - Inch to Mil}
\usage{
unitconversion.inch.to.mil(x = 1)

unitconversion.in.to.mil(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of inches}
}
\value{
x, but converted to mils
}
\description{
Performs a conversion of lengths from inches to mils.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
