% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.joule.to.therm.us.R
\name{unitconversion.joule.to.therm.us}
\alias{unitconversion.joule.to.therm.us}
\title{Unit Conversion - Energy - Joule to U.S. Therm}
\usage{
unitconversion.joule.to.therm.us(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of joules}
}
\value{
x, but converted to therms
}
\description{
Performs a conversion of energy from joules to U.S. therms.
}
\references{
Thompson, Ambler and Taylor, Barry. NIST Special Publication 811 - Guide for theUse of the International System of Units (SI). 2008.  
https://nvlpubs.nist.gov/nistpubs/Legacy/SP/nistspecialpublication811e2008.pdf
}
