% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.kilogram.to.metric.ton.R
\name{unitconversion.kilogram.to.metric.ton}
\alias{unitconversion.kilogram.to.metric.ton}
\title{Unit Conversion - Mass/Weight - Kilogram to Metric Ton}
\usage{
unitconversion.kilogram.to.metric.ton(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of kilograms}
}
\value{
x, but converted to metric tons
}
\description{
Performs a conversion of weights from kilograms to metric tons.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
