% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.kilometer.to.mile.R
\name{unitconversion.kilometer.to.mile}
\alias{unitconversion.kilometer.to.mile}
\alias{unitconversion.km.to.mi}
\title{Unit Conversion - Length - Kilometer to Miles}
\usage{
unitconversion.kilometer.to.mile(x = 1)

unitconversion.km.to.mi(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of kilometers}
}
\value{
x, but converted to miles
}
\description{
Performs a conversion of lengths from kilometers to miles (international).
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
