% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.link.to.chain.R
\name{unitconversion.link.to.chain}
\alias{unitconversion.link.to.chain}
\alias{unitconversion.li.to.ch}
\title{Unit Conversion - Length - Link to Chain}
\usage{
unitconversion.link.to.chain(x = 1)

unitconversion.li.to.ch(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of links}
}
\value{
x, but converted to chains
}
\description{
Performs a conversion of lengths from links (surveyor's chain) to chains.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
