% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.link.to.foot.R
\name{unitconversion.link.to.foot}
\alias{unitconversion.link.to.foot}
\alias{unitconversion.li.to.ft}
\title{Unit Conversion - Length - Link to Foot}
\usage{
unitconversion.link.to.foot(x = 1)

unitconversion.li.to.ft(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of links}
}
\value{
x, but converted to feet
}
\description{
Performs a conversion of lengths from links (surveyor's chain) to feet (U.S. survey).
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
