% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.meter.to.astronomical.unit.R
\name{unitconversion.meter.to.astronomical.unit}
\alias{unitconversion.meter.to.astronomical.unit}
\alias{unitconversion.m.to.au}
\title{Unit Conversion - Length - Meter to Astronomical Unit}
\usage{
unitconversion.meter.to.astronomical.unit(x = 1)

unitconversion.m.to.au(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of meters}
}
\value{
x, but converted to astronomical units
}
\description{
Performs a conversion of length from meters to astronomical units.
}
\references{
Thompson, Ambler and Taylor, Barry. NIST Special Publication 811 - Guide for theUse of the International System of Units (SI). 2008.  
https://nvlpubs.nist.gov/nistpubs/Legacy/SP/nistspecialpublication811e2008.pdf
}
