% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.meter.to.nautical.mile.R
\name{unitconversion.meter.to.nautical.mile}
\alias{unitconversion.meter.to.nautical.mile}
\alias{unitconversion.m.to.nmi}
\title{Unit Conversion - Length -  Meter to Nautical Mile}
\usage{
unitconversion.meter.to.nautical.mile(x = 1)

unitconversion.m.to.nmi(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of meters}
}
\value{
x, but converted to nautical miles
}
\description{
Performs a conversion of lengths from meters to nautical miles.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
