% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.metric.ton.to.kilogram.R
\name{unitconversion.metric.ton.to.kilogram}
\alias{unitconversion.metric.ton.to.kilogram}
\title{Unit Conversion - Mass/Weight - Metric Ton to Kilogram}
\usage{
unitconversion.metric.ton.to.kilogram(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of metric tons}
}
\value{
x, but converted to kilograms
}
\description{
Performs a conversion of weights from metric tons to kilograms.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
