% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.micrometer.to.mil.R
\name{unitconversion.micrometer.to.mil}
\alias{unitconversion.micrometer.to.mil}
\alias{unitconversion.um.to.mil}
\title{Unit Conversion - Thickness - Micrometer to Mil}
\usage{
unitconversion.micrometer.to.mil(x = 1)

unitconversion.um.to.mil(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of micrometers}
}
\value{
x, but converted to mils
}
\description{
Performs a conversion of lengths from micrometers to mils. 
Note, the character 'u' is used in place of 'mu' in "micro-" SI unit abbreviations due to naming/keyboard limitations.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E - General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
