% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.mil.to.inch.R
\name{unitconversion.mil.to.inch}
\alias{unitconversion.mil.to.inch}
\alias{unitconversion.mil.to.in}
\title{Unit Conversion - Thickness - Mil to Inch}
\usage{
unitconversion.mil.to.inch(x = 1)

unitconversion.mil.to.in(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of mils}
}
\value{
x, but converted to inches
}
\description{
Performs a conversion of lengths from mils to inches.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
