% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.mil.to.micrometer.R
\name{unitconversion.mil.to.micrometer}
\alias{unitconversion.mil.to.micrometer}
\alias{unitconversion.mil.to.um}
\title{Unit Conversion - Thickness - Mil to Micrometer}
\usage{
unitconversion.mil.to.micrometer(x = 1)

unitconversion.mil.to.um(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of mils}
}
\value{
x, but converted to micrometers
}
\description{
Performs a conversion of lengths from mils to micrometers. 
Note, the character 'u' is used in place of 'mu' in "micro-" SI unit abbreviations due to naming/keyboard limitations.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E - General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
