% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.mil.to.millimeter.R
\name{unitconversion.mil.to.millimeter}
\alias{unitconversion.mil.to.millimeter}
\alias{unitconversion.mil.to.mm}
\title{Unit Conversion - Thickness - Mil to Millimeter}
\usage{
unitconversion.mil.to.millimeter(x = 1)

unitconversion.mil.to.mm(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of mils}
}
\value{
x, but converted to millimeters
}
\description{
Performs a conversion of lengths from mils to millimeters.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
