% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.mile.to.furlong.R
\name{unitconversion.mile.to.furlong}
\alias{unitconversion.mile.to.furlong}
\alias{unitconversion.mi.to.fur}
\title{Unit Conversion - Length - Mile to Furlong}
\usage{
unitconversion.mile.to.furlong(x = 1)

unitconversion.mi.to.fur(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of miles}
}
\value{
x, but converted to furlongs
}
\description{
Performs a conversion of lengths from miles (U.S. statute/survey miles) to furlongs.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
