% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.mile.to.kilometer.R
\name{unitconversion.mile.to.kilometer}
\alias{unitconversion.mile.to.kilometer}
\alias{unitconversion.mi.to.km}
\title{Unit Conversion - Length - Mile to Kilometer}
\usage{
unitconversion.mile.to.kilometer(x = 1)

unitconversion.mi.to.km(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of miles}
}
\value{
x, but converted to kilometers
}
\description{
Performs a conversion of lengths from miles to kilometers.
}
\references{
NIST. Approximate Conversions from U.S. Customary Measures to Metric. 2022. Accessed 3/16/2022. 
https://www.nist.gov/pml/weights-and-measures/approximate-conversions-us-customary-measures-metric
}
