% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.milligram.to.carat.R
\name{unitconversion.milligram.to.carat}
\alias{unitconversion.milligram.to.carat}
\alias{unitconversion.mg.to.ct}
\title{Unit Conversion - Mass/Weight - Milligram to Carat}
\usage{
unitconversion.milligram.to.carat(x = 1)

unitconversion.mg.to.ct(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of milligrams}
}
\value{
x, but converted to carats
}
\description{
Performs a conversion of weights from milligrams to carats.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
