% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.minim.to.fluid.dram.R
\name{unitconversion.minim.to.fluid.dram}
\alias{unitconversion.minim.to.fluid.dram}
\title{Unit Conversion - Liquid Volume - Minim to Fluid Dram}
\usage{
unitconversion.minim.to.fluid.dram(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of minims}
}
\value{
x, but converted to fluid drams
}
\description{
Performs a conversion of volumes from minims to fluid drams.
}
