% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.nautical.mile.to.meter.R
\name{unitconversion.nautical.mile.to.meter}
\alias{unitconversion.nautical.mile.to.meter}
\alias{unitconversion.nmi.to.m}
\title{Unit Conversion - Length - Nautical Mile to Meter}
\usage{
unitconversion.nautical.mile.to.meter(x = 1)

unitconversion.nmi.to.m(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of nautical miles}
}
\value{
x, but converted to meters
}
\description{
Performs a conversion of lengths from nautical miles to meters.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
