% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.ounce.to.carat.R
\name{unitconversion.ounce.to.carat}
\alias{unitconversion.ounce.to.carat}
\alias{unitconversion.oz.to.ct}
\title{Unit Conversion - Mass/Weight - U.S. Ounce to Carat}
\usage{
unitconversion.ounce.to.carat(x = 1)

unitconversion.oz.to.ct(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of ounces}
}
\value{
x, but converted to carats
}
\description{
Performs a conversion of weights from ounces  (avoirdupois) to carats.
}
