% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.ounce.to.grain.R
\name{unitconversion.ounce.to.grain}
\alias{unitconversion.ounce.to.grain}
\alias{unitconversion.oz.to.gr}
\title{Unit Conversion - Mass/Weight - U.S. Ounce to Grain}
\usage{
unitconversion.ounce.to.grain(x = 1)

unitconversion.oz.to.gr(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of ounces}
}
\value{
x, but converted to grains
}
\description{
Performs a conversion of weights from ounces (avoirdupois) to grains.
}
