% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.ounce.to.troy.ounce.R
\name{unitconversion.ounce.to.troy.ounce}
\alias{unitconversion.ounce.to.troy.ounce}
\alias{unitconversion.oz.to.ozt}
\title{Unit Conversion - Mass/Weight - U.S. Ounce to Troy Ounce}
\usage{
unitconversion.ounce.to.troy.ounce(x = 1)

unitconversion.oz.to.ozt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of ounces}
}
\value{
x, but converted to troy ounces
}
\description{
Performs a conversion of weights from ounces (avoirdupois) to troy ounces.
}
