% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.peck.to.dry.quart.R
\name{unitconversion.peck.to.dry.quart}
\alias{unitconversion.peck.to.dry.quart}
\alias{unitconversion.pk.to.dry.qt}
\title{Unit Conversion - Dry Volume - Peck to Dry Quart}
\usage{
unitconversion.peck.to.dry.quart(x = 1)

unitconversion.pk.to.dry.qt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pecks}
}
\value{
x, but converted to dry quarts
}
\description{
Performs a conversion of volumes from dry pecks to dry quarts.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
