% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.pennyweight.to.grain.R
\name{unitconversion.pennyweight.to.grain}
\alias{unitconversion.pennyweight.to.grain}
\alias{unitconversion.dwt.to.gr}
\title{Unit Conversion - Mass/Weight - Troy Pennyweight to Grain}
\usage{
unitconversion.pennyweight.to.grain(x = 1)

unitconversion.dwt.to.gr(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pennyweight}
}
\value{
x, but converted to grain
}
\description{
Performs a conversion of weights from troy pennyweights to grains.
}
