% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.pennyweight.to.troy.ounce.R
\name{unitconversion.pennyweight.to.troy.ounce}
\alias{unitconversion.pennyweight.to.troy.ounce}
\alias{unitconversion.dwt.to.ozt}
\title{Unit Conversion - Mass/Weight - Troy Pennyweight to Troy Ounce}
\usage{
unitconversion.pennyweight.to.troy.ounce(x = 1)

unitconversion.dwt.to.ozt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pennyweights}
}
\value{
x, but converted to troy ounces
}
\description{
Performs a conversion of weights from troy pennyweights to troy ounces.
}
