% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.pint.to.fluid.ounce.R
\name{unitconversion.pint.to.fluid.ounce}
\alias{unitconversion.pint.to.fluid.ounce}
\alias{unitconversion.pt.to.fl.oz}
\title{Unit Conversion - Liquid Volume - Pint to Fluid Ounce}
\usage{
unitconversion.pint.to.fluid.ounce(x = 1)

unitconversion.pt.to.fl.oz(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pints}
}
\value{
x, but converted to fluid ounces
}
\description{
Performs a conversion of volumes from pints to fluid ounces.
}
