% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.pint.to.gill.R
\name{unitconversion.pint.to.gill}
\alias{unitconversion.pint.to.gill}
\alias{unitconversion.pt.to.gi}
\title{Unit Conversion - Liquid Volume - Pint to Gill}
\usage{
unitconversion.pint.to.gill(x = 1)

unitconversion.pt.to.gi(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pints}
}
\value{
x, but converted to gills
}
\description{
Performs a conversion of volumes from pints to gills.
}
