% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.pint.to.quart.R
\name{unitconversion.pint.to.quart}
\alias{unitconversion.pint.to.quart}
\alias{unitconversion.pt.to.qt}
\title{Unit Conversion - Liquid Volume - Pint to Quart}
\usage{
unitconversion.pint.to.quart(x = 1)

unitconversion.pt.to.qt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pints}
}
\value{
x, but converted to quarts
}
\description{
Performs a conversion of volumes from pints to quarts.
}
