% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.pound.to.grain.R
\name{unitconversion.pound.to.grain}
\alias{unitconversion.pound.to.grain}
\alias{unitconversion.lb.to.gr}
\title{Unit Conversion - Mass/Weight - U.S. Pound to Grain}
\usage{
unitconversion.pound.to.grain(x = 1)

unitconversion.lb.to.gr(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pounds}
}
\value{
x, but converted to grains
}
\description{
Performs a conversion of weights from pounds (avoirdupois) to grains.
}
