% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.quart.to.cubic.inch.R
\name{unitconversion.quart.to.cubic.inch}
\alias{unitconversion.quart.to.cubic.inch}
\alias{unitconversion.qt.to.cu.in}
\title{Unit Conversion - Liquid Volume - Quart to Cubic Inch}
\usage{
unitconversion.quart.to.cubic.inch(x = 1)

unitconversion.qt.to.cu.in(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of quarts}
}
\value{
x, but converted to cubic inches
}
\description{
Performs a conversion of volumes from quarts to cubic inches.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
