% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.quart.to.gallon.R
\name{unitconversion.quart.to.gallon}
\alias{unitconversion.quart.to.gallon}
\alias{unitconversion.qt.to.gal}
\title{Unit Conversion - Liquid Volume - Quart to Gallon}
\usage{
unitconversion.quart.to.gallon(x = 1)

unitconversion.qt.to.gal(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of quarts}
}
\value{
x, but converted to gallons
}
\description{
Performs a conversion of volumes from quarts to gallons.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
