% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.quart.to.pint.R
\name{unitconversion.quart.to.pint}
\alias{unitconversion.quart.to.pint}
\alias{unitconversion.qt.to.pt}
\title{Unit Conversion - Liquid Volume - Quart to Pint}
\usage{
unitconversion.quart.to.pint(x = 1)

unitconversion.qt.to.pt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of quarts}
}
\value{
x, but converted to pints
}
\description{
Performs a conversion of volumes from quarts to pints.
}
