% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.radian.to.degree.R
\name{unitconversion.radian.to.degree}
\alias{unitconversion.radian.to.degree}
\title{Unit Conversion - Angles - Radians to Degrees}
\usage{
unitconversion.radian.to.degree(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of radians}
}
\value{
x, but converted to degrees
}
\description{
Performs a conversion of angle measures from radians to degrees.
}
