% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.rod.to.foot.R
\name{unitconversion.rod.to.foot}
\alias{unitconversion.rod.to.foot}
\alias{unitconversion.rd.to.ft}
\alias{unitconversion.pole.to.foot}
\alias{unitconversion.perch.to.foot}
\title{Unit Conversion - Length - Rod to Foot}
\usage{
unitconversion.rod.to.foot(x = 1)

unitconversion.rd.to.ft(x = 1)

unitconversion.pole.to.foot(x = 1)

unitconversion.perch.to.foot(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of rods}
}
\value{
x, but converted to feet
}
\description{
Performs a conversion of lengths from rods (also called poles and perches) to feet (survey feet).
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
