% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.rod.to.furlong.R
\name{unitconversion.rod.to.furlong}
\alias{unitconversion.rod.to.furlong}
\alias{unitconversion.rd.to.fur}
\alias{unitconversion.pole.to.furlong}
\alias{unitconversion.perch.to.furlong}
\title{Unit Conversion - Length - Rod to Furlong}
\usage{
unitconversion.rod.to.furlong(x = 1)

unitconversion.rd.to.fur(x = 1)

unitconversion.pole.to.furlong(x = 1)

unitconversion.perch.to.furlong(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of rods}
}
\value{
x, but converted to furlongs
}
\description{
Performs a conversion of lengths from rods (also called poles and perches) to furlongs.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
