% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.section.to.square.mile.R
\name{unitconversion.section.to.square.mile}
\alias{unitconversion.section.to.square.mile}
\title{Unit Conversion - Area - Section to Square Mile}
\usage{
unitconversion.section.to.square.mile(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of square miles or sections}
}
\value{
x since conversion factor is 1
}
\description{
Performs a conversion of areas from sections to square miles (U.S. Survey/Statute).
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
