% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.atto.to.base.R
\name{unitconversion.siprefix.atto.to.base}
\alias{unitconversion.siprefix.atto.to.base}
\alias{unitconversion.attosecond.to.second}
\alias{unitconversion.as.to.s}
\alias{unitconversion.attometer.to.meter}
\alias{unitconversion.am.to.m}
\alias{unitconversion.attogram.to.gram}
\alias{unitconversion.ag.to.g}
\alias{unitconversion.attoampere.to.ampere}
\alias{unitconversion.aA.to.A}
\alias{unitconversion.attokelvin.to.kelvin}
\alias{unitconversion.aK.to.K}
\alias{unitconversion.attomole.to.mole}
\alias{unitconversion.amol.to.mol}
\alias{unitconversion.attocandela.to.candela}
\alias{unitconversion.acd.to.cd}
\alias{unitconversion.attoradian.to.radian}
\alias{unitconversion.arad.to.rad}
\alias{unitconversion.attosteradian.to.steradian}
\alias{unitconversion.asr.to.sr}
\alias{unitconversion.attohertz.to.hertz}
\alias{unitconversion.aHz.to.Hz}
\alias{unitconversion.attonewton.to.newton}
\alias{unitconversion.aN.to.N}
\alias{unitconversion.attopascal.to.pascal}
\alias{unitconversion.aPa.to.Pa}
\alias{unitconversion.attojoule.to.joule}
\alias{unitconversion.aJ.to.J}
\alias{unitconversion.attowatt.to.watt}
\alias{unitconversion.aW.to.W}
\alias{unitconversion.attocoulomb.to.coulomb}
\alias{unitconversion.aC.to.C}
\alias{unitconversion.attovolt.to.volt}
\alias{unitconversion.aV.to.V}
\alias{unitconversion.attofarad.to.farad}
\alias{unitconversion.aF.to.F}
\alias{unitconversion.attoohm.to.ohm}
\alias{unitconversion.attosiemens.to.siemens}
\alias{unitconversion.aS.to.S}
\alias{unitconversion.attoweber.to.weber}
\alias{unitconversion.aWb.to.Wb}
\alias{unitconversion.attotesla.to.tesla}
\alias{unitconversion.aT.to.T}
\alias{unitconversion.attohenry.to.henry}
\alias{unitconversion.aH.to.H}
\alias{unitconversion.attolumen.to.lumen}
\alias{unitconversion.alm.to.lm}
\alias{unitconversion.attolux.to.lux}
\alias{unitconversion.alx.to.lx}
\alias{unitconversion.attobecquerel.to.becquerel}
\alias{unitconversion.aBq.to.Bq}
\alias{unitconversion.attogray.to.gray}
\alias{unitconversion.aGy.to.Gy}
\alias{unitconversion.attosievert.to.sievert}
\alias{unitconversion.aSv.to.Sv}
\alias{unitconversion.attokatal.to.katal}
\alias{unitconversion.akat.to.kat}
\title{Unit Conversion - SI Prefixes - Atto- to Base}
\usage{
unitconversion.siprefix.atto.to.base(x = 1)

unitconversion.attosecond.to.second(x = 1)

unitconversion.as.to.s(x = 1)

unitconversion.attometer.to.meter(x = 1)

unitconversion.am.to.m(x = 1)

unitconversion.attogram.to.gram(x = 1)

unitconversion.ag.to.g(x = 1)

unitconversion.attoampere.to.ampere(x = 1)

unitconversion.aA.to.A(x = 1)

unitconversion.attokelvin.to.kelvin(x = 1)

unitconversion.aK.to.K(x = 1)

unitconversion.attomole.to.mole(x = 1)

unitconversion.amol.to.mol(x = 1)

unitconversion.attocandela.to.candela(x = 1)

unitconversion.acd.to.cd(x = 1)

unitconversion.attoradian.to.radian(x = 1)

unitconversion.arad.to.rad(x = 1)

unitconversion.attosteradian.to.steradian(x = 1)

unitconversion.asr.to.sr(x = 1)

unitconversion.attohertz.to.hertz(x = 1)

unitconversion.aHz.to.Hz(x = 1)

unitconversion.attonewton.to.newton(x = 1)

unitconversion.aN.to.N(x = 1)

unitconversion.attopascal.to.pascal(x = 1)

unitconversion.aPa.to.Pa(x = 1)

unitconversion.attojoule.to.joule(x = 1)

unitconversion.aJ.to.J(x = 1)

unitconversion.attowatt.to.watt(x = 1)

unitconversion.aW.to.W(x = 1)

unitconversion.attocoulomb.to.coulomb(x = 1)

unitconversion.aC.to.C(x = 1)

unitconversion.attovolt.to.volt(x = 1)

unitconversion.aV.to.V(x = 1)

unitconversion.attofarad.to.farad(x = 1)

unitconversion.aF.to.F(x = 1)

unitconversion.attoohm.to.ohm(x = 1)

unitconversion.attosiemens.to.siemens(x = 1)

unitconversion.aS.to.S(x = 1)

unitconversion.attoweber.to.weber(x = 1)

unitconversion.aWb.to.Wb(x = 1)

unitconversion.attotesla.to.tesla(x = 1)

unitconversion.aT.to.T(x = 1)

unitconversion.attohenry.to.henry(x = 1)

unitconversion.aH.to.H(x = 1)

unitconversion.attolumen.to.lumen(x = 1)

unitconversion.alm.to.lm(x = 1)

unitconversion.attolux.to.lux(x = 1)

unitconversion.alx.to.lx(x = 1)

unitconversion.attobecquerel.to.becquerel(x = 1)

unitconversion.aBq.to.Bq(x = 1)

unitconversion.attogray.to.gray(x = 1)

unitconversion.aGy.to.Gy(x = 1)

unitconversion.attosievert.to.sievert(x = 1)

unitconversion.aSv.to.Sv(x = 1)

unitconversion.attokatal.to.katal(x = 1)

unitconversion.akat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of atto-units}
}
\value{
x, but converted to base units
}
\description{
Atto- quintillionths, 1/1 000 000 000 000 000 000, or 10^(-18)
}
\details{
Performs a conversion from atto-units to base units (ex. attograms to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
