% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.atto.R
\name{unitconversion.siprefix.base.to.atto}
\alias{unitconversion.siprefix.base.to.atto}
\alias{unitconversion.second.to.attosecond}
\alias{unitconversion.s.to.as}
\alias{unitconversion.meter.to.attometer}
\alias{unitconversion.m.to.am}
\alias{unitconversion.gram.to.attogram}
\alias{unitconversion.g.to.ag}
\alias{unitconversion.ampere.to.attoampere}
\alias{unitconversion.A.to.aA}
\alias{unitconversion.kelvin.to.attokelvin}
\alias{unitconversion.K.to.aK}
\alias{unitconversion.mole.to.attomole}
\alias{unitconversion.mol.to.amol}
\alias{unitconversion.candela.to.attocandela}
\alias{unitconversion.cd.to.acd}
\alias{unitconversion.radian.to.attoradian}
\alias{unitconversion.rad.to.arad}
\alias{unitconversion.steradian.to.attosteradian}
\alias{unitconversion.sr.to.asr}
\alias{unitconversion.hertz.to.attohertz}
\alias{unitconversion.Hz.to.aHz}
\alias{unitconversion.newton.to.attonewton}
\alias{unitconversion.N.to.aN}
\alias{unitconversion.pascal.to.attopascal}
\alias{unitconversion.Pa.to.aPa}
\alias{unitconversion.joule.to.attojoule}
\alias{unitconversion.J.to.aJ}
\alias{unitconversion.watt.to.attowatt}
\alias{unitconversion.W.to.aW}
\alias{unitconversion.coulomb.to.attocoulomb}
\alias{unitconversion.C.to.aC}
\alias{unitconversion.volt.to.attovolt}
\alias{unitconversion.V.to.aV}
\alias{unitconversion.farad.to.attofarad}
\alias{unitconversion.F.to.aF}
\alias{unitconversion.ohm.to.attoohm}
\alias{unitconversion.siemens.to.attosiemens}
\alias{unitconversion.S.to.aS}
\alias{unitconversion.weber.to.attoweber}
\alias{unitconversion.Wb.to.aWb}
\alias{unitconversion.tesla.to.attotesla}
\alias{unitconversion.T.to.aT}
\alias{unitconversion.henry.to.attohenry}
\alias{unitconversion.H.to.aH}
\alias{unitconversion.lumen.to.attolumen}
\alias{unitconversion.lm.to.alm}
\alias{unitconversion.lux.to.attolux}
\alias{unitconversion.lx.to.alx}
\alias{unitconversion.becquerel.to.attobecquerel}
\alias{unitconversion.Bq.to.aBq}
\alias{unitconversion.gray.to.attogray}
\alias{unitconversion.Gy.to.aGy}
\alias{unitconversion.sievert.to.attosievert}
\alias{unitconversion.Sv.to.aSv}
\alias{unitconversion.katal.to.attokatal}
\alias{unitconversion.kat.to.akat}
\title{Unit Conversion - SI Prefixes - Base to Atto-}
\usage{
unitconversion.siprefix.base.to.atto(x = 1)

unitconversion.second.to.attosecond(x = 1)

unitconversion.s.to.as(x = 1)

unitconversion.meter.to.attometer(x = 1)

unitconversion.m.to.am(x = 1)

unitconversion.gram.to.attogram(x = 1)

unitconversion.g.to.ag(x = 1)

unitconversion.ampere.to.attoampere(x = 1)

unitconversion.A.to.aA(x = 1)

unitconversion.kelvin.to.attokelvin(x = 1)

unitconversion.K.to.aK(x = 1)

unitconversion.mole.to.attomole(x = 1)

unitconversion.mol.to.amol(x = 1)

unitconversion.candela.to.attocandela(x = 1)

unitconversion.cd.to.acd(x = 1)

unitconversion.radian.to.attoradian(x = 1)

unitconversion.rad.to.arad(x = 1)

unitconversion.steradian.to.attosteradian(x = 1)

unitconversion.sr.to.asr(x = 1)

unitconversion.hertz.to.attohertz(x = 1)

unitconversion.Hz.to.aHz(x = 1)

unitconversion.newton.to.attonewton(x = 1)

unitconversion.N.to.aN(x = 1)

unitconversion.pascal.to.attopascal(x = 1)

unitconversion.Pa.to.aPa(x = 1)

unitconversion.joule.to.attojoule(x = 1)

unitconversion.J.to.aJ(x = 1)

unitconversion.watt.to.attowatt(x = 1)

unitconversion.W.to.aW(x = 1)

unitconversion.coulomb.to.attocoulomb(x = 1)

unitconversion.C.to.aC(x = 1)

unitconversion.volt.to.attovolt(x = 1)

unitconversion.V.to.aV(x = 1)

unitconversion.farad.to.attofarad(x = 1)

unitconversion.F.to.aF(x = 1)

unitconversion.ohm.to.attoohm(x = 1)

unitconversion.siemens.to.attosiemens(x = 1)

unitconversion.S.to.aS(x = 1)

unitconversion.weber.to.attoweber(x = 1)

unitconversion.Wb.to.aWb(x = 1)

unitconversion.tesla.to.attotesla(x = 1)

unitconversion.T.to.aT(x = 1)

unitconversion.henry.to.attohenry(x = 1)

unitconversion.H.to.aH(x = 1)

unitconversion.lumen.to.attolumen(x = 1)

unitconversion.lm.to.alm(x = 1)

unitconversion.lux.to.attolux(x = 1)

unitconversion.lx.to.alx(x = 1)

unitconversion.becquerel.to.attobecquerel(x = 1)

unitconversion.Bq.to.aBq(x = 1)

unitconversion.gray.to.attogray(x = 1)

unitconversion.Gy.to.aGy(x = 1)

unitconversion.sievert.to.attosievert(x = 1)

unitconversion.Sv.to.aSv(x = 1)

unitconversion.katal.to.attokatal(x = 1)

unitconversion.kat.to.akat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to atto-units
}
\description{
Atto- quintillionths, 1/1 000 000 000 000 000 000, or 10^(-18)
}
\details{
Performs a conversion from base units to atto-units (ex. grams to attograms).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
