% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.centi.R
\name{unitconversion.siprefix.base.to.centi}
\alias{unitconversion.siprefix.base.to.centi}
\alias{unitconversion.second.to.centisecond}
\alias{unitconversion.s.to.cs}
\alias{unitconversion.meter.to.centimeter}
\alias{unitconversion.m.to.cm}
\alias{unitconversion.gram.to.centigram}
\alias{unitconversion.g.to.cg}
\alias{unitconversion.ampere.to.centiampere}
\alias{unitconversion.A.to.cA}
\alias{unitconversion.kelvin.to.centikelvin}
\alias{unitconversion.K.to.cK}
\alias{unitconversion.mole.to.centimole}
\alias{unitconversion.mol.to.cmol}
\alias{unitconversion.candela.to.centicandela}
\alias{unitconversion.cd.to.ccd}
\alias{unitconversion.radian.to.centiradian}
\alias{unitconversion.rad.to.crad}
\alias{unitconversion.steradian.to.centisteradian}
\alias{unitconversion.sr.to.csr}
\alias{unitconversion.hertz.to.centihertz}
\alias{unitconversion.Hz.to.cHz}
\alias{unitconversion.newton.to.centinewton}
\alias{unitconversion.N.to.cN}
\alias{unitconversion.pascal.to.centipascal}
\alias{unitconversion.Pa.to.cPa}
\alias{unitconversion.joule.to.centijoule}
\alias{unitconversion.J.to.cJ}
\alias{unitconversion.watt.to.centiwatt}
\alias{unitconversion.W.to.cW}
\alias{unitconversion.coulomb.to.centicoulomb}
\alias{unitconversion.C.to.cC}
\alias{unitconversion.volt.to.centivolt}
\alias{unitconversion.V.to.cV}
\alias{unitconversion.farad.to.centifarad}
\alias{unitconversion.F.to.cF}
\alias{unitconversion.ohm.to.centiohm}
\alias{unitconversion.siemens.to.centisiemens}
\alias{unitconversion.S.to.cS}
\alias{unitconversion.weber.to.centiweber}
\alias{unitconversion.Wb.to.cWb}
\alias{unitconversion.tesla.to.centitesla}
\alias{unitconversion.T.to.cT}
\alias{unitconversion.henry.to.centihenry}
\alias{unitconversion.H.to.cH}
\alias{unitconversion.lumen.to.centilumen}
\alias{unitconversion.lm.to.clm}
\alias{unitconversion.lux.to.centilux}
\alias{unitconversion.lx.to.clx}
\alias{unitconversion.becquerel.to.centibecquerel}
\alias{unitconversion.Bq.to.cBq}
\alias{unitconversion.gray.to.centigray}
\alias{unitconversion.Gy.to.cGy}
\alias{unitconversion.sievert.to.centisievert}
\alias{unitconversion.Sv.to.cSv}
\alias{unitconversion.katal.to.centikatal}
\alias{unitconversion.kat.to.ckat}
\title{Unit Conversion - SI Prefixes - Base to Centi-}
\usage{
unitconversion.siprefix.base.to.centi(x = 1)

unitconversion.second.to.centisecond(x = 1)

unitconversion.s.to.cs(x = 1)

unitconversion.meter.to.centimeter(x = 1)

unitconversion.m.to.cm(x = 1)

unitconversion.gram.to.centigram(x = 1)

unitconversion.g.to.cg(x = 1)

unitconversion.ampere.to.centiampere(x = 1)

unitconversion.A.to.cA(x = 1)

unitconversion.kelvin.to.centikelvin(x = 1)

unitconversion.K.to.cK(x = 1)

unitconversion.mole.to.centimole(x = 1)

unitconversion.mol.to.cmol(x = 1)

unitconversion.candela.to.centicandela(x = 1)

unitconversion.cd.to.ccd(x = 1)

unitconversion.radian.to.centiradian(x = 1)

unitconversion.rad.to.crad(x = 1)

unitconversion.steradian.to.centisteradian(x = 1)

unitconversion.sr.to.csr(x = 1)

unitconversion.hertz.to.centihertz(x = 1)

unitconversion.Hz.to.cHz(x = 1)

unitconversion.newton.to.centinewton(x = 1)

unitconversion.N.to.cN(x = 1)

unitconversion.pascal.to.centipascal(x = 1)

unitconversion.Pa.to.cPa(x = 1)

unitconversion.joule.to.centijoule(x = 1)

unitconversion.J.to.cJ(x = 1)

unitconversion.watt.to.centiwatt(x = 1)

unitconversion.W.to.cW(x = 1)

unitconversion.coulomb.to.centicoulomb(x = 1)

unitconversion.C.to.cC(x = 1)

unitconversion.volt.to.centivolt(x = 1)

unitconversion.V.to.cV(x = 1)

unitconversion.farad.to.centifarad(x = 1)

unitconversion.F.to.cF(x = 1)

unitconversion.ohm.to.centiohm(x = 1)

unitconversion.siemens.to.centisiemens(x = 1)

unitconversion.S.to.cS(x = 1)

unitconversion.weber.to.centiweber(x = 1)

unitconversion.Wb.to.cWb(x = 1)

unitconversion.tesla.to.centitesla(x = 1)

unitconversion.T.to.cT(x = 1)

unitconversion.henry.to.centihenry(x = 1)

unitconversion.H.to.cH(x = 1)

unitconversion.lumen.to.centilumen(x = 1)

unitconversion.lm.to.clm(x = 1)

unitconversion.lux.to.centilux(x = 1)

unitconversion.lx.to.clx(x = 1)

unitconversion.becquerel.to.centibecquerel(x = 1)

unitconversion.Bq.to.cBq(x = 1)

unitconversion.gray.to.centigray(x = 1)

unitconversion.Gy.to.cGy(x = 1)

unitconversion.sievert.to.centisievert(x = 1)

unitconversion.Sv.to.cSv(x = 1)

unitconversion.katal.to.centikatal(x = 1)

unitconversion.kat.to.ckat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to centi-units
}
\description{
Centi- hundredths, 1/100, or 10^(-2)
}
\details{
Performs a conversion from base units to centi-units (ex. grams to centigrams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
