% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.deci.R
\name{unitconversion.siprefix.base.to.deci}
\alias{unitconversion.siprefix.base.to.deci}
\alias{unitconversion.second.to.decisecond}
\alias{unitconversion.s.to.ds}
\alias{unitconversion.meter.to.decimeter}
\alias{unitconversion.m.to.dm}
\alias{unitconversion.gram.to.decigram}
\alias{unitconversion.g.to.dg}
\alias{unitconversion.ampere.to.deciampere}
\alias{unitconversion.A.to.dA}
\alias{unitconversion.kelvin.to.decikelvin}
\alias{unitconversion.K.to.dK}
\alias{unitconversion.mole.to.decimole}
\alias{unitconversion.mol.to.dmol}
\alias{unitconversion.candela.to.decicandela}
\alias{unitconversion.cd.to.dcd}
\alias{unitconversion.radian.to.deciradian}
\alias{unitconversion.rad.to.drad}
\alias{unitconversion.steradian.to.decisteradian}
\alias{unitconversion.sr.to.dsr}
\alias{unitconversion.hertz.to.decihertz}
\alias{unitconversion.Hz.to.dHz}
\alias{unitconversion.newton.to.decinewton}
\alias{unitconversion.N.to.dN}
\alias{unitconversion.pascal.to.decipascal}
\alias{unitconversion.Pa.to.dPa}
\alias{unitconversion.joule.to.decijoule}
\alias{unitconversion.J.to.dJ}
\alias{unitconversion.watt.to.deciwatt}
\alias{unitconversion.W.to.dW}
\alias{unitconversion.coulomb.to.decicoulomb}
\alias{unitconversion.C.to.dC}
\alias{unitconversion.volt.to.decivolt}
\alias{unitconversion.V.to.dV}
\alias{unitconversion.farad.to.decifarad}
\alias{unitconversion.F.to.dF}
\alias{unitconversion.ohm.to.deciohm}
\alias{unitconversion.siemens.to.decisiemens}
\alias{unitconversion.S.to.dS}
\alias{unitconversion.weber.to.deciweber}
\alias{unitconversion.Wb.to.dWb}
\alias{unitconversion.tesla.to.decitesla}
\alias{unitconversion.T.to.dT}
\alias{unitconversion.henry.to.decihenry}
\alias{unitconversion.H.to.dH}
\alias{unitconversion.lumen.to.decilumen}
\alias{unitconversion.lm.to.dlm}
\alias{unitconversion.lux.to.decilux}
\alias{unitconversion.lx.to.dlx}
\alias{unitconversion.becquerel.to.decibecquerel}
\alias{unitconversion.Bq.to.dBq}
\alias{unitconversion.gray.to.decigray}
\alias{unitconversion.Gy.to.dGy}
\alias{unitconversion.sievert.to.decisievert}
\alias{unitconversion.Sv.to.dSv}
\alias{unitconversion.katal.to.decikatal}
\alias{unitconversion.kat.to.dkat}
\title{Unit Conversion - SI Prefixes - Base to Deci-}
\usage{
unitconversion.siprefix.base.to.deci(x = 1)

unitconversion.second.to.decisecond(x = 1)

unitconversion.s.to.ds(x = 1)

unitconversion.meter.to.decimeter(x = 1)

unitconversion.m.to.dm(x = 1)

unitconversion.gram.to.decigram(x = 1)

unitconversion.g.to.dg(x = 1)

unitconversion.ampere.to.deciampere(x = 1)

unitconversion.A.to.dA(x = 1)

unitconversion.kelvin.to.decikelvin(x = 1)

unitconversion.K.to.dK(x = 1)

unitconversion.mole.to.decimole(x = 1)

unitconversion.mol.to.dmol(x = 1)

unitconversion.candela.to.decicandela(x = 1)

unitconversion.cd.to.dcd(x = 1)

unitconversion.radian.to.deciradian(x = 1)

unitconversion.rad.to.drad(x = 1)

unitconversion.steradian.to.decisteradian(x = 1)

unitconversion.sr.to.dsr(x = 1)

unitconversion.hertz.to.decihertz(x = 1)

unitconversion.Hz.to.dHz(x = 1)

unitconversion.newton.to.decinewton(x = 1)

unitconversion.N.to.dN(x = 1)

unitconversion.pascal.to.decipascal(x = 1)

unitconversion.Pa.to.dPa(x = 1)

unitconversion.joule.to.decijoule(x = 1)

unitconversion.J.to.dJ(x = 1)

unitconversion.watt.to.deciwatt(x = 1)

unitconversion.W.to.dW(x = 1)

unitconversion.coulomb.to.decicoulomb(x = 1)

unitconversion.C.to.dC(x = 1)

unitconversion.volt.to.decivolt(x = 1)

unitconversion.V.to.dV(x = 1)

unitconversion.farad.to.decifarad(x = 1)

unitconversion.F.to.dF(x = 1)

unitconversion.ohm.to.deciohm(x = 1)

unitconversion.siemens.to.decisiemens(x = 1)

unitconversion.S.to.dS(x = 1)

unitconversion.weber.to.deciweber(x = 1)

unitconversion.Wb.to.dWb(x = 1)

unitconversion.tesla.to.decitesla(x = 1)

unitconversion.T.to.dT(x = 1)

unitconversion.henry.to.decihenry(x = 1)

unitconversion.H.to.dH(x = 1)

unitconversion.lumen.to.decilumen(x = 1)

unitconversion.lm.to.dlm(x = 1)

unitconversion.lux.to.decilux(x = 1)

unitconversion.lx.to.dlx(x = 1)

unitconversion.becquerel.to.decibecquerel(x = 1)

unitconversion.Bq.to.dBq(x = 1)

unitconversion.gray.to.decigray(x = 1)

unitconversion.Gy.to.dGy(x = 1)

unitconversion.sievert.to.decisievert(x = 1)

unitconversion.Sv.to.dSv(x = 1)

unitconversion.katal.to.decikatal(x = 1)

unitconversion.kat.to.dkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to deci-units
}
\description{
Deci- tenths, 1/10, or 10^(-1)
}
\details{
Performs a conversion from base units to deci-units (ex. grams to decigrams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
