% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.deka.R
\name{unitconversion.siprefix.base.to.deka}
\alias{unitconversion.siprefix.base.to.deka}
\alias{unitconversion.second.to.dekasecond}
\alias{unitconversion.s.to.das}
\alias{unitconversion.meter.to.dekameter}
\alias{unitconversion.m.to.dam}
\alias{unitconversion.gram.to.dekagram}
\alias{unitconversion.g.to.dag}
\alias{unitconversion.ampere.to.dekaampere}
\alias{unitconversion.A.to.daA}
\alias{unitconversion.kelvin.to.dekakelvin}
\alias{unitconversion.K.to.daK}
\alias{unitconversion.mole.to.dekamole}
\alias{unitconversion.mol.to.damol}
\alias{unitconversion.candela.to.dekacandela}
\alias{unitconversion.cd.to.dacd}
\alias{unitconversion.radian.to.dekaradian}
\alias{unitconversion.rad.to.darad}
\alias{unitconversion.steradian.to.dekasteradian}
\alias{unitconversion.sr.to.dasr}
\alias{unitconversion.hertz.to.dekahertz}
\alias{unitconversion.Hz.to.daHz}
\alias{unitconversion.newton.to.dekanewton}
\alias{unitconversion.N.to.daN}
\alias{unitconversion.pascal.to.dekapascal}
\alias{unitconversion.Pa.to.daPa}
\alias{unitconversion.joule.to.dekajoule}
\alias{unitconversion.J.to.daJ}
\alias{unitconversion.watt.to.dekawatt}
\alias{unitconversion.W.to.daW}
\alias{unitconversion.coulomb.to.dekacoulomb}
\alias{unitconversion.C.to.daC}
\alias{unitconversion.volt.to.dekavolt}
\alias{unitconversion.V.to.daV}
\alias{unitconversion.farad.to.dekafarad}
\alias{unitconversion.F.to.daF}
\alias{unitconversion.ohm.to.dekaohm}
\alias{unitconversion.siemens.to.dekasiemens}
\alias{unitconversion.S.to.daS}
\alias{unitconversion.weber.to.dekaweber}
\alias{unitconversion.Wb.to.daWb}
\alias{unitconversion.tesla.to.dekatesla}
\alias{unitconversion.T.to.daT}
\alias{unitconversion.henry.to.dekahenry}
\alias{unitconversion.H.to.daH}
\alias{unitconversion.lumen.to.dekalumen}
\alias{unitconversion.lm.to.dalm}
\alias{unitconversion.lux.to.dekalux}
\alias{unitconversion.lx.to.dalx}
\alias{unitconversion.becquerel.to.dekabecquerel}
\alias{unitconversion.Bq.to.daBq}
\alias{unitconversion.gray.to.dekagray}
\alias{unitconversion.Gy.to.daGy}
\alias{unitconversion.sievert.to.dekasievert}
\alias{unitconversion.Sv.to.daSv}
\alias{unitconversion.katal.to.dekakatal}
\alias{unitconversion.kat.to.dakat}
\title{Unit Conversion - SI Prefixes - Base to Deka-}
\usage{
unitconversion.siprefix.base.to.deka(x = 1)

unitconversion.second.to.dekasecond(x = 1)

unitconversion.s.to.das(x = 1)

unitconversion.meter.to.dekameter(x = 1)

unitconversion.m.to.dam(x = 1)

unitconversion.gram.to.dekagram(x = 1)

unitconversion.g.to.dag(x = 1)

unitconversion.ampere.to.dekaampere(x = 1)

unitconversion.A.to.daA(x = 1)

unitconversion.kelvin.to.dekakelvin(x = 1)

unitconversion.K.to.daK(x = 1)

unitconversion.mole.to.dekamole(x = 1)

unitconversion.mol.to.damol(x = 1)

unitconversion.candela.to.dekacandela(x = 1)

unitconversion.cd.to.dacd(x = 1)

unitconversion.radian.to.dekaradian(x = 1)

unitconversion.rad.to.darad(x = 1)

unitconversion.steradian.to.dekasteradian(x = 1)

unitconversion.sr.to.dasr(x = 1)

unitconversion.hertz.to.dekahertz(x = 1)

unitconversion.Hz.to.daHz(x = 1)

unitconversion.newton.to.dekanewton(x = 1)

unitconversion.N.to.daN(x = 1)

unitconversion.pascal.to.dekapascal(x = 1)

unitconversion.Pa.to.daPa(x = 1)

unitconversion.joule.to.dekajoule(x = 1)

unitconversion.J.to.daJ(x = 1)

unitconversion.watt.to.dekawatt(x = 1)

unitconversion.W.to.daW(x = 1)

unitconversion.coulomb.to.dekacoulomb(x = 1)

unitconversion.C.to.daC(x = 1)

unitconversion.volt.to.dekavolt(x = 1)

unitconversion.V.to.daV(x = 1)

unitconversion.farad.to.dekafarad(x = 1)

unitconversion.F.to.daF(x = 1)

unitconversion.ohm.to.dekaohm(x = 1)

unitconversion.siemens.to.dekasiemens(x = 1)

unitconversion.S.to.daS(x = 1)

unitconversion.weber.to.dekaweber(x = 1)

unitconversion.Wb.to.daWb(x = 1)

unitconversion.tesla.to.dekatesla(x = 1)

unitconversion.T.to.daT(x = 1)

unitconversion.henry.to.dekahenry(x = 1)

unitconversion.H.to.daH(x = 1)

unitconversion.lumen.to.dekalumen(x = 1)

unitconversion.lm.to.dalm(x = 1)

unitconversion.lux.to.dekalux(x = 1)

unitconversion.lx.to.dalx(x = 1)

unitconversion.becquerel.to.dekabecquerel(x = 1)

unitconversion.Bq.to.daBq(x = 1)

unitconversion.gray.to.dekagray(x = 1)

unitconversion.Gy.to.daGy(x = 1)

unitconversion.sievert.to.dekasievert(x = 1)

unitconversion.Sv.to.daSv(x = 1)

unitconversion.katal.to.dekakatal(x = 1)

unitconversion.kat.to.dakat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to deka-units
}
\description{
Deka- tens, 10, or 10^1
}
\details{
Performs a conversion from base units to deka-units (ex. grams to dekagrams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
