% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.exa.R
\name{unitconversion.siprefix.base.to.exa}
\alias{unitconversion.siprefix.base.to.exa}
\alias{unitconversion.second.to.exasecond}
\alias{unitconversion.s.to.Es}
\alias{unitconversion.meter.to.exameter}
\alias{unitconversion.m.to.Em}
\alias{unitconversion.gram.to.exagram}
\alias{unitconversion.g.to.Eg}
\alias{unitconversion.ampere.to.exaampere}
\alias{unitconversion.A.to.EA}
\alias{unitconversion.kelvin.to.exakelvin}
\alias{unitconversion.K.to.EK}
\alias{unitconversion.mole.to.examole}
\alias{unitconversion.mol.to.Emol}
\alias{unitconversion.candela.to.exacandela}
\alias{unitconversion.cd.to.Ecd}
\alias{unitconversion.radian.to.exaradian}
\alias{unitconversion.rad.to.Erad}
\alias{unitconversion.steradian.to.exasteradian}
\alias{unitconversion.sr.to.Esr}
\alias{unitconversion.hertz.to.exahertz}
\alias{unitconversion.Hz.to.EHz}
\alias{unitconversion.newton.to.exanewton}
\alias{unitconversion.N.to.EN}
\alias{unitconversion.pascal.to.exapascal}
\alias{unitconversion.Pa.to.EPa}
\alias{unitconversion.joule.to.exajoule}
\alias{unitconversion.J.to.EJ}
\alias{unitconversion.watt.to.exawatt}
\alias{unitconversion.W.to.EW}
\alias{unitconversion.coulomb.to.exacoulomb}
\alias{unitconversion.C.to.EC}
\alias{unitconversion.volt.to.exavolt}
\alias{unitconversion.V.to.EV}
\alias{unitconversion.farad.to.exafarad}
\alias{unitconversion.F.to.EF}
\alias{unitconversion.ohm.to.exaohm}
\alias{unitconversion.siemens.to.exasiemens}
\alias{unitconversion.S.to.ES}
\alias{unitconversion.weber.to.exaweber}
\alias{unitconversion.Wb.to.EWb}
\alias{unitconversion.tesla.to.exatesla}
\alias{unitconversion.T.to.ET}
\alias{unitconversion.henry.to.exahenry}
\alias{unitconversion.H.to.EH}
\alias{unitconversion.lumen.to.exalumen}
\alias{unitconversion.lm.to.Elm}
\alias{unitconversion.lux.to.exalux}
\alias{unitconversion.lx.to.Elx}
\alias{unitconversion.becquerel.to.exabecquerel}
\alias{unitconversion.Bq.to.EBq}
\alias{unitconversion.gray.to.exagray}
\alias{unitconversion.Gy.to.EGy}
\alias{unitconversion.sievert.to.exasievert}
\alias{unitconversion.Sv.to.ESv}
\alias{unitconversion.katal.to.exakatal}
\alias{unitconversion.kat.to.Ekat}
\title{Unit Conversion - SI Prefixes - Base to Exa-}
\usage{
unitconversion.siprefix.base.to.exa(x = 1)

unitconversion.second.to.exasecond(x = 1)

unitconversion.s.to.Es(x = 1)

unitconversion.meter.to.exameter(x = 1)

unitconversion.m.to.Em(x = 1)

unitconversion.gram.to.exagram(x = 1)

unitconversion.g.to.Eg(x = 1)

unitconversion.ampere.to.exaampere(x = 1)

unitconversion.A.to.EA(x = 1)

unitconversion.kelvin.to.exakelvin(x = 1)

unitconversion.K.to.EK(x = 1)

unitconversion.mole.to.examole(x = 1)

unitconversion.mol.to.Emol(x = 1)

unitconversion.candela.to.exacandela(x = 1)

unitconversion.cd.to.Ecd(x = 1)

unitconversion.radian.to.exaradian(x = 1)

unitconversion.rad.to.Erad(x = 1)

unitconversion.steradian.to.exasteradian(x = 1)

unitconversion.sr.to.Esr(x = 1)

unitconversion.hertz.to.exahertz(x = 1)

unitconversion.Hz.to.EHz(x = 1)

unitconversion.newton.to.exanewton(x = 1)

unitconversion.N.to.EN(x = 1)

unitconversion.pascal.to.exapascal(x = 1)

unitconversion.Pa.to.EPa(x = 1)

unitconversion.joule.to.exajoule(x = 1)

unitconversion.J.to.EJ(x = 1)

unitconversion.watt.to.exawatt(x = 1)

unitconversion.W.to.EW(x = 1)

unitconversion.coulomb.to.exacoulomb(x = 1)

unitconversion.C.to.EC(x = 1)

unitconversion.volt.to.exavolt(x = 1)

unitconversion.V.to.EV(x = 1)

unitconversion.farad.to.exafarad(x = 1)

unitconversion.F.to.EF(x = 1)

unitconversion.ohm.to.exaohm(x = 1)

unitconversion.siemens.to.exasiemens(x = 1)

unitconversion.S.to.ES(x = 1)

unitconversion.weber.to.exaweber(x = 1)

unitconversion.Wb.to.EWb(x = 1)

unitconversion.tesla.to.exatesla(x = 1)

unitconversion.T.to.ET(x = 1)

unitconversion.henry.to.exahenry(x = 1)

unitconversion.H.to.EH(x = 1)

unitconversion.lumen.to.exalumen(x = 1)

unitconversion.lm.to.Elm(x = 1)

unitconversion.lux.to.exalux(x = 1)

unitconversion.lx.to.Elx(x = 1)

unitconversion.becquerel.to.exabecquerel(x = 1)

unitconversion.Bq.to.EBq(x = 1)

unitconversion.gray.to.exagray(x = 1)

unitconversion.Gy.to.EGy(x = 1)

unitconversion.sievert.to.exasievert(x = 1)

unitconversion.Sv.to.ESv(x = 1)

unitconversion.katal.to.exakatal(x = 1)

unitconversion.kat.to.Ekat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to exa-units
}
\description{
Exa- quintillions, 1 000 000 000 000 000 000, or 10^18
}
\details{
Performs a conversion from base units to exa-units (ex. grams to exagrams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
