% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.femto.R
\name{unitconversion.siprefix.base.to.femto}
\alias{unitconversion.siprefix.base.to.femto}
\alias{unitconversion.second.to.femtosecond}
\alias{unitconversion.s.to.fs}
\alias{unitconversion.meter.to.femtometer}
\alias{unitconversion.m.to.fm}
\alias{unitconversion.gram.to.femtogram}
\alias{unitconversion.g.to.fg}
\alias{unitconversion.ampere.to.femtoampere}
\alias{unitconversion.A.to.fA}
\alias{unitconversion.kelvin.to.femtokelvin}
\alias{unitconversion.K.to.fK}
\alias{unitconversion.mole.to.femtomole}
\alias{unitconversion.mol.to.fmol}
\alias{unitconversion.candela.to.femtocandela}
\alias{unitconversion.cd.to.fcd}
\alias{unitconversion.radian.to.femtoradian}
\alias{unitconversion.rad.to.frad}
\alias{unitconversion.steradian.to.femtosteradian}
\alias{unitconversion.sr.to.fsr}
\alias{unitconversion.hertz.to.femtohertz}
\alias{unitconversion.Hz.to.fHz}
\alias{unitconversion.newton.to.femtonewton}
\alias{unitconversion.N.to.fN}
\alias{unitconversion.pascal.to.femtopascal}
\alias{unitconversion.Pa.to.fPa}
\alias{unitconversion.joule.to.femtojoule}
\alias{unitconversion.J.to.fJ}
\alias{unitconversion.watt.to.femtowatt}
\alias{unitconversion.W.to.fW}
\alias{unitconversion.coulomb.to.femtocoulomb}
\alias{unitconversion.C.to.fC}
\alias{unitconversion.volt.to.femtovolt}
\alias{unitconversion.V.to.fV}
\alias{unitconversion.farad.to.femtofarad}
\alias{unitconversion.F.to.fF}
\alias{unitconversion.ohm.to.femtoohm}
\alias{unitconversion.siemens.to.femtosiemens}
\alias{unitconversion.S.to.fS}
\alias{unitconversion.weber.to.femtoweber}
\alias{unitconversion.Wb.to.fWb}
\alias{unitconversion.tesla.to.femtotesla}
\alias{unitconversion.T.to.fT}
\alias{unitconversion.henry.to.femtohenry}
\alias{unitconversion.H.to.fH}
\alias{unitconversion.lumen.to.femtolumen}
\alias{unitconversion.lm.to.flm}
\alias{unitconversion.lux.to.femtolux}
\alias{unitconversion.lx.to.flx}
\alias{unitconversion.becquerel.to.femtobecquerel}
\alias{unitconversion.Bq.to.fBq}
\alias{unitconversion.gray.to.femtogray}
\alias{unitconversion.Gy.to.fGy}
\alias{unitconversion.sievert.to.femtosievert}
\alias{unitconversion.Sv.to.fSv}
\alias{unitconversion.katal.to.femtokatal}
\alias{unitconversion.kat.to.fkat}
\title{Unit Conversion - SI Prefixes - Base to Femto-}
\usage{
unitconversion.siprefix.base.to.femto(x = 1)

unitconversion.second.to.femtosecond(x = 1)

unitconversion.s.to.fs(x = 1)

unitconversion.meter.to.femtometer(x = 1)

unitconversion.m.to.fm(x = 1)

unitconversion.gram.to.femtogram(x = 1)

unitconversion.g.to.fg(x = 1)

unitconversion.ampere.to.femtoampere(x = 1)

unitconversion.A.to.fA(x = 1)

unitconversion.kelvin.to.femtokelvin(x = 1)

unitconversion.K.to.fK(x = 1)

unitconversion.mole.to.femtomole(x = 1)

unitconversion.mol.to.fmol(x = 1)

unitconversion.candela.to.femtocandela(x = 1)

unitconversion.cd.to.fcd(x = 1)

unitconversion.radian.to.femtoradian(x = 1)

unitconversion.rad.to.frad(x = 1)

unitconversion.steradian.to.femtosteradian(x = 1)

unitconversion.sr.to.fsr(x = 1)

unitconversion.hertz.to.femtohertz(x = 1)

unitconversion.Hz.to.fHz(x = 1)

unitconversion.newton.to.femtonewton(x = 1)

unitconversion.N.to.fN(x = 1)

unitconversion.pascal.to.femtopascal(x = 1)

unitconversion.Pa.to.fPa(x = 1)

unitconversion.joule.to.femtojoule(x = 1)

unitconversion.J.to.fJ(x = 1)

unitconversion.watt.to.femtowatt(x = 1)

unitconversion.W.to.fW(x = 1)

unitconversion.coulomb.to.femtocoulomb(x = 1)

unitconversion.C.to.fC(x = 1)

unitconversion.volt.to.femtovolt(x = 1)

unitconversion.V.to.fV(x = 1)

unitconversion.farad.to.femtofarad(x = 1)

unitconversion.F.to.fF(x = 1)

unitconversion.ohm.to.femtoohm(x = 1)

unitconversion.siemens.to.femtosiemens(x = 1)

unitconversion.S.to.fS(x = 1)

unitconversion.weber.to.femtoweber(x = 1)

unitconversion.Wb.to.fWb(x = 1)

unitconversion.tesla.to.femtotesla(x = 1)

unitconversion.T.to.fT(x = 1)

unitconversion.henry.to.femtohenry(x = 1)

unitconversion.H.to.fH(x = 1)

unitconversion.lumen.to.femtolumen(x = 1)

unitconversion.lm.to.flm(x = 1)

unitconversion.lux.to.femtolux(x = 1)

unitconversion.lx.to.flx(x = 1)

unitconversion.becquerel.to.femtobecquerel(x = 1)

unitconversion.Bq.to.fBq(x = 1)

unitconversion.gray.to.femtogray(x = 1)

unitconversion.Gy.to.fGy(x = 1)

unitconversion.sievert.to.femtosievert(x = 1)

unitconversion.Sv.to.fSv(x = 1)

unitconversion.katal.to.femtokatal(x = 1)

unitconversion.kat.to.fkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to femto-units
}
\description{
Femto- quadrillionths, 1/1 000 000 000 000 000, or 10^(-15)
}
\details{
Performs a conversion from base units to femto-units (ex. grams to femtograms).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
