% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.giga.R
\name{unitconversion.siprefix.base.to.giga}
\alias{unitconversion.siprefix.base.to.giga}
\alias{unitconversion.second.to.gigasecond}
\alias{unitconversion.s.to.Gs}
\alias{unitconversion.meter.to.gigameter}
\alias{unitconversion.m.to.Gm}
\alias{unitconversion.gram.to.gigagram}
\alias{unitconversion.g.to.Gg}
\alias{unitconversion.ampere.to.gigaampere}
\alias{unitconversion.A.to.GA}
\alias{unitconversion.kelvin.to.gigakelvin}
\alias{unitconversion.K.to.GK}
\alias{unitconversion.mole.to.gigamole}
\alias{unitconversion.mol.to.Gmol}
\alias{unitconversion.candela.to.gigacandela}
\alias{unitconversion.cd.to.Gcd}
\alias{unitconversion.radian.to.gigaradian}
\alias{unitconversion.rad.to.Grad}
\alias{unitconversion.steradian.to.gigasteradian}
\alias{unitconversion.sr.to.Gsr}
\alias{unitconversion.hertz.to.gigahertz}
\alias{unitconversion.Hz.to.GHz}
\alias{unitconversion.newton.to.giganewton}
\alias{unitconversion.N.to.GN}
\alias{unitconversion.pascal.to.gigapascal}
\alias{unitconversion.Pa.to.GPa}
\alias{unitconversion.joule.to.gigajoule}
\alias{unitconversion.J.to.GJ}
\alias{unitconversion.watt.to.gigawatt}
\alias{unitconversion.W.to.GW}
\alias{unitconversion.coulomb.to.gigacoulomb}
\alias{unitconversion.C.to.GC}
\alias{unitconversion.volt.to.gigavolt}
\alias{unitconversion.V.to.GV}
\alias{unitconversion.farad.to.gigafarad}
\alias{unitconversion.F.to.GF}
\alias{unitconversion.ohm.to.gigaohm}
\alias{unitconversion.siemens.to.gigasiemens}
\alias{unitconversion.S.to.GS}
\alias{unitconversion.weber.to.gigaweber}
\alias{unitconversion.Wb.to.GWb}
\alias{unitconversion.tesla.to.gigatesla}
\alias{unitconversion.T.to.GT}
\alias{unitconversion.henry.to.gigahenry}
\alias{unitconversion.H.to.GH}
\alias{unitconversion.lumen.to.gigalumen}
\alias{unitconversion.lm.to.Glm}
\alias{unitconversion.lux.to.gigalux}
\alias{unitconversion.lx.to.Glx}
\alias{unitconversion.becquerel.to.gigabecquerel}
\alias{unitconversion.Bq.to.GBq}
\alias{unitconversion.gray.to.gigagray}
\alias{unitconversion.Gy.to.GGy}
\alias{unitconversion.sievert.to.gigasievert}
\alias{unitconversion.Sv.to.GSv}
\alias{unitconversion.katal.to.gigakatal}
\alias{unitconversion.kat.to.Gkat}
\title{Unit Conversion - SI Prefixes - Base to Giga-}
\usage{
unitconversion.siprefix.base.to.giga(x = 1)

unitconversion.second.to.gigasecond(x = 1)

unitconversion.s.to.Gs(x = 1)

unitconversion.meter.to.gigameter(x = 1)

unitconversion.m.to.Gm(x = 1)

unitconversion.gram.to.gigagram(x = 1)

unitconversion.g.to.Gg(x = 1)

unitconversion.ampere.to.gigaampere(x = 1)

unitconversion.A.to.GA(x = 1)

unitconversion.kelvin.to.gigakelvin(x = 1)

unitconversion.K.to.GK(x = 1)

unitconversion.mole.to.gigamole(x = 1)

unitconversion.mol.to.Gmol(x = 1)

unitconversion.candela.to.gigacandela(x = 1)

unitconversion.cd.to.Gcd(x = 1)

unitconversion.radian.to.gigaradian(x = 1)

unitconversion.rad.to.Grad(x = 1)

unitconversion.steradian.to.gigasteradian(x = 1)

unitconversion.sr.to.Gsr(x = 1)

unitconversion.hertz.to.gigahertz(x = 1)

unitconversion.Hz.to.GHz(x = 1)

unitconversion.newton.to.giganewton(x = 1)

unitconversion.N.to.GN(x = 1)

unitconversion.pascal.to.gigapascal(x = 1)

unitconversion.Pa.to.GPa(x = 1)

unitconversion.joule.to.gigajoule(x = 1)

unitconversion.J.to.GJ(x = 1)

unitconversion.watt.to.gigawatt(x = 1)

unitconversion.W.to.GW(x = 1)

unitconversion.coulomb.to.gigacoulomb(x = 1)

unitconversion.C.to.GC(x = 1)

unitconversion.volt.to.gigavolt(x = 1)

unitconversion.V.to.GV(x = 1)

unitconversion.farad.to.gigafarad(x = 1)

unitconversion.F.to.GF(x = 1)

unitconversion.ohm.to.gigaohm(x = 1)

unitconversion.siemens.to.gigasiemens(x = 1)

unitconversion.S.to.GS(x = 1)

unitconversion.weber.to.gigaweber(x = 1)

unitconversion.Wb.to.GWb(x = 1)

unitconversion.tesla.to.gigatesla(x = 1)

unitconversion.T.to.GT(x = 1)

unitconversion.henry.to.gigahenry(x = 1)

unitconversion.H.to.GH(x = 1)

unitconversion.lumen.to.gigalumen(x = 1)

unitconversion.lm.to.Glm(x = 1)

unitconversion.lux.to.gigalux(x = 1)

unitconversion.lx.to.Glx(x = 1)

unitconversion.becquerel.to.gigabecquerel(x = 1)

unitconversion.Bq.to.GBq(x = 1)

unitconversion.gray.to.gigagray(x = 1)

unitconversion.Gy.to.GGy(x = 1)

unitconversion.sievert.to.gigasievert(x = 1)

unitconversion.Sv.to.GSv(x = 1)

unitconversion.katal.to.gigakatal(x = 1)

unitconversion.kat.to.Gkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to giga-units
}
\description{
Giga- billions, 1 000 000 000, or 10^9
}
\details{
Performs a conversion from base units to giga-units (ex. grams to gigagrams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
