% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.hecto.R
\name{unitconversion.siprefix.base.to.hecto}
\alias{unitconversion.siprefix.base.to.hecto}
\alias{unitconversion.second.to.hectosecond}
\alias{unitconversion.s.to.hs}
\alias{unitconversion.meter.to.hectometer}
\alias{unitconversion.m.to.hm}
\alias{unitconversion.gram.to.hectogram}
\alias{unitconversion.g.to.hg}
\alias{unitconversion.ampere.to.hectoampere}
\alias{unitconversion.A.to.hA}
\alias{unitconversion.kelvin.to.hectokelvin}
\alias{unitconversion.K.to.hK}
\alias{unitconversion.mole.to.hectomole}
\alias{unitconversion.mol.to.hmol}
\alias{unitconversion.candela.to.hectocandela}
\alias{unitconversion.cd.to.hcd}
\alias{unitconversion.radian.to.hectoradian}
\alias{unitconversion.rad.to.hrad}
\alias{unitconversion.steradian.to.hectosteradian}
\alias{unitconversion.sr.to.hsr}
\alias{unitconversion.hertz.to.hectohertz}
\alias{unitconversion.Hz.to.hHz}
\alias{unitconversion.newton.to.hectonewton}
\alias{unitconversion.N.to.hN}
\alias{unitconversion.pascal.to.hectopascal}
\alias{unitconversion.Pa.to.hPa}
\alias{unitconversion.joule.to.hectojoule}
\alias{unitconversion.J.to.hJ}
\alias{unitconversion.watt.to.hectowatt}
\alias{unitconversion.W.to.hW}
\alias{unitconversion.coulomb.to.hectocoulomb}
\alias{unitconversion.C.to.hC}
\alias{unitconversion.volt.to.hectovolt}
\alias{unitconversion.V.to.hV}
\alias{unitconversion.farad.to.hectofarad}
\alias{unitconversion.F.to.hF}
\alias{unitconversion.ohm.to.hectoohm}
\alias{unitconversion.siemens.to.hectosiemens}
\alias{unitconversion.S.to.hS}
\alias{unitconversion.weber.to.hectoweber}
\alias{unitconversion.Wb.to.hWb}
\alias{unitconversion.tesla.to.hectotesla}
\alias{unitconversion.T.to.hT}
\alias{unitconversion.henry.to.hectohenry}
\alias{unitconversion.H.to.hH}
\alias{unitconversion.lumen.to.hectolumen}
\alias{unitconversion.lm.to.hlm}
\alias{unitconversion.lux.to.hectolux}
\alias{unitconversion.lx.to.hlx}
\alias{unitconversion.becquerel.to.hectobecquerel}
\alias{unitconversion.Bq.to.hBq}
\alias{unitconversion.gray.to.hectogray}
\alias{unitconversion.Gy.to.hGy}
\alias{unitconversion.sievert.to.hectosievert}
\alias{unitconversion.Sv.to.hSv}
\alias{unitconversion.katal.to.hectokatal}
\alias{unitconversion.kat.to.hkat}
\title{Unit Conversion - SI Prefixes - Base to Hecto-}
\usage{
unitconversion.siprefix.base.to.hecto(x = 1)

unitconversion.second.to.hectosecond(x = 1)

unitconversion.s.to.hs(x = 1)

unitconversion.meter.to.hectometer(x = 1)

unitconversion.m.to.hm(x = 1)

unitconversion.gram.to.hectogram(x = 1)

unitconversion.g.to.hg(x = 1)

unitconversion.ampere.to.hectoampere(x = 1)

unitconversion.A.to.hA(x = 1)

unitconversion.kelvin.to.hectokelvin(x = 1)

unitconversion.K.to.hK(x = 1)

unitconversion.mole.to.hectomole(x = 1)

unitconversion.mol.to.hmol(x = 1)

unitconversion.candela.to.hectocandela(x = 1)

unitconversion.cd.to.hcd(x = 1)

unitconversion.radian.to.hectoradian(x = 1)

unitconversion.rad.to.hrad(x = 1)

unitconversion.steradian.to.hectosteradian(x = 1)

unitconversion.sr.to.hsr(x = 1)

unitconversion.hertz.to.hectohertz(x = 1)

unitconversion.Hz.to.hHz(x = 1)

unitconversion.newton.to.hectonewton(x = 1)

unitconversion.N.to.hN(x = 1)

unitconversion.pascal.to.hectopascal(x = 1)

unitconversion.Pa.to.hPa(x = 1)

unitconversion.joule.to.hectojoule(x = 1)

unitconversion.J.to.hJ(x = 1)

unitconversion.watt.to.hectowatt(x = 1)

unitconversion.W.to.hW(x = 1)

unitconversion.coulomb.to.hectocoulomb(x = 1)

unitconversion.C.to.hC(x = 1)

unitconversion.volt.to.hectovolt(x = 1)

unitconversion.V.to.hV(x = 1)

unitconversion.farad.to.hectofarad(x = 1)

unitconversion.F.to.hF(x = 1)

unitconversion.ohm.to.hectoohm(x = 1)

unitconversion.siemens.to.hectosiemens(x = 1)

unitconversion.S.to.hS(x = 1)

unitconversion.weber.to.hectoweber(x = 1)

unitconversion.Wb.to.hWb(x = 1)

unitconversion.tesla.to.hectotesla(x = 1)

unitconversion.T.to.hT(x = 1)

unitconversion.henry.to.hectohenry(x = 1)

unitconversion.H.to.hH(x = 1)

unitconversion.lumen.to.hectolumen(x = 1)

unitconversion.lm.to.hlm(x = 1)

unitconversion.lux.to.hectolux(x = 1)

unitconversion.lx.to.hlx(x = 1)

unitconversion.becquerel.to.hectobecquerel(x = 1)

unitconversion.Bq.to.hBq(x = 1)

unitconversion.gray.to.hectogray(x = 1)

unitconversion.Gy.to.hGy(x = 1)

unitconversion.sievert.to.hectosievert(x = 1)

unitconversion.Sv.to.hSv(x = 1)

unitconversion.katal.to.hectokatal(x = 1)

unitconversion.kat.to.hkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to hecto-units
}
\description{
Hecto- hundreds, 100, or 10^2
}
\details{
Performs a conversion from base units to hecto-units (ex. grams to hectograms).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
