% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.kilo.R
\name{unitconversion.siprefix.base.to.kilo}
\alias{unitconversion.siprefix.base.to.kilo}
\alias{unitconversion.second.to.kilosecond}
\alias{unitconversion.s.to.ks}
\alias{unitconversion.meter.to.kilometer}
\alias{unitconversion.m.to.km}
\alias{unitconversion.gram.to.kilogram}
\alias{unitconversion.g.to.kg}
\alias{unitconversion.ampere.to.kiloampere}
\alias{unitconversion.A.to.kA}
\alias{unitconversion.kelvin.to.kilokelvin}
\alias{unitconversion.K.to.kK}
\alias{unitconversion.mole.to.kilomole}
\alias{unitconversion.mol.to.kmol}
\alias{unitconversion.candela.to.kilocandela}
\alias{unitconversion.cd.to.kcd}
\alias{unitconversion.radian.to.kiloradian}
\alias{unitconversion.rad.to.krad}
\alias{unitconversion.steradian.to.kilosteradian}
\alias{unitconversion.sr.to.ksr}
\alias{unitconversion.hertz.to.kilohertz}
\alias{unitconversion.Hz.to.kHz}
\alias{unitconversion.newton.to.kilonewton}
\alias{unitconversion.N.to.kN}
\alias{unitconversion.pascal.to.kilopascal}
\alias{unitconversion.Pa.to.kPa}
\alias{unitconversion.joule.to.kilojoule}
\alias{unitconversion.J.to.kJ}
\alias{unitconversion.watt.to.kilowatt}
\alias{unitconversion.W.to.kW}
\alias{unitconversion.coulomb.to.kilocoulomb}
\alias{unitconversion.C.to.kC}
\alias{unitconversion.volt.to.kilovolt}
\alias{unitconversion.V.to.kV}
\alias{unitconversion.farad.to.kilofarad}
\alias{unitconversion.F.to.kF}
\alias{unitconversion.ohm.to.kiloohm}
\alias{unitconversion.siemens.to.kilosiemens}
\alias{unitconversion.S.to.kS}
\alias{unitconversion.weber.to.kiloweber}
\alias{unitconversion.Wb.to.kWb}
\alias{unitconversion.tesla.to.kilotesla}
\alias{unitconversion.T.to.kT}
\alias{unitconversion.henry.to.kilohenry}
\alias{unitconversion.H.to.kH}
\alias{unitconversion.lumen.to.kilolumen}
\alias{unitconversion.lm.to.klm}
\alias{unitconversion.lux.to.kilolux}
\alias{unitconversion.lx.to.klx}
\alias{unitconversion.becquerel.to.kilobecquerel}
\alias{unitconversion.Bq.to.kBq}
\alias{unitconversion.gray.to.kilogray}
\alias{unitconversion.Gy.to.kGy}
\alias{unitconversion.sievert.to.kilosievert}
\alias{unitconversion.Sv.to.kSv}
\alias{unitconversion.katal.to.kilokatal}
\alias{unitconversion.kat.to.kkat}
\title{Unit Conversion - SI Prefixes - Base to Kilo-}
\usage{
unitconversion.siprefix.base.to.kilo(x = 1)

unitconversion.second.to.kilosecond(x = 1)

unitconversion.s.to.ks(x = 1)

unitconversion.meter.to.kilometer(x = 1)

unitconversion.m.to.km(x = 1)

unitconversion.gram.to.kilogram(x = 1)

unitconversion.g.to.kg(x = 1)

unitconversion.ampere.to.kiloampere(x = 1)

unitconversion.A.to.kA(x = 1)

unitconversion.kelvin.to.kilokelvin(x = 1)

unitconversion.K.to.kK(x = 1)

unitconversion.mole.to.kilomole(x = 1)

unitconversion.mol.to.kmol(x = 1)

unitconversion.candela.to.kilocandela(x = 1)

unitconversion.cd.to.kcd(x = 1)

unitconversion.radian.to.kiloradian(x = 1)

unitconversion.rad.to.krad(x = 1)

unitconversion.steradian.to.kilosteradian(x = 1)

unitconversion.sr.to.ksr(x = 1)

unitconversion.hertz.to.kilohertz(x = 1)

unitconversion.Hz.to.kHz(x = 1)

unitconversion.newton.to.kilonewton(x = 1)

unitconversion.N.to.kN(x = 1)

unitconversion.pascal.to.kilopascal(x = 1)

unitconversion.Pa.to.kPa(x = 1)

unitconversion.joule.to.kilojoule(x = 1)

unitconversion.J.to.kJ(x = 1)

unitconversion.watt.to.kilowatt(x = 1)

unitconversion.W.to.kW(x = 1)

unitconversion.coulomb.to.kilocoulomb(x = 1)

unitconversion.C.to.kC(x = 1)

unitconversion.volt.to.kilovolt(x = 1)

unitconversion.V.to.kV(x = 1)

unitconversion.farad.to.kilofarad(x = 1)

unitconversion.F.to.kF(x = 1)

unitconversion.ohm.to.kiloohm(x = 1)

unitconversion.siemens.to.kilosiemens(x = 1)

unitconversion.S.to.kS(x = 1)

unitconversion.weber.to.kiloweber(x = 1)

unitconversion.Wb.to.kWb(x = 1)

unitconversion.tesla.to.kilotesla(x = 1)

unitconversion.T.to.kT(x = 1)

unitconversion.henry.to.kilohenry(x = 1)

unitconversion.H.to.kH(x = 1)

unitconversion.lumen.to.kilolumen(x = 1)

unitconversion.lm.to.klm(x = 1)

unitconversion.lux.to.kilolux(x = 1)

unitconversion.lx.to.klx(x = 1)

unitconversion.becquerel.to.kilobecquerel(x = 1)

unitconversion.Bq.to.kBq(x = 1)

unitconversion.gray.to.kilogray(x = 1)

unitconversion.Gy.to.kGy(x = 1)

unitconversion.sievert.to.kilosievert(x = 1)

unitconversion.Sv.to.kSv(x = 1)

unitconversion.katal.to.kilokatal(x = 1)

unitconversion.kat.to.kkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to kilo-units
}
\description{
Kilo- thousands, 1000, or 10^3
}
\details{
Performs a conversion from base units to kilo-units (ex. grams to kilograms).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
