% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.mega.R
\name{unitconversion.siprefix.base.to.mega}
\alias{unitconversion.siprefix.base.to.mega}
\alias{unitconversion.second.to.megasecond}
\alias{unitconversion.s.to.Ms}
\alias{unitconversion.meter.to.megameter}
\alias{unitconversion.m.to.Mm}
\alias{unitconversion.gram.to.megagram}
\alias{unitconversion.g.to.Mg}
\alias{unitconversion.ampere.to.megaampere}
\alias{unitconversion.A.to.MA}
\alias{unitconversion.kelvin.to.megakelvin}
\alias{unitconversion.K.to.MK}
\alias{unitconversion.mole.to.megamole}
\alias{unitconversion.mol.to.Mmol}
\alias{unitconversion.candela.to.megacandela}
\alias{unitconversion.cd.to.Mcd}
\alias{unitconversion.radian.to.megaradian}
\alias{unitconversion.rad.to.Mrad}
\alias{unitconversion.steradian.to.megasteradian}
\alias{unitconversion.sr.to.Msr}
\alias{unitconversion.hertz.to.megahertz}
\alias{unitconversion.Hz.to.MHz}
\alias{unitconversion.newton.to.meganewton}
\alias{unitconversion.N.to.MN}
\alias{unitconversion.pascal.to.megapascal}
\alias{unitconversion.Pa.to.MPa}
\alias{unitconversion.joule.to.megajoule}
\alias{unitconversion.J.to.MJ}
\alias{unitconversion.watt.to.megawatt}
\alias{unitconversion.W.to.MW}
\alias{unitconversion.coulomb.to.megacoulomb}
\alias{unitconversion.C.to.MC}
\alias{unitconversion.volt.to.megavolt}
\alias{unitconversion.V.to.MV}
\alias{unitconversion.farad.to.megafarad}
\alias{unitconversion.F.to.MF}
\alias{unitconversion.ohm.to.megaohm}
\alias{unitconversion.siemens.to.megasiemens}
\alias{unitconversion.S.to.MS}
\alias{unitconversion.weber.to.megaweber}
\alias{unitconversion.Wb.to.MWb}
\alias{unitconversion.tesla.to.megatesla}
\alias{unitconversion.T.to.MT}
\alias{unitconversion.henry.to.megahenry}
\alias{unitconversion.H.to.MH}
\alias{unitconversion.lumen.to.megalumen}
\alias{unitconversion.lm.to.Mlm}
\alias{unitconversion.lux.to.megalux}
\alias{unitconversion.lx.to.Mlx}
\alias{unitconversion.becquerel.to.megabecquerel}
\alias{unitconversion.Bq.to.MBq}
\alias{unitconversion.gray.to.megagray}
\alias{unitconversion.Gy.to.MGy}
\alias{unitconversion.sievert.to.megasievert}
\alias{unitconversion.Sv.to.MSv}
\alias{unitconversion.katal.to.megakatal}
\alias{unitconversion.kat.to.Mkat}
\title{Unit Conversion - SI Prefixes - Base to Mega-}
\usage{
unitconversion.siprefix.base.to.mega(x = 1)

unitconversion.second.to.megasecond(x = 1)

unitconversion.s.to.Ms(x = 1)

unitconversion.meter.to.megameter(x = 1)

unitconversion.m.to.Mm(x = 1)

unitconversion.gram.to.megagram(x = 1)

unitconversion.g.to.Mg(x = 1)

unitconversion.ampere.to.megaampere(x = 1)

unitconversion.A.to.MA(x = 1)

unitconversion.kelvin.to.megakelvin(x = 1)

unitconversion.K.to.MK(x = 1)

unitconversion.mole.to.megamole(x = 1)

unitconversion.mol.to.Mmol(x = 1)

unitconversion.candela.to.megacandela(x = 1)

unitconversion.cd.to.Mcd(x = 1)

unitconversion.radian.to.megaradian(x = 1)

unitconversion.rad.to.Mrad(x = 1)

unitconversion.steradian.to.megasteradian(x = 1)

unitconversion.sr.to.Msr(x = 1)

unitconversion.hertz.to.megahertz(x = 1)

unitconversion.Hz.to.MHz(x = 1)

unitconversion.newton.to.meganewton(x = 1)

unitconversion.N.to.MN(x = 1)

unitconversion.pascal.to.megapascal(x = 1)

unitconversion.Pa.to.MPa(x = 1)

unitconversion.joule.to.megajoule(x = 1)

unitconversion.J.to.MJ(x = 1)

unitconversion.watt.to.megawatt(x = 1)

unitconversion.W.to.MW(x = 1)

unitconversion.coulomb.to.megacoulomb(x = 1)

unitconversion.C.to.MC(x = 1)

unitconversion.volt.to.megavolt(x = 1)

unitconversion.V.to.MV(x = 1)

unitconversion.farad.to.megafarad(x = 1)

unitconversion.F.to.MF(x = 1)

unitconversion.ohm.to.megaohm(x = 1)

unitconversion.siemens.to.megasiemens(x = 1)

unitconversion.S.to.MS(x = 1)

unitconversion.weber.to.megaweber(x = 1)

unitconversion.Wb.to.MWb(x = 1)

unitconversion.tesla.to.megatesla(x = 1)

unitconversion.T.to.MT(x = 1)

unitconversion.henry.to.megahenry(x = 1)

unitconversion.H.to.MH(x = 1)

unitconversion.lumen.to.megalumen(x = 1)

unitconversion.lm.to.Mlm(x = 1)

unitconversion.lux.to.megalux(x = 1)

unitconversion.lx.to.Mlx(x = 1)

unitconversion.becquerel.to.megabecquerel(x = 1)

unitconversion.Bq.to.MBq(x = 1)

unitconversion.gray.to.megagray(x = 1)

unitconversion.Gy.to.MGy(x = 1)

unitconversion.sievert.to.megasievert(x = 1)

unitconversion.Sv.to.MSv(x = 1)

unitconversion.katal.to.megakatal(x = 1)

unitconversion.kat.to.Mkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to mega-units
}
\description{
Mega- millions, 1 000 000, or 10^6
}
\details{
Performs a conversion from base units to mega-units (ex. grams to megagrams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
