% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.micro.R
\name{unitconversion.siprefix.base.to.micro}
\alias{unitconversion.siprefix.base.to.micro}
\alias{unitconversion.second.to.microsecond}
\alias{unitconversion.s.to.us}
\alias{unitconversion.meter.to.micrometer}
\alias{unitconversion.m.to.um}
\alias{unitconversion.gram.to.microgram}
\alias{unitconversion.g.to.ug}
\alias{unitconversion.ampere.to.microampere}
\alias{unitconversion.A.to.uA}
\alias{unitconversion.kelvin.to.microkelvin}
\alias{unitconversion.K.to.uK}
\alias{unitconversion.mole.to.micromole}
\alias{unitconversion.mol.to.umol}
\alias{unitconversion.candela.to.microcandela}
\alias{unitconversion.cd.to.ucd}
\alias{unitconversion.radian.to.microradian}
\alias{unitconversion.rad.to.urad}
\alias{unitconversion.steradian.to.microsteradian}
\alias{unitconversion.sr.to.usr}
\alias{unitconversion.hertz.to.microhertz}
\alias{unitconversion.Hz.to.uHz}
\alias{unitconversion.newton.to.micronewton}
\alias{unitconversion.N.to.uN}
\alias{unitconversion.pascal.to.micropascal}
\alias{unitconversion.Pa.to.uPa}
\alias{unitconversion.joule.to.microjoule}
\alias{unitconversion.J.to.uJ}
\alias{unitconversion.watt.to.microwatt}
\alias{unitconversion.W.to.uW}
\alias{unitconversion.coulomb.to.microcoulomb}
\alias{unitconversion.C.to.uC}
\alias{unitconversion.volt.to.microvolt}
\alias{unitconversion.V.to.uV}
\alias{unitconversion.farad.to.microfarad}
\alias{unitconversion.F.to.uF}
\alias{unitconversion.ohm.to.microohm}
\alias{unitconversion.siemens.to.microsiemens}
\alias{unitconversion.S.to.uS}
\alias{unitconversion.weber.to.microweber}
\alias{unitconversion.Wb.to.uWb}
\alias{unitconversion.tesla.to.microtesla}
\alias{unitconversion.T.to.uT}
\alias{unitconversion.henry.to.microhenry}
\alias{unitconversion.H.to.uH}
\alias{unitconversion.lumen.to.microlumen}
\alias{unitconversion.lm.to.ulm}
\alias{unitconversion.lux.to.microlux}
\alias{unitconversion.lx.to.ulx}
\alias{unitconversion.becquerel.to.microbecquerel}
\alias{unitconversion.Bq.to.uBq}
\alias{unitconversion.gray.to.microgray}
\alias{unitconversion.Gy.to.uGy}
\alias{unitconversion.sievert.to.microsievert}
\alias{unitconversion.Sv.to.uSv}
\alias{unitconversion.katal.to.microkatal}
\alias{unitconversion.kat.to.ukat}
\title{Unit Conversion - SI Prefixes - Base to Micro-}
\usage{
unitconversion.siprefix.base.to.micro(x = 1)

unitconversion.second.to.microsecond(x = 1)

unitconversion.s.to.us(x = 1)

unitconversion.meter.to.micrometer(x = 1)

unitconversion.m.to.um(x = 1)

unitconversion.gram.to.microgram(x = 1)

unitconversion.g.to.ug(x = 1)

unitconversion.ampere.to.microampere(x = 1)

unitconversion.A.to.uA(x = 1)

unitconversion.kelvin.to.microkelvin(x = 1)

unitconversion.K.to.uK(x = 1)

unitconversion.mole.to.micromole(x = 1)

unitconversion.mol.to.umol(x = 1)

unitconversion.candela.to.microcandela(x = 1)

unitconversion.cd.to.ucd(x = 1)

unitconversion.radian.to.microradian(x = 1)

unitconversion.rad.to.urad(x = 1)

unitconversion.steradian.to.microsteradian(x = 1)

unitconversion.sr.to.usr(x = 1)

unitconversion.hertz.to.microhertz(x = 1)

unitconversion.Hz.to.uHz(x = 1)

unitconversion.newton.to.micronewton(x = 1)

unitconversion.N.to.uN(x = 1)

unitconversion.pascal.to.micropascal(x = 1)

unitconversion.Pa.to.uPa(x = 1)

unitconversion.joule.to.microjoule(x = 1)

unitconversion.J.to.uJ(x = 1)

unitconversion.watt.to.microwatt(x = 1)

unitconversion.W.to.uW(x = 1)

unitconversion.coulomb.to.microcoulomb(x = 1)

unitconversion.C.to.uC(x = 1)

unitconversion.volt.to.microvolt(x = 1)

unitconversion.V.to.uV(x = 1)

unitconversion.farad.to.microfarad(x = 1)

unitconversion.F.to.uF(x = 1)

unitconversion.ohm.to.microohm(x = 1)

unitconversion.siemens.to.microsiemens(x = 1)

unitconversion.S.to.uS(x = 1)

unitconversion.weber.to.microweber(x = 1)

unitconversion.Wb.to.uWb(x = 1)

unitconversion.tesla.to.microtesla(x = 1)

unitconversion.T.to.uT(x = 1)

unitconversion.henry.to.microhenry(x = 1)

unitconversion.H.to.uH(x = 1)

unitconversion.lumen.to.microlumen(x = 1)

unitconversion.lm.to.ulm(x = 1)

unitconversion.lux.to.microlux(x = 1)

unitconversion.lx.to.ulx(x = 1)

unitconversion.becquerel.to.microbecquerel(x = 1)

unitconversion.Bq.to.uBq(x = 1)

unitconversion.gray.to.microgray(x = 1)

unitconversion.Gy.to.uGy(x = 1)

unitconversion.sievert.to.microsievert(x = 1)

unitconversion.Sv.to.uSv(x = 1)

unitconversion.katal.to.microkatal(x = 1)

unitconversion.kat.to.ukat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to micro-units
}
\description{
Micro- millionths, 1/1 000 000, or 10^(-6)
}
\details{
Performs a conversion from base units to micro-units (ex. grams to micrograms).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
