% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.milli.R
\name{unitconversion.siprefix.base.to.milli}
\alias{unitconversion.siprefix.base.to.milli}
\alias{unitconversion.second.to.millisecond}
\alias{unitconversion.s.to.ms}
\alias{unitconversion.meter.to.millimeter}
\alias{unitconversion.m.to.mm}
\alias{unitconversion.gram.to.milligram}
\alias{unitconversion.g.to.mg}
\alias{unitconversion.ampere.to.milliampere}
\alias{unitconversion.A.to.mA}
\alias{unitconversion.kelvin.to.millikelvin}
\alias{unitconversion.K.to.mK}
\alias{unitconversion.mole.to.millimole}
\alias{unitconversion.mol.to.mmol}
\alias{unitconversion.candela.to.millicandela}
\alias{unitconversion.cd.to.mcd}
\alias{unitconversion.radian.to.milliradian}
\alias{unitconversion.rad.to.mrad}
\alias{unitconversion.steradian.to.millisteradian}
\alias{unitconversion.sr.to.msr}
\alias{unitconversion.hertz.to.millihertz}
\alias{unitconversion.Hz.to.mHz}
\alias{unitconversion.newton.to.millinewton}
\alias{unitconversion.N.to.mN}
\alias{unitconversion.pascal.to.millipascal}
\alias{unitconversion.Pa.to.mPa}
\alias{unitconversion.joule.to.millijoule}
\alias{unitconversion.J.to.mJ}
\alias{unitconversion.watt.to.milliwatt}
\alias{unitconversion.W.to.mW}
\alias{unitconversion.coulomb.to.millicoulomb}
\alias{unitconversion.C.to.mC}
\alias{unitconversion.volt.to.millivolt}
\alias{unitconversion.V.to.mV}
\alias{unitconversion.farad.to.millifarad}
\alias{unitconversion.F.to.mF}
\alias{unitconversion.ohm.to.milliohm}
\alias{unitconversion.siemens.to.millisiemens}
\alias{unitconversion.S.to.mS}
\alias{unitconversion.weber.to.milliweber}
\alias{unitconversion.Wb.to.mWb}
\alias{unitconversion.tesla.to.millitesla}
\alias{unitconversion.T.to.mT}
\alias{unitconversion.henry.to.millihenry}
\alias{unitconversion.H.to.mH}
\alias{unitconversion.lumen.to.millilumen}
\alias{unitconversion.lm.to.mlm}
\alias{unitconversion.lux.to.millilux}
\alias{unitconversion.lx.to.mlx}
\alias{unitconversion.becquerel.to.millibecquerel}
\alias{unitconversion.Bq.to.mBq}
\alias{unitconversion.gray.to.milligray}
\alias{unitconversion.Gy.to.mGy}
\alias{unitconversion.sievert.to.millisievert}
\alias{unitconversion.Sv.to.mSv}
\alias{unitconversion.katal.to.millikatal}
\alias{unitconversion.kat.to.mkat}
\title{Unit Conversion - SI Prefixes - Base to Milli-}
\usage{
unitconversion.siprefix.base.to.milli(x = 1)

unitconversion.second.to.millisecond(x = 1)

unitconversion.s.to.ms(x = 1)

unitconversion.meter.to.millimeter(x = 1)

unitconversion.m.to.mm(x = 1)

unitconversion.gram.to.milligram(x = 1)

unitconversion.g.to.mg(x = 1)

unitconversion.ampere.to.milliampere(x = 1)

unitconversion.A.to.mA(x = 1)

unitconversion.kelvin.to.millikelvin(x = 1)

unitconversion.K.to.mK(x = 1)

unitconversion.mole.to.millimole(x = 1)

unitconversion.mol.to.mmol(x = 1)

unitconversion.candela.to.millicandela(x = 1)

unitconversion.cd.to.mcd(x = 1)

unitconversion.radian.to.milliradian(x = 1)

unitconversion.rad.to.mrad(x = 1)

unitconversion.steradian.to.millisteradian(x = 1)

unitconversion.sr.to.msr(x = 1)

unitconversion.hertz.to.millihertz(x = 1)

unitconversion.Hz.to.mHz(x = 1)

unitconversion.newton.to.millinewton(x = 1)

unitconversion.N.to.mN(x = 1)

unitconversion.pascal.to.millipascal(x = 1)

unitconversion.Pa.to.mPa(x = 1)

unitconversion.joule.to.millijoule(x = 1)

unitconversion.J.to.mJ(x = 1)

unitconversion.watt.to.milliwatt(x = 1)

unitconversion.W.to.mW(x = 1)

unitconversion.coulomb.to.millicoulomb(x = 1)

unitconversion.C.to.mC(x = 1)

unitconversion.volt.to.millivolt(x = 1)

unitconversion.V.to.mV(x = 1)

unitconversion.farad.to.millifarad(x = 1)

unitconversion.F.to.mF(x = 1)

unitconversion.ohm.to.milliohm(x = 1)

unitconversion.siemens.to.millisiemens(x = 1)

unitconversion.S.to.mS(x = 1)

unitconversion.weber.to.milliweber(x = 1)

unitconversion.Wb.to.mWb(x = 1)

unitconversion.tesla.to.millitesla(x = 1)

unitconversion.T.to.mT(x = 1)

unitconversion.henry.to.millihenry(x = 1)

unitconversion.H.to.mH(x = 1)

unitconversion.lumen.to.millilumen(x = 1)

unitconversion.lm.to.mlm(x = 1)

unitconversion.lux.to.millilux(x = 1)

unitconversion.lx.to.mlx(x = 1)

unitconversion.becquerel.to.millibecquerel(x = 1)

unitconversion.Bq.to.mBq(x = 1)

unitconversion.gray.to.milligray(x = 1)

unitconversion.Gy.to.mGy(x = 1)

unitconversion.sievert.to.millisievert(x = 1)

unitconversion.Sv.to.mSv(x = 1)

unitconversion.katal.to.millikatal(x = 1)

unitconversion.kat.to.mkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to milli-units
}
\description{
Milli- thousandths, 1/1000, or 10^(-3)
}
\details{
Performs a conversion from base units to milli-units (ex. grams to milligrams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
