% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.nano.R
\name{unitconversion.siprefix.base.to.nano}
\alias{unitconversion.siprefix.base.to.nano}
\alias{unitconversion.second.to.nanosecond}
\alias{unitconversion.s.to.ns}
\alias{unitconversion.meter.to.nanometer}
\alias{unitconversion.m.to.nm}
\alias{unitconversion.gram.to.nanogram}
\alias{unitconversion.g.to.ng}
\alias{unitconversion.ampere.to.nanoampere}
\alias{unitconversion.A.to.nA}
\alias{unitconversion.kelvin.to.nanokelvin}
\alias{unitconversion.K.to.nK}
\alias{unitconversion.mole.to.nanomole}
\alias{unitconversion.mol.to.nmol}
\alias{unitconversion.candela.to.nanocandela}
\alias{unitconversion.cd.to.ncd}
\alias{unitconversion.radian.to.nanoradian}
\alias{unitconversion.rad.to.nrad}
\alias{unitconversion.steradian.to.nanosteradian}
\alias{unitconversion.sr.to.nsr}
\alias{unitconversion.hertz.to.nanohertz}
\alias{unitconversion.Hz.to.nHz}
\alias{unitconversion.newton.to.nanonewton}
\alias{unitconversion.N.to.nN}
\alias{unitconversion.pascal.to.nanopascal}
\alias{unitconversion.Pa.to.nPa}
\alias{unitconversion.joule.to.nanojoule}
\alias{unitconversion.J.to.nJ}
\alias{unitconversion.watt.to.nanowatt}
\alias{unitconversion.W.to.nW}
\alias{unitconversion.coulomb.to.nanocoulomb}
\alias{unitconversion.C.to.nC}
\alias{unitconversion.volt.to.nanovolt}
\alias{unitconversion.V.to.nV}
\alias{unitconversion.farad.to.nanofarad}
\alias{unitconversion.F.to.nF}
\alias{unitconversion.ohm.to.nanoohm}
\alias{unitconversion.siemens.to.nanosiemens}
\alias{unitconversion.S.to.nS}
\alias{unitconversion.weber.to.nanoweber}
\alias{unitconversion.Wb.to.nWb}
\alias{unitconversion.tesla.to.nanotesla}
\alias{unitconversion.T.to.nT}
\alias{unitconversion.henry.to.nanohenry}
\alias{unitconversion.H.to.nH}
\alias{unitconversion.lumen.to.nanolumen}
\alias{unitconversion.lm.to.nlm}
\alias{unitconversion.lux.to.nanolux}
\alias{unitconversion.lx.to.nlx}
\alias{unitconversion.becquerel.to.nanobecquerel}
\alias{unitconversion.Bq.to.nBq}
\alias{unitconversion.gray.to.nanogray}
\alias{unitconversion.Gy.to.nGy}
\alias{unitconversion.sievert.to.nanosievert}
\alias{unitconversion.Sv.to.nSv}
\alias{unitconversion.katal.to.nanokatal}
\alias{unitconversion.kat.to.nkat}
\title{Unit Conversion - SI Prefixes - Base to Nano-}
\usage{
unitconversion.siprefix.base.to.nano(x = 1)

unitconversion.second.to.nanosecond(x = 1)

unitconversion.s.to.ns(x = 1)

unitconversion.meter.to.nanometer(x = 1)

unitconversion.m.to.nm(x = 1)

unitconversion.gram.to.nanogram(x = 1)

unitconversion.g.to.ng(x = 1)

unitconversion.ampere.to.nanoampere(x = 1)

unitconversion.A.to.nA(x = 1)

unitconversion.kelvin.to.nanokelvin(x = 1)

unitconversion.K.to.nK(x = 1)

unitconversion.mole.to.nanomole(x = 1)

unitconversion.mol.to.nmol(x = 1)

unitconversion.candela.to.nanocandela(x = 1)

unitconversion.cd.to.ncd(x = 1)

unitconversion.radian.to.nanoradian(x = 1)

unitconversion.rad.to.nrad(x = 1)

unitconversion.steradian.to.nanosteradian(x = 1)

unitconversion.sr.to.nsr(x = 1)

unitconversion.hertz.to.nanohertz(x = 1)

unitconversion.Hz.to.nHz(x = 1)

unitconversion.newton.to.nanonewton(x = 1)

unitconversion.N.to.nN(x = 1)

unitconversion.pascal.to.nanopascal(x = 1)

unitconversion.Pa.to.nPa(x = 1)

unitconversion.joule.to.nanojoule(x = 1)

unitconversion.J.to.nJ(x = 1)

unitconversion.watt.to.nanowatt(x = 1)

unitconversion.W.to.nW(x = 1)

unitconversion.coulomb.to.nanocoulomb(x = 1)

unitconversion.C.to.nC(x = 1)

unitconversion.volt.to.nanovolt(x = 1)

unitconversion.V.to.nV(x = 1)

unitconversion.farad.to.nanofarad(x = 1)

unitconversion.F.to.nF(x = 1)

unitconversion.ohm.to.nanoohm(x = 1)

unitconversion.siemens.to.nanosiemens(x = 1)

unitconversion.S.to.nS(x = 1)

unitconversion.weber.to.nanoweber(x = 1)

unitconversion.Wb.to.nWb(x = 1)

unitconversion.tesla.to.nanotesla(x = 1)

unitconversion.T.to.nT(x = 1)

unitconversion.henry.to.nanohenry(x = 1)

unitconversion.H.to.nH(x = 1)

unitconversion.lumen.to.nanolumen(x = 1)

unitconversion.lm.to.nlm(x = 1)

unitconversion.lux.to.nanolux(x = 1)

unitconversion.lx.to.nlx(x = 1)

unitconversion.becquerel.to.nanobecquerel(x = 1)

unitconversion.Bq.to.nBq(x = 1)

unitconversion.gray.to.nanogray(x = 1)

unitconversion.Gy.to.nGy(x = 1)

unitconversion.sievert.to.nanosievert(x = 1)

unitconversion.Sv.to.nSv(x = 1)

unitconversion.katal.to.nanokatal(x = 1)

unitconversion.kat.to.nkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to nano-units
}
\description{
Nano- billionths, 1/1 000 000 000, or 10^(-9)
}
\details{
Performs a conversion from base units to nano-units (ex. grams to nanograms).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
