% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.peta.R
\name{unitconversion.siprefix.base.to.peta}
\alias{unitconversion.siprefix.base.to.peta}
\alias{unitconversion.second.to.petasecond}
\alias{unitconversion.s.to.Ps}
\alias{unitconversion.meter.to.petameter}
\alias{unitconversion.m.to.Pm}
\alias{unitconversion.gram.to.petagram}
\alias{unitconversion.g.to.Pg}
\alias{unitconversion.ampere.to.petaampere}
\alias{unitconversion.A.to.PA}
\alias{unitconversion.kelvin.to.petakelvin}
\alias{unitconversion.K.to.PK}
\alias{unitconversion.mole.to.petamole}
\alias{unitconversion.mol.to.Pmol}
\alias{unitconversion.candela.to.petacandela}
\alias{unitconversion.cd.to.Pcd}
\alias{unitconversion.radian.to.petaradian}
\alias{unitconversion.rad.to.Prad}
\alias{unitconversion.steradian.to.petasteradian}
\alias{unitconversion.sr.to.Psr}
\alias{unitconversion.hertz.to.petahertz}
\alias{unitconversion.Hz.to.PHz}
\alias{unitconversion.newton.to.petanewton}
\alias{unitconversion.N.to.PN}
\alias{unitconversion.pascal.to.petapascal}
\alias{unitconversion.Pa.to.PPa}
\alias{unitconversion.joule.to.petajoule}
\alias{unitconversion.J.to.PJ}
\alias{unitconversion.watt.to.petawatt}
\alias{unitconversion.W.to.PW}
\alias{unitconversion.coulomb.to.petacoulomb}
\alias{unitconversion.C.to.PC}
\alias{unitconversion.volt.to.petavolt}
\alias{unitconversion.V.to.PV}
\alias{unitconversion.farad.to.petafarad}
\alias{unitconversion.F.to.PF}
\alias{unitconversion.ohm.to.petaohm}
\alias{unitconversion.siemens.to.petasiemens}
\alias{unitconversion.S.to.PS}
\alias{unitconversion.weber.to.petaweber}
\alias{unitconversion.Wb.to.PWb}
\alias{unitconversion.tesla.to.petatesla}
\alias{unitconversion.T.to.PT}
\alias{unitconversion.henry.to.petahenry}
\alias{unitconversion.H.to.PH}
\alias{unitconversion.lumen.to.petalumen}
\alias{unitconversion.lm.to.Plm}
\alias{unitconversion.lux.to.petalux}
\alias{unitconversion.lx.to.Plx}
\alias{unitconversion.becquerel.to.petabecquerel}
\alias{unitconversion.Bq.to.PBq}
\alias{unitconversion.gray.to.petagray}
\alias{unitconversion.Gy.to.PGy}
\alias{unitconversion.sievert.to.petasievert}
\alias{unitconversion.Sv.to.PSv}
\alias{unitconversion.katal.to.petakatal}
\alias{unitconversion.kat.to.Pkat}
\title{Unit Conversion - SI Prefixes - Base to Peta-}
\usage{
unitconversion.siprefix.base.to.peta(x = 1)

unitconversion.second.to.petasecond(x = 1)

unitconversion.s.to.Ps(x = 1)

unitconversion.meter.to.petameter(x = 1)

unitconversion.m.to.Pm(x = 1)

unitconversion.gram.to.petagram(x = 1)

unitconversion.g.to.Pg(x = 1)

unitconversion.ampere.to.petaampere(x = 1)

unitconversion.A.to.PA(x = 1)

unitconversion.kelvin.to.petakelvin(x = 1)

unitconversion.K.to.PK(x = 1)

unitconversion.mole.to.petamole(x = 1)

unitconversion.mol.to.Pmol(x = 1)

unitconversion.candela.to.petacandela(x = 1)

unitconversion.cd.to.Pcd(x = 1)

unitconversion.radian.to.petaradian(x = 1)

unitconversion.rad.to.Prad(x = 1)

unitconversion.steradian.to.petasteradian(x = 1)

unitconversion.sr.to.Psr(x = 1)

unitconversion.hertz.to.petahertz(x = 1)

unitconversion.Hz.to.PHz(x = 1)

unitconversion.newton.to.petanewton(x = 1)

unitconversion.N.to.PN(x = 1)

unitconversion.pascal.to.petapascal(x = 1)

unitconversion.Pa.to.PPa(x = 1)

unitconversion.joule.to.petajoule(x = 1)

unitconversion.J.to.PJ(x = 1)

unitconversion.watt.to.petawatt(x = 1)

unitconversion.W.to.PW(x = 1)

unitconversion.coulomb.to.petacoulomb(x = 1)

unitconversion.C.to.PC(x = 1)

unitconversion.volt.to.petavolt(x = 1)

unitconversion.V.to.PV(x = 1)

unitconversion.farad.to.petafarad(x = 1)

unitconversion.F.to.PF(x = 1)

unitconversion.ohm.to.petaohm(x = 1)

unitconversion.siemens.to.petasiemens(x = 1)

unitconversion.S.to.PS(x = 1)

unitconversion.weber.to.petaweber(x = 1)

unitconversion.Wb.to.PWb(x = 1)

unitconversion.tesla.to.petatesla(x = 1)

unitconversion.T.to.PT(x = 1)

unitconversion.henry.to.petahenry(x = 1)

unitconversion.H.to.PH(x = 1)

unitconversion.lumen.to.petalumen(x = 1)

unitconversion.lm.to.Plm(x = 1)

unitconversion.lux.to.petalux(x = 1)

unitconversion.lx.to.Plx(x = 1)

unitconversion.becquerel.to.petabecquerel(x = 1)

unitconversion.Bq.to.PBq(x = 1)

unitconversion.gray.to.petagray(x = 1)

unitconversion.Gy.to.PGy(x = 1)

unitconversion.sievert.to.petasievert(x = 1)

unitconversion.Sv.to.PSv(x = 1)

unitconversion.katal.to.petakatal(x = 1)

unitconversion.kat.to.Pkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to peta-units
}
\description{
Peta- quadrillions, 1 000 000 000 000 000, or 10^15
}
\details{
Performs a conversion from base units to peta-units (ex. grams to petagrams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
