% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.pico.R
\name{unitconversion.siprefix.base.to.pico}
\alias{unitconversion.siprefix.base.to.pico}
\alias{unitconversion.second.to.picosecond}
\alias{unitconversion.s.to.ps}
\alias{unitconversion.meter.to.picometer}
\alias{unitconversion.m.to.pm}
\alias{unitconversion.gram.to.picogram}
\alias{unitconversion.g.to.pg}
\alias{unitconversion.ampere.to.picoampere}
\alias{unitconversion.A.to.pA}
\alias{unitconversion.kelvin.to.picokelvin}
\alias{unitconversion.K.to.pK}
\alias{unitconversion.mole.to.picomole}
\alias{unitconversion.mol.to.pmol}
\alias{unitconversion.candela.to.picocandela}
\alias{unitconversion.cd.to.pcd}
\alias{unitconversion.radian.to.picoradian}
\alias{unitconversion.rad.to.prad}
\alias{unitconversion.steradian.to.picosteradian}
\alias{unitconversion.sr.to.psr}
\alias{unitconversion.hertz.to.picohertz}
\alias{unitconversion.Hz.to.pHz}
\alias{unitconversion.newton.to.piconewton}
\alias{unitconversion.N.to.pN}
\alias{unitconversion.pascal.to.picopascal}
\alias{unitconversion.Pa.to.pPa}
\alias{unitconversion.joule.to.picojoule}
\alias{unitconversion.J.to.pJ}
\alias{unitconversion.watt.to.picowatt}
\alias{unitconversion.W.to.pW}
\alias{unitconversion.coulomb.to.picocoulomb}
\alias{unitconversion.C.to.pC}
\alias{unitconversion.volt.to.picovolt}
\alias{unitconversion.V.to.pV}
\alias{unitconversion.farad.to.picofarad}
\alias{unitconversion.F.to.pF}
\alias{unitconversion.ohm.to.picoohm}
\alias{unitconversion.siemens.to.picosiemens}
\alias{unitconversion.S.to.pS}
\alias{unitconversion.weber.to.picoweber}
\alias{unitconversion.Wb.to.pWb}
\alias{unitconversion.tesla.to.picotesla}
\alias{unitconversion.T.to.pT}
\alias{unitconversion.henry.to.picohenry}
\alias{unitconversion.H.to.pH}
\alias{unitconversion.lumen.to.picolumen}
\alias{unitconversion.lm.to.plm}
\alias{unitconversion.lux.to.picolux}
\alias{unitconversion.lx.to.plx}
\alias{unitconversion.becquerel.to.picobecquerel}
\alias{unitconversion.Bq.to.pBq}
\alias{unitconversion.gray.to.picogray}
\alias{unitconversion.Gy.to.pGy}
\alias{unitconversion.sievert.to.picosievert}
\alias{unitconversion.Sv.to.pSv}
\alias{unitconversion.katal.to.picokatal}
\alias{unitconversion.kat.to.pkat}
\title{Unit Conversion - SI Prefixes - Base to Pico-}
\usage{
unitconversion.siprefix.base.to.pico(x = 1)

unitconversion.second.to.picosecond(x = 1)

unitconversion.s.to.ps(x = 1)

unitconversion.meter.to.picometer(x = 1)

unitconversion.m.to.pm(x = 1)

unitconversion.gram.to.picogram(x = 1)

unitconversion.g.to.pg(x = 1)

unitconversion.ampere.to.picoampere(x = 1)

unitconversion.A.to.pA(x = 1)

unitconversion.kelvin.to.picokelvin(x = 1)

unitconversion.K.to.pK(x = 1)

unitconversion.mole.to.picomole(x = 1)

unitconversion.mol.to.pmol(x = 1)

unitconversion.candela.to.picocandela(x = 1)

unitconversion.cd.to.pcd(x = 1)

unitconversion.radian.to.picoradian(x = 1)

unitconversion.rad.to.prad(x = 1)

unitconversion.steradian.to.picosteradian(x = 1)

unitconversion.sr.to.psr(x = 1)

unitconversion.hertz.to.picohertz(x = 1)

unitconversion.Hz.to.pHz(x = 1)

unitconversion.newton.to.piconewton(x = 1)

unitconversion.N.to.pN(x = 1)

unitconversion.pascal.to.picopascal(x = 1)

unitconversion.Pa.to.pPa(x = 1)

unitconversion.joule.to.picojoule(x = 1)

unitconversion.J.to.pJ(x = 1)

unitconversion.watt.to.picowatt(x = 1)

unitconversion.W.to.pW(x = 1)

unitconversion.coulomb.to.picocoulomb(x = 1)

unitconversion.C.to.pC(x = 1)

unitconversion.volt.to.picovolt(x = 1)

unitconversion.V.to.pV(x = 1)

unitconversion.farad.to.picofarad(x = 1)

unitconversion.F.to.pF(x = 1)

unitconversion.ohm.to.picoohm(x = 1)

unitconversion.siemens.to.picosiemens(x = 1)

unitconversion.S.to.pS(x = 1)

unitconversion.weber.to.picoweber(x = 1)

unitconversion.Wb.to.pWb(x = 1)

unitconversion.tesla.to.picotesla(x = 1)

unitconversion.T.to.pT(x = 1)

unitconversion.henry.to.picohenry(x = 1)

unitconversion.H.to.pH(x = 1)

unitconversion.lumen.to.picolumen(x = 1)

unitconversion.lm.to.plm(x = 1)

unitconversion.lux.to.picolux(x = 1)

unitconversion.lx.to.plx(x = 1)

unitconversion.becquerel.to.picobecquerel(x = 1)

unitconversion.Bq.to.pBq(x = 1)

unitconversion.gray.to.picogray(x = 1)

unitconversion.Gy.to.pGy(x = 1)

unitconversion.sievert.to.picosievert(x = 1)

unitconversion.Sv.to.pSv(x = 1)

unitconversion.katal.to.picokatal(x = 1)

unitconversion.kat.to.pkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to pico-units
}
\description{
Pico- trillionths, 1/1 000 000 000 000, or 10^(-12)
}
\details{
Performs a conversion from base units to pico-units (ex. grams to picograms).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
