% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.tera.R
\name{unitconversion.siprefix.base.to.tera}
\alias{unitconversion.siprefix.base.to.tera}
\alias{unitconversion.second.to.terasecond}
\alias{unitconversion.s.to.Ts}
\alias{unitconversion.meter.to.terameter}
\alias{unitconversion.m.to.Tm}
\alias{unitconversion.gram.to.teragram}
\alias{unitconversion.g.to.Tg}
\alias{unitconversion.ampere.to.teraampere}
\alias{unitconversion.A.to.TA}
\alias{unitconversion.kelvin.to.terakelvin}
\alias{unitconversion.K.to.TK}
\alias{unitconversion.mole.to.teramole}
\alias{unitconversion.mol.to.Tmol}
\alias{unitconversion.candela.to.teracandela}
\alias{unitconversion.cd.to.Tcd}
\alias{unitconversion.radian.to.teraradian}
\alias{unitconversion.rad.to.Trad}
\alias{unitconversion.steradian.to.terasteradian}
\alias{unitconversion.sr.to.Tsr}
\alias{unitconversion.hertz.to.terahertz}
\alias{unitconversion.Hz.to.THz}
\alias{unitconversion.newton.to.teranewton}
\alias{unitconversion.N.to.TN}
\alias{unitconversion.pascal.to.terapascal}
\alias{unitconversion.Pa.to.TPa}
\alias{unitconversion.joule.to.terajoule}
\alias{unitconversion.J.to.TJ}
\alias{unitconversion.watt.to.terawatt}
\alias{unitconversion.W.to.TW}
\alias{unitconversion.coulomb.to.teracoulomb}
\alias{unitconversion.C.to.TC}
\alias{unitconversion.volt.to.teravolt}
\alias{unitconversion.V.to.TV}
\alias{unitconversion.farad.to.terafarad}
\alias{unitconversion.F.to.TF}
\alias{unitconversion.ohm.to.teraohm}
\alias{unitconversion.siemens.to.terasiemens}
\alias{unitconversion.S.to.TS}
\alias{unitconversion.weber.to.teraweber}
\alias{unitconversion.Wb.to.TWb}
\alias{unitconversion.tesla.to.teratesla}
\alias{unitconversion.T.to.TT}
\alias{unitconversion.henry.to.terahenry}
\alias{unitconversion.H.to.TH}
\alias{unitconversion.lumen.to.teralumen}
\alias{unitconversion.lm.to.Tlm}
\alias{unitconversion.lux.to.teralux}
\alias{unitconversion.lx.to.Tlx}
\alias{unitconversion.becquerel.to.terabecquerel}
\alias{unitconversion.Bq.to.TBq}
\alias{unitconversion.gray.to.teragray}
\alias{unitconversion.Gy.to.TGy}
\alias{unitconversion.sievert.to.terasievert}
\alias{unitconversion.Sv.to.TSv}
\alias{unitconversion.katal.to.terakatal}
\alias{unitconversion.kat.to.Tkat}
\title{Unit Conversion - SI Prefixes - Base to Tera-}
\usage{
unitconversion.siprefix.base.to.tera(x = 1)

unitconversion.second.to.terasecond(x = 1)

unitconversion.s.to.Ts(x = 1)

unitconversion.meter.to.terameter(x = 1)

unitconversion.m.to.Tm(x = 1)

unitconversion.gram.to.teragram(x = 1)

unitconversion.g.to.Tg(x = 1)

unitconversion.ampere.to.teraampere(x = 1)

unitconversion.A.to.TA(x = 1)

unitconversion.kelvin.to.terakelvin(x = 1)

unitconversion.K.to.TK(x = 1)

unitconversion.mole.to.teramole(x = 1)

unitconversion.mol.to.Tmol(x = 1)

unitconversion.candela.to.teracandela(x = 1)

unitconversion.cd.to.Tcd(x = 1)

unitconversion.radian.to.teraradian(x = 1)

unitconversion.rad.to.Trad(x = 1)

unitconversion.steradian.to.terasteradian(x = 1)

unitconversion.sr.to.Tsr(x = 1)

unitconversion.hertz.to.terahertz(x = 1)

unitconversion.Hz.to.THz(x = 1)

unitconversion.newton.to.teranewton(x = 1)

unitconversion.N.to.TN(x = 1)

unitconversion.pascal.to.terapascal(x = 1)

unitconversion.Pa.to.TPa(x = 1)

unitconversion.joule.to.terajoule(x = 1)

unitconversion.J.to.TJ(x = 1)

unitconversion.watt.to.terawatt(x = 1)

unitconversion.W.to.TW(x = 1)

unitconversion.coulomb.to.teracoulomb(x = 1)

unitconversion.C.to.TC(x = 1)

unitconversion.volt.to.teravolt(x = 1)

unitconversion.V.to.TV(x = 1)

unitconversion.farad.to.terafarad(x = 1)

unitconversion.F.to.TF(x = 1)

unitconversion.ohm.to.teraohm(x = 1)

unitconversion.siemens.to.terasiemens(x = 1)

unitconversion.S.to.TS(x = 1)

unitconversion.weber.to.teraweber(x = 1)

unitconversion.Wb.to.TWb(x = 1)

unitconversion.tesla.to.teratesla(x = 1)

unitconversion.T.to.TT(x = 1)

unitconversion.henry.to.terahenry(x = 1)

unitconversion.H.to.TH(x = 1)

unitconversion.lumen.to.teralumen(x = 1)

unitconversion.lm.to.Tlm(x = 1)

unitconversion.lux.to.teralux(x = 1)

unitconversion.lx.to.Tlx(x = 1)

unitconversion.becquerel.to.terabecquerel(x = 1)

unitconversion.Bq.to.TBq(x = 1)

unitconversion.gray.to.teragray(x = 1)

unitconversion.Gy.to.TGy(x = 1)

unitconversion.sievert.to.terasievert(x = 1)

unitconversion.Sv.to.TSv(x = 1)

unitconversion.katal.to.terakatal(x = 1)

unitconversion.kat.to.Tkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to tera-units
}
\description{
Tera- trillions, 1 000 000 000 000, or 10^12
}
\details{
Performs a conversion from base units to tera-units (ex. grams to teragrams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
