% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.yocto.R
\name{unitconversion.siprefix.base.to.yocto}
\alias{unitconversion.siprefix.base.to.yocto}
\alias{unitconversion.second.to.yoctosecond}
\alias{unitconversion.s.to.ys}
\alias{unitconversion.meter.to.yoctometer}
\alias{unitconversion.m.to.ym}
\alias{unitconversion.gram.to.yoctogram}
\alias{unitconversion.g.to.yg}
\alias{unitconversion.ampere.to.yoctoampere}
\alias{unitconversion.A.to.yA}
\alias{unitconversion.kelvin.to.yoctokelvin}
\alias{unitconversion.K.to.yK}
\alias{unitconversion.mole.to.yoctomole}
\alias{unitconversion.mol.to.ymol}
\alias{unitconversion.candela.to.yoctocandela}
\alias{unitconversion.cd.to.ycd}
\alias{unitconversion.radian.to.yoctoradian}
\alias{unitconversion.rad.to.yrad}
\alias{unitconversion.steradian.to.yoctosteradian}
\alias{unitconversion.sr.to.ysr}
\alias{unitconversion.hertz.to.yoctohertz}
\alias{unitconversion.Hz.to.yHz}
\alias{unitconversion.newton.to.yoctonewton}
\alias{unitconversion.N.to.yN}
\alias{unitconversion.pascal.to.yoctopascal}
\alias{unitconversion.Pa.to.yPa}
\alias{unitconversion.joule.to.yoctojoule}
\alias{unitconversion.J.to.yJ}
\alias{unitconversion.watt.to.yoctowatt}
\alias{unitconversion.W.to.yW}
\alias{unitconversion.coulomb.to.yoctocoulomb}
\alias{unitconversion.C.to.yC}
\alias{unitconversion.volt.to.yoctovolt}
\alias{unitconversion.V.to.yV}
\alias{unitconversion.farad.to.yoctofarad}
\alias{unitconversion.F.to.yF}
\alias{unitconversion.ohm.to.yoctoohm}
\alias{unitconversion.siemens.to.yoctosiemens}
\alias{unitconversion.S.to.yS}
\alias{unitconversion.weber.to.yoctoweber}
\alias{unitconversion.Wb.to.yWb}
\alias{unitconversion.tesla.to.yoctotesla}
\alias{unitconversion.T.to.yT}
\alias{unitconversion.henry.to.yoctohenry}
\alias{unitconversion.H.to.yH}
\alias{unitconversion.lumen.to.yoctolumen}
\alias{unitconversion.lm.to.ylm}
\alias{unitconversion.lux.to.yoctolux}
\alias{unitconversion.lx.to.ylx}
\alias{unitconversion.becquerel.to.yoctobecquerel}
\alias{unitconversion.Bq.to.yBq}
\alias{unitconversion.gray.to.yoctogray}
\alias{unitconversion.Gy.to.yGy}
\alias{unitconversion.sievert.to.yoctosievert}
\alias{unitconversion.Sv.to.ySv}
\alias{unitconversion.katal.to.yoctokatal}
\alias{unitconversion.kat.to.ykat}
\title{Unit Conversion - SI Prefixes - Base to Yocto-}
\usage{
unitconversion.siprefix.base.to.yocto(x = 1)

unitconversion.second.to.yoctosecond(x = 1)

unitconversion.s.to.ys(x = 1)

unitconversion.meter.to.yoctometer(x = 1)

unitconversion.m.to.ym(x = 1)

unitconversion.gram.to.yoctogram(x = 1)

unitconversion.g.to.yg(x = 1)

unitconversion.ampere.to.yoctoampere(x = 1)

unitconversion.A.to.yA(x = 1)

unitconversion.kelvin.to.yoctokelvin(x = 1)

unitconversion.K.to.yK(x = 1)

unitconversion.mole.to.yoctomole(x = 1)

unitconversion.mol.to.ymol(x = 1)

unitconversion.candela.to.yoctocandela(x = 1)

unitconversion.cd.to.ycd(x = 1)

unitconversion.radian.to.yoctoradian(x = 1)

unitconversion.rad.to.yrad(x = 1)

unitconversion.steradian.to.yoctosteradian(x = 1)

unitconversion.sr.to.ysr(x = 1)

unitconversion.hertz.to.yoctohertz(x = 1)

unitconversion.Hz.to.yHz(x = 1)

unitconversion.newton.to.yoctonewton(x = 1)

unitconversion.N.to.yN(x = 1)

unitconversion.pascal.to.yoctopascal(x = 1)

unitconversion.Pa.to.yPa(x = 1)

unitconversion.joule.to.yoctojoule(x = 1)

unitconversion.J.to.yJ(x = 1)

unitconversion.watt.to.yoctowatt(x = 1)

unitconversion.W.to.yW(x = 1)

unitconversion.coulomb.to.yoctocoulomb(x = 1)

unitconversion.C.to.yC(x = 1)

unitconversion.volt.to.yoctovolt(x = 1)

unitconversion.V.to.yV(x = 1)

unitconversion.farad.to.yoctofarad(x = 1)

unitconversion.F.to.yF(x = 1)

unitconversion.ohm.to.yoctoohm(x = 1)

unitconversion.siemens.to.yoctosiemens(x = 1)

unitconversion.S.to.yS(x = 1)

unitconversion.weber.to.yoctoweber(x = 1)

unitconversion.Wb.to.yWb(x = 1)

unitconversion.tesla.to.yoctotesla(x = 1)

unitconversion.T.to.yT(x = 1)

unitconversion.henry.to.yoctohenry(x = 1)

unitconversion.H.to.yH(x = 1)

unitconversion.lumen.to.yoctolumen(x = 1)

unitconversion.lm.to.ylm(x = 1)

unitconversion.lux.to.yoctolux(x = 1)

unitconversion.lx.to.ylx(x = 1)

unitconversion.becquerel.to.yoctobecquerel(x = 1)

unitconversion.Bq.to.yBq(x = 1)

unitconversion.gray.to.yoctogray(x = 1)

unitconversion.Gy.to.yGy(x = 1)

unitconversion.sievert.to.yoctosievert(x = 1)

unitconversion.Sv.to.ySv(x = 1)

unitconversion.katal.to.yoctokatal(x = 1)

unitconversion.kat.to.ykat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to yocto-units
}
\description{
Yocto- means 1/1 000 000 000 000 000 000 000 000 or 10^(-24)
}
\details{
Performs a conversion from base units to yocto-units (ex. grams to yoctograms).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
