% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.yotta.R
\name{unitconversion.siprefix.base.to.yotta}
\alias{unitconversion.siprefix.base.to.yotta}
\alias{unitconversion.second.to.yottasecond}
\alias{unitconversion.s.to.Ys}
\alias{unitconversion.meter.to.yottameter}
\alias{unitconversion.m.to.Ym}
\alias{unitconversion.gram.to.yottagram}
\alias{unitconversion.g.to.Yg}
\alias{unitconversion.ampere.to.yottaampere}
\alias{unitconversion.A.to.YA}
\alias{unitconversion.kelvin.to.yottakelvin}
\alias{unitconversion.K.to.YK}
\alias{unitconversion.mole.to.yottamole}
\alias{unitconversion.mol.to.Ymol}
\alias{unitconversion.candela.to.yottacandela}
\alias{unitconversion.cd.to.Ycd}
\alias{unitconversion.radian.to.yottaradian}
\alias{unitconversion.rad.to.Yrad}
\alias{unitconversion.steradian.to.yottasteradian}
\alias{unitconversion.sr.to.Ysr}
\alias{unitconversion.hertz.to.yottahertz}
\alias{unitconversion.Hz.to.YHz}
\alias{unitconversion.newton.to.yottanewton}
\alias{unitconversion.N.to.YN}
\alias{unitconversion.pascal.to.yottapascal}
\alias{unitconversion.Pa.to.YPa}
\alias{unitconversion.joule.to.yottajoule}
\alias{unitconversion.J.to.YJ}
\alias{unitconversion.watt.to.yottawatt}
\alias{unitconversion.W.to.YW}
\alias{unitconversion.coulomb.to.yottacoulomb}
\alias{unitconversion.C.to.YC}
\alias{unitconversion.volt.to.yottavolt}
\alias{unitconversion.V.to.YV}
\alias{unitconversion.farad.to.yottafarad}
\alias{unitconversion.F.to.YF}
\alias{unitconversion.ohm.to.yottaohm}
\alias{unitconversion.siemens.to.yottasiemens}
\alias{unitconversion.S.to.YS}
\alias{unitconversion.weber.to.yottaweber}
\alias{unitconversion.Wb.to.YWb}
\alias{unitconversion.tesla.to.yottatesla}
\alias{unitconversion.T.to.YT}
\alias{unitconversion.henry.to.yottahenry}
\alias{unitconversion.H.to.YH}
\alias{unitconversion.lumen.to.yottalumen}
\alias{unitconversion.lm.to.Ylm}
\alias{unitconversion.lux.to.yottalux}
\alias{unitconversion.lx.to.Ylx}
\alias{unitconversion.becquerel.to.yottabecquerel}
\alias{unitconversion.Bq.to.YBq}
\alias{unitconversion.gray.to.yottagray}
\alias{unitconversion.Gy.to.YGy}
\alias{unitconversion.sievert.to.yottasievert}
\alias{unitconversion.Sv.to.YSv}
\alias{unitconversion.katal.to.yottakatal}
\alias{unitconversion.kat.to.Ykat}
\title{Unit Conversion - SI Prefixes - Base to Yotta-}
\usage{
unitconversion.siprefix.base.to.yotta(x = 1)

unitconversion.second.to.yottasecond(x = 1)

unitconversion.s.to.Ys(x = 1)

unitconversion.meter.to.yottameter(x = 1)

unitconversion.m.to.Ym(x = 1)

unitconversion.gram.to.yottagram(x = 1)

unitconversion.g.to.Yg(x = 1)

unitconversion.ampere.to.yottaampere(x = 1)

unitconversion.A.to.YA(x = 1)

unitconversion.kelvin.to.yottakelvin(x = 1)

unitconversion.K.to.YK(x = 1)

unitconversion.mole.to.yottamole(x = 1)

unitconversion.mol.to.Ymol(x = 1)

unitconversion.candela.to.yottacandela(x = 1)

unitconversion.cd.to.Ycd(x = 1)

unitconversion.radian.to.yottaradian(x = 1)

unitconversion.rad.to.Yrad(x = 1)

unitconversion.steradian.to.yottasteradian(x = 1)

unitconversion.sr.to.Ysr(x = 1)

unitconversion.hertz.to.yottahertz(x = 1)

unitconversion.Hz.to.YHz(x = 1)

unitconversion.newton.to.yottanewton(x = 1)

unitconversion.N.to.YN(x = 1)

unitconversion.pascal.to.yottapascal(x = 1)

unitconversion.Pa.to.YPa(x = 1)

unitconversion.joule.to.yottajoule(x = 1)

unitconversion.J.to.YJ(x = 1)

unitconversion.watt.to.yottawatt(x = 1)

unitconversion.W.to.YW(x = 1)

unitconversion.coulomb.to.yottacoulomb(x = 1)

unitconversion.C.to.YC(x = 1)

unitconversion.volt.to.yottavolt(x = 1)

unitconversion.V.to.YV(x = 1)

unitconversion.farad.to.yottafarad(x = 1)

unitconversion.F.to.YF(x = 1)

unitconversion.ohm.to.yottaohm(x = 1)

unitconversion.siemens.to.yottasiemens(x = 1)

unitconversion.S.to.YS(x = 1)

unitconversion.weber.to.yottaweber(x = 1)

unitconversion.Wb.to.YWb(x = 1)

unitconversion.tesla.to.yottatesla(x = 1)

unitconversion.T.to.YT(x = 1)

unitconversion.henry.to.yottahenry(x = 1)

unitconversion.H.to.YH(x = 1)

unitconversion.lumen.to.yottalumen(x = 1)

unitconversion.lm.to.Ylm(x = 1)

unitconversion.lux.to.yottalux(x = 1)

unitconversion.lx.to.Ylx(x = 1)

unitconversion.becquerel.to.yottabecquerel(x = 1)

unitconversion.Bq.to.YBq(x = 1)

unitconversion.gray.to.yottagray(x = 1)

unitconversion.Gy.to.YGy(x = 1)

unitconversion.sievert.to.yottasievert(x = 1)

unitconversion.Sv.to.YSv(x = 1)

unitconversion.katal.to.yottakatal(x = 1)

unitconversion.kat.to.Ykat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to yotta-units
}
\description{
Yotta- means 1 000 000 000 000 000 000 000 000 or 10^24
}
\details{
Performs a conversion from base units to yotta-units (ex. grams to yottagrams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
