% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.zepto.R
\name{unitconversion.siprefix.base.to.zepto}
\alias{unitconversion.siprefix.base.to.zepto}
\alias{unitconversion.second.to.zeptosecond}
\alias{unitconversion.s.to.zs}
\alias{unitconversion.meter.to.zeptometer}
\alias{unitconversion.m.to.zm}
\alias{unitconversion.gram.to.zeptogram}
\alias{unitconversion.g.to.zg}
\alias{unitconversion.ampere.to.zeptoampere}
\alias{unitconversion.A.to.zA}
\alias{unitconversion.kelvin.to.zeptokelvin}
\alias{unitconversion.K.to.zK}
\alias{unitconversion.mole.to.zeptomole}
\alias{unitconversion.mol.to.zmol}
\alias{unitconversion.candela.to.zeptocandela}
\alias{unitconversion.cd.to.zcd}
\alias{unitconversion.radian.to.zeptoradian}
\alias{unitconversion.rad.to.zrad}
\alias{unitconversion.steradian.to.zeptosteradian}
\alias{unitconversion.sr.to.zsr}
\alias{unitconversion.hertz.to.zeptohertz}
\alias{unitconversion.Hz.to.zHz}
\alias{unitconversion.newton.to.zeptonewton}
\alias{unitconversion.N.to.zN}
\alias{unitconversion.pascal.to.zeptopascal}
\alias{unitconversion.Pa.to.zPa}
\alias{unitconversion.joule.to.zeptojoule}
\alias{unitconversion.J.to.zJ}
\alias{unitconversion.watt.to.zeptowatt}
\alias{unitconversion.W.to.zW}
\alias{unitconversion.coulomb.to.zeptocoulomb}
\alias{unitconversion.C.to.zC}
\alias{unitconversion.volt.to.zeptovolt}
\alias{unitconversion.V.to.zV}
\alias{unitconversion.farad.to.zeptofarad}
\alias{unitconversion.F.to.zF}
\alias{unitconversion.ohm.to.zeptoohm}
\alias{unitconversion.siemens.to.zeptosiemens}
\alias{unitconversion.S.to.zS}
\alias{unitconversion.weber.to.zeptoweber}
\alias{unitconversion.Wb.to.zWb}
\alias{unitconversion.tesla.to.zeptotesla}
\alias{unitconversion.T.to.zT}
\alias{unitconversion.henry.to.zeptohenry}
\alias{unitconversion.H.to.zH}
\alias{unitconversion.lumen.to.zeptolumen}
\alias{unitconversion.lm.to.zlm}
\alias{unitconversion.lux.to.zeptolux}
\alias{unitconversion.lx.to.zlx}
\alias{unitconversion.becquerel.to.zeptobecquerel}
\alias{unitconversion.Bq.to.zBq}
\alias{unitconversion.gray.to.zeptogray}
\alias{unitconversion.Gy.to.zGy}
\alias{unitconversion.sievert.to.zeptosievert}
\alias{unitconversion.Sv.to.zSv}
\alias{unitconversion.katal.to.zeptokatal}
\alias{unitconversion.kat.to.zkat}
\title{Unit Conversion - SI Prefixes - Base to Zepto-}
\usage{
unitconversion.siprefix.base.to.zepto(x = 1)

unitconversion.second.to.zeptosecond(x = 1)

unitconversion.s.to.zs(x = 1)

unitconversion.meter.to.zeptometer(x = 1)

unitconversion.m.to.zm(x = 1)

unitconversion.gram.to.zeptogram(x = 1)

unitconversion.g.to.zg(x = 1)

unitconversion.ampere.to.zeptoampere(x = 1)

unitconversion.A.to.zA(x = 1)

unitconversion.kelvin.to.zeptokelvin(x = 1)

unitconversion.K.to.zK(x = 1)

unitconversion.mole.to.zeptomole(x = 1)

unitconversion.mol.to.zmol(x = 1)

unitconversion.candela.to.zeptocandela(x = 1)

unitconversion.cd.to.zcd(x = 1)

unitconversion.radian.to.zeptoradian(x = 1)

unitconversion.rad.to.zrad(x = 1)

unitconversion.steradian.to.zeptosteradian(x = 1)

unitconversion.sr.to.zsr(x = 1)

unitconversion.hertz.to.zeptohertz(x = 1)

unitconversion.Hz.to.zHz(x = 1)

unitconversion.newton.to.zeptonewton(x = 1)

unitconversion.N.to.zN(x = 1)

unitconversion.pascal.to.zeptopascal(x = 1)

unitconversion.Pa.to.zPa(x = 1)

unitconversion.joule.to.zeptojoule(x = 1)

unitconversion.J.to.zJ(x = 1)

unitconversion.watt.to.zeptowatt(x = 1)

unitconversion.W.to.zW(x = 1)

unitconversion.coulomb.to.zeptocoulomb(x = 1)

unitconversion.C.to.zC(x = 1)

unitconversion.volt.to.zeptovolt(x = 1)

unitconversion.V.to.zV(x = 1)

unitconversion.farad.to.zeptofarad(x = 1)

unitconversion.F.to.zF(x = 1)

unitconversion.ohm.to.zeptoohm(x = 1)

unitconversion.siemens.to.zeptosiemens(x = 1)

unitconversion.S.to.zS(x = 1)

unitconversion.weber.to.zeptoweber(x = 1)

unitconversion.Wb.to.zWb(x = 1)

unitconversion.tesla.to.zeptotesla(x = 1)

unitconversion.T.to.zT(x = 1)

unitconversion.henry.to.zeptohenry(x = 1)

unitconversion.H.to.zH(x = 1)

unitconversion.lumen.to.zeptolumen(x = 1)

unitconversion.lm.to.zlm(x = 1)

unitconversion.lux.to.zeptolux(x = 1)

unitconversion.lx.to.zlx(x = 1)

unitconversion.becquerel.to.zeptobecquerel(x = 1)

unitconversion.Bq.to.zBq(x = 1)

unitconversion.gray.to.zeptogray(x = 1)

unitconversion.Gy.to.zGy(x = 1)

unitconversion.sievert.to.zeptosievert(x = 1)

unitconversion.Sv.to.zSv(x = 1)

unitconversion.katal.to.zeptokatal(x = 1)

unitconversion.kat.to.zkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to zepto-units
}
\description{
Zepto- means 1/1 000 000 000 000 000 000 000 or 10^(-21)
}
\details{
Performs a conversion from base units to zepto-units (ex. grams to zeptograms).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
