% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.base.to.zetta.R
\name{unitconversion.siprefix.base.to.zetta}
\alias{unitconversion.siprefix.base.to.zetta}
\alias{unitconversion.second.to.zettasecond}
\alias{unitconversion.s.to.Zs}
\alias{unitconversion.meter.to.zettameter}
\alias{unitconversion.m.to.Zm}
\alias{unitconversion.gram.to.zettagram}
\alias{unitconversion.g.to.Zg}
\alias{unitconversion.ampere.to.zettaampere}
\alias{unitconversion.A.to.ZA}
\alias{unitconversion.kelvin.to.zettakelvin}
\alias{unitconversion.K.to.ZK}
\alias{unitconversion.mole.to.zettamole}
\alias{unitconversion.mol.to.Zmol}
\alias{unitconversion.candela.to.zettacandela}
\alias{unitconversion.cd.to.Zcd}
\alias{unitconversion.radian.to.zettaradian}
\alias{unitconversion.rad.to.Zrad}
\alias{unitconversion.steradian.to.zettasteradian}
\alias{unitconversion.sr.to.Zsr}
\alias{unitconversion.hertz.to.zettahertz}
\alias{unitconversion.Hz.to.ZHz}
\alias{unitconversion.newton.to.zettanewton}
\alias{unitconversion.N.to.ZN}
\alias{unitconversion.pascal.to.zettapascal}
\alias{unitconversion.Pa.to.ZPa}
\alias{unitconversion.joule.to.zettajoule}
\alias{unitconversion.J.to.ZJ}
\alias{unitconversion.watt.to.zettawatt}
\alias{unitconversion.W.to.ZW}
\alias{unitconversion.coulomb.to.zettacoulomb}
\alias{unitconversion.C.to.ZC}
\alias{unitconversion.volt.to.zettavolt}
\alias{unitconversion.V.to.ZV}
\alias{unitconversion.farad.to.zettafarad}
\alias{unitconversion.F.to.ZF}
\alias{unitconversion.ohm.to.zettaohm}
\alias{unitconversion.siemens.to.zettasiemens}
\alias{unitconversion.S.to.ZS}
\alias{unitconversion.weber.to.zettaweber}
\alias{unitconversion.Wb.to.ZWb}
\alias{unitconversion.tesla.to.zettatesla}
\alias{unitconversion.T.to.ZT}
\alias{unitconversion.henry.to.zettahenry}
\alias{unitconversion.H.to.ZH}
\alias{unitconversion.lumen.to.zettalumen}
\alias{unitconversion.lm.to.Zlm}
\alias{unitconversion.lux.to.zettalux}
\alias{unitconversion.lx.to.Zlx}
\alias{unitconversion.becquerel.to.zettabecquerel}
\alias{unitconversion.Bq.to.ZBq}
\alias{unitconversion.gray.to.zettagray}
\alias{unitconversion.Gy.to.ZGy}
\alias{unitconversion.sievert.to.zettasievert}
\alias{unitconversion.Sv.to.ZSv}
\alias{unitconversion.katal.to.zettakatal}
\alias{unitconversion.kat.to.Zkat}
\title{Unit Conversion - SI Prefixes - Base to Zetta-}
\usage{
unitconversion.siprefix.base.to.zetta(x = 1)

unitconversion.second.to.zettasecond(x = 1)

unitconversion.s.to.Zs(x = 1)

unitconversion.meter.to.zettameter(x = 1)

unitconversion.m.to.Zm(x = 1)

unitconversion.gram.to.zettagram(x = 1)

unitconversion.g.to.Zg(x = 1)

unitconversion.ampere.to.zettaampere(x = 1)

unitconversion.A.to.ZA(x = 1)

unitconversion.kelvin.to.zettakelvin(x = 1)

unitconversion.K.to.ZK(x = 1)

unitconversion.mole.to.zettamole(x = 1)

unitconversion.mol.to.Zmol(x = 1)

unitconversion.candela.to.zettacandela(x = 1)

unitconversion.cd.to.Zcd(x = 1)

unitconversion.radian.to.zettaradian(x = 1)

unitconversion.rad.to.Zrad(x = 1)

unitconversion.steradian.to.zettasteradian(x = 1)

unitconversion.sr.to.Zsr(x = 1)

unitconversion.hertz.to.zettahertz(x = 1)

unitconversion.Hz.to.ZHz(x = 1)

unitconversion.newton.to.zettanewton(x = 1)

unitconversion.N.to.ZN(x = 1)

unitconversion.pascal.to.zettapascal(x = 1)

unitconversion.Pa.to.ZPa(x = 1)

unitconversion.joule.to.zettajoule(x = 1)

unitconversion.J.to.ZJ(x = 1)

unitconversion.watt.to.zettawatt(x = 1)

unitconversion.W.to.ZW(x = 1)

unitconversion.coulomb.to.zettacoulomb(x = 1)

unitconversion.C.to.ZC(x = 1)

unitconversion.volt.to.zettavolt(x = 1)

unitconversion.V.to.ZV(x = 1)

unitconversion.farad.to.zettafarad(x = 1)

unitconversion.F.to.ZF(x = 1)

unitconversion.ohm.to.zettaohm(x = 1)

unitconversion.siemens.to.zettasiemens(x = 1)

unitconversion.S.to.ZS(x = 1)

unitconversion.weber.to.zettaweber(x = 1)

unitconversion.Wb.to.ZWb(x = 1)

unitconversion.tesla.to.zettatesla(x = 1)

unitconversion.T.to.ZT(x = 1)

unitconversion.henry.to.zettahenry(x = 1)

unitconversion.H.to.ZH(x = 1)

unitconversion.lumen.to.zettalumen(x = 1)

unitconversion.lm.to.Zlm(x = 1)

unitconversion.lux.to.zettalux(x = 1)

unitconversion.lx.to.Zlx(x = 1)

unitconversion.becquerel.to.zettabecquerel(x = 1)

unitconversion.Bq.to.ZBq(x = 1)

unitconversion.gray.to.zettagray(x = 1)

unitconversion.Gy.to.ZGy(x = 1)

unitconversion.sievert.to.zettasievert(x = 1)

unitconversion.Sv.to.ZSv(x = 1)

unitconversion.katal.to.zettakatal(x = 1)

unitconversion.kat.to.Zkat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of base units}
}
\value{
x, but converted to zetta-units
}
\description{
Zetta- means 1 000 000 000 000 000 000 000 or 10^21
}
\details{
Performs a conversion from base units to zetta-units (ex. grams to zettagrams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
