% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.centi.to.base.R
\name{unitconversion.siprefix.centi.to.base}
\alias{unitconversion.siprefix.centi.to.base}
\alias{unitconversion.centisecond.to.second}
\alias{unitconversion.cs.to.s}
\alias{unitconversion.centimeter.to.meter}
\alias{unitconversion.cm.to.m}
\alias{unitconversion.centigram.to.gram}
\alias{unitconversion.cg.to.g}
\alias{unitconversion.centiampere.to.ampere}
\alias{unitconversion.cA.to.A}
\alias{unitconversion.centikelvin.to.kelvin}
\alias{unitconversion.cK.to.K}
\alias{unitconversion.centimole.to.mole}
\alias{unitconversion.cmol.to.mol}
\alias{unitconversion.centicandela.to.candela}
\alias{unitconversion.ccd.to.cd}
\alias{unitconversion.centiradian.to.radian}
\alias{unitconversion.crad.to.rad}
\alias{unitconversion.centisteradian.to.steradian}
\alias{unitconversion.csr.to.sr}
\alias{unitconversion.centihertz.to.hertz}
\alias{unitconversion.cHz.to.Hz}
\alias{unitconversion.centinewton.to.newton}
\alias{unitconversion.cN.to.N}
\alias{unitconversion.centipascal.to.pascal}
\alias{unitconversion.cPa.to.Pa}
\alias{unitconversion.centijoule.to.joule}
\alias{unitconversion.cJ.to.J}
\alias{unitconversion.centiwatt.to.watt}
\alias{unitconversion.cW.to.W}
\alias{unitconversion.centicoulomb.to.coulomb}
\alias{unitconversion.cC.to.C}
\alias{unitconversion.centivolt.to.volt}
\alias{unitconversion.cV.to.V}
\alias{unitconversion.centifarad.to.farad}
\alias{unitconversion.cF.to.F}
\alias{unitconversion.centiohm.to.ohm}
\alias{unitconversion.centisiemens.to.siemens}
\alias{unitconversion.cS.to.S}
\alias{unitconversion.centiweber.to.weber}
\alias{unitconversion.cWb.to.Wb}
\alias{unitconversion.centitesla.to.tesla}
\alias{unitconversion.cT.to.T}
\alias{unitconversion.centihenry.to.henry}
\alias{unitconversion.cH.to.H}
\alias{unitconversion.centilumen.to.lumen}
\alias{unitconversion.clm.to.lm}
\alias{unitconversion.centilux.to.lux}
\alias{unitconversion.clx.to.lx}
\alias{unitconversion.centibecquerel.to.becquerel}
\alias{unitconversion.cBq.to.Bq}
\alias{unitconversion.centigray.to.gray}
\alias{unitconversion.cGy.to.Gy}
\alias{unitconversion.centisievert.to.sievert}
\alias{unitconversion.cSv.to.Sv}
\alias{unitconversion.centikatal.to.katal}
\alias{unitconversion.ckat.to.kat}
\title{Unit Conversion - SI Prefixes - Centi- to Base}
\usage{
unitconversion.siprefix.centi.to.base(x = 1)

unitconversion.centisecond.to.second(x = 1)

unitconversion.cs.to.s(x = 1)

unitconversion.centimeter.to.meter(x = 1)

unitconversion.cm.to.m(x = 1)

unitconversion.centigram.to.gram(x = 1)

unitconversion.cg.to.g(x = 1)

unitconversion.centiampere.to.ampere(x = 1)

unitconversion.cA.to.A(x = 1)

unitconversion.centikelvin.to.kelvin(x = 1)

unitconversion.cK.to.K(x = 1)

unitconversion.centimole.to.mole(x = 1)

unitconversion.cmol.to.mol(x = 1)

unitconversion.centicandela.to.candela(x = 1)

unitconversion.ccd.to.cd(x = 1)

unitconversion.centiradian.to.radian(x = 1)

unitconversion.crad.to.rad(x = 1)

unitconversion.centisteradian.to.steradian(x = 1)

unitconversion.csr.to.sr(x = 1)

unitconversion.centihertz.to.hertz(x = 1)

unitconversion.cHz.to.Hz(x = 1)

unitconversion.centinewton.to.newton(x = 1)

unitconversion.cN.to.N(x = 1)

unitconversion.centipascal.to.pascal(x = 1)

unitconversion.cPa.to.Pa(x = 1)

unitconversion.centijoule.to.joule(x = 1)

unitconversion.cJ.to.J(x = 1)

unitconversion.centiwatt.to.watt(x = 1)

unitconversion.cW.to.W(x = 1)

unitconversion.centicoulomb.to.coulomb(x = 1)

unitconversion.cC.to.C(x = 1)

unitconversion.centivolt.to.volt(x = 1)

unitconversion.cV.to.V(x = 1)

unitconversion.centifarad.to.farad(x = 1)

unitconversion.cF.to.F(x = 1)

unitconversion.centiohm.to.ohm(x = 1)

unitconversion.centisiemens.to.siemens(x = 1)

unitconversion.cS.to.S(x = 1)

unitconversion.centiweber.to.weber(x = 1)

unitconversion.cWb.to.Wb(x = 1)

unitconversion.centitesla.to.tesla(x = 1)

unitconversion.cT.to.T(x = 1)

unitconversion.centihenry.to.henry(x = 1)

unitconversion.cH.to.H(x = 1)

unitconversion.centilumen.to.lumen(x = 1)

unitconversion.clm.to.lm(x = 1)

unitconversion.centilux.to.lux(x = 1)

unitconversion.clx.to.lx(x = 1)

unitconversion.centibecquerel.to.becquerel(x = 1)

unitconversion.cBq.to.Bq(x = 1)

unitconversion.centigray.to.gray(x = 1)

unitconversion.cGy.to.Gy(x = 1)

unitconversion.centisievert.to.sievert(x = 1)

unitconversion.cSv.to.Sv(x = 1)

unitconversion.centikatal.to.katal(x = 1)

unitconversion.ckat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of centi-units}
}
\value{
x, but converted to base units
}
\description{
Centi- hundredths, 1/100, or 10^(-2)
}
\details{
Performs a conversion from centi-units to base units (ex. centigrams to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
