% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.deci.to.base.R
\name{unitconversion.siprefix.deci.to.base}
\alias{unitconversion.siprefix.deci.to.base}
\alias{unitconversion.decisecond.to.second}
\alias{unitconversion.ds.to.s}
\alias{unitconversion.decimeter.to.meter}
\alias{unitconversion.dm.to.m}
\alias{unitconversion.decigram.to.gram}
\alias{unitconversion.dg.to.g}
\alias{unitconversion.deciampere.to.ampere}
\alias{unitconversion.dA.to.A}
\alias{unitconversion.decikelvin.to.kelvin}
\alias{unitconversion.dK.to.K}
\alias{unitconversion.decimole.to.mole}
\alias{unitconversion.dmol.to.mol}
\alias{unitconversion.decicandela.to.candela}
\alias{unitconversion.dcd.to.cd}
\alias{unitconversion.deciradian.to.radian}
\alias{unitconversion.drad.to.rad}
\alias{unitconversion.decisteradian.to.steradian}
\alias{unitconversion.dsr.to.sr}
\alias{unitconversion.decihertz.to.hertz}
\alias{unitconversion.dHz.to.Hz}
\alias{unitconversion.decinewton.to.newton}
\alias{unitconversion.dN.to.N}
\alias{unitconversion.decipascal.to.pascal}
\alias{unitconversion.dPa.to.Pa}
\alias{unitconversion.decijoule.to.joule}
\alias{unitconversion.dJ.to.J}
\alias{unitconversion.deciwatt.to.watt}
\alias{unitconversion.dW.to.W}
\alias{unitconversion.decicoulomb.to.coulomb}
\alias{unitconversion.dC.to.C}
\alias{unitconversion.decivolt.to.volt}
\alias{unitconversion.dV.to.V}
\alias{unitconversion.decifarad.to.farad}
\alias{unitconversion.dF.to.F}
\alias{unitconversion.deciohm.to.ohm}
\alias{unitconversion.decisiemens.to.siemens}
\alias{unitconversion.dS.to.S}
\alias{unitconversion.deciweber.to.weber}
\alias{unitconversion.dWb.to.Wb}
\alias{unitconversion.decitesla.to.tesla}
\alias{unitconversion.dT.to.T}
\alias{unitconversion.decihenry.to.henry}
\alias{unitconversion.dH.to.H}
\alias{unitconversion.decilumen.to.lumen}
\alias{unitconversion.dlm.to.lm}
\alias{unitconversion.decilux.to.lux}
\alias{unitconversion.dlx.to.lx}
\alias{unitconversion.decibecquerel.to.becquerel}
\alias{unitconversion.dBq.to.Bq}
\alias{unitconversion.decigray.to.gray}
\alias{unitconversion.dGy.to.Gy}
\alias{unitconversion.decisievert.to.sievert}
\alias{unitconversion.dSv.to.Sv}
\alias{unitconversion.decikatal.to.katal}
\alias{unitconversion.dkat.to.kat}
\title{Unit Conversion - SI Prefixes - Deci- to Base}
\usage{
unitconversion.siprefix.deci.to.base(x = 1)

unitconversion.decisecond.to.second(x = 1)

unitconversion.ds.to.s(x = 1)

unitconversion.decimeter.to.meter(x = 1)

unitconversion.dm.to.m(x = 1)

unitconversion.decigram.to.gram(x = 1)

unitconversion.dg.to.g(x = 1)

unitconversion.deciampere.to.ampere(x = 1)

unitconversion.dA.to.A(x = 1)

unitconversion.decikelvin.to.kelvin(x = 1)

unitconversion.dK.to.K(x = 1)

unitconversion.decimole.to.mole(x = 1)

unitconversion.dmol.to.mol(x = 1)

unitconversion.decicandela.to.candela(x = 1)

unitconversion.dcd.to.cd(x = 1)

unitconversion.deciradian.to.radian(x = 1)

unitconversion.drad.to.rad(x = 1)

unitconversion.decisteradian.to.steradian(x = 1)

unitconversion.dsr.to.sr(x = 1)

unitconversion.decihertz.to.hertz(x = 1)

unitconversion.dHz.to.Hz(x = 1)

unitconversion.decinewton.to.newton(x = 1)

unitconversion.dN.to.N(x = 1)

unitconversion.decipascal.to.pascal(x = 1)

unitconversion.dPa.to.Pa(x = 1)

unitconversion.decijoule.to.joule(x = 1)

unitconversion.dJ.to.J(x = 1)

unitconversion.deciwatt.to.watt(x = 1)

unitconversion.dW.to.W(x = 1)

unitconversion.decicoulomb.to.coulomb(x = 1)

unitconversion.dC.to.C(x = 1)

unitconversion.decivolt.to.volt(x = 1)

unitconversion.dV.to.V(x = 1)

unitconversion.decifarad.to.farad(x = 1)

unitconversion.dF.to.F(x = 1)

unitconversion.deciohm.to.ohm(x = 1)

unitconversion.decisiemens.to.siemens(x = 1)

unitconversion.dS.to.S(x = 1)

unitconversion.deciweber.to.weber(x = 1)

unitconversion.dWb.to.Wb(x = 1)

unitconversion.decitesla.to.tesla(x = 1)

unitconversion.dT.to.T(x = 1)

unitconversion.decihenry.to.henry(x = 1)

unitconversion.dH.to.H(x = 1)

unitconversion.decilumen.to.lumen(x = 1)

unitconversion.dlm.to.lm(x = 1)

unitconversion.decilux.to.lux(x = 1)

unitconversion.dlx.to.lx(x = 1)

unitconversion.decibecquerel.to.becquerel(x = 1)

unitconversion.dBq.to.Bq(x = 1)

unitconversion.decigray.to.gray(x = 1)

unitconversion.dGy.to.Gy(x = 1)

unitconversion.decisievert.to.sievert(x = 1)

unitconversion.dSv.to.Sv(x = 1)

unitconversion.decikatal.to.katal(x = 1)

unitconversion.dkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of deci-units}
}
\value{
x, but converted to base units
}
\description{
Deci- tenths, 1/10, or 10^(-1)
}
\details{
Performs a conversion from deci-units to base units (ex. decigrams to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
