% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.deka.to.base.R
\name{unitconversion.siprefix.deka.to.base}
\alias{unitconversion.siprefix.deka.to.base}
\alias{unitconversion.dekasecond.to.second}
\alias{unitconversion.das.to.s}
\alias{unitconversion.dekameter.to.meter}
\alias{unitconversion.dam.to.m}
\alias{unitconversion.dekagram.to.gram}
\alias{unitconversion.dag.to.g}
\alias{unitconversion.dekaampere.to.ampere}
\alias{unitconversion.daA.to.A}
\alias{unitconversion.dekakelvin.to.kelvin}
\alias{unitconversion.daK.to.K}
\alias{unitconversion.dekamole.to.mole}
\alias{unitconversion.damol.to.mol}
\alias{unitconversion.dekacandela.to.candela}
\alias{unitconversion.dacd.to.cd}
\alias{unitconversion.dekaradian.to.radian}
\alias{unitconversion.darad.to.rad}
\alias{unitconversion.dekasteradian.to.steradian}
\alias{unitconversion.dasr.to.sr}
\alias{unitconversion.dekahertz.to.hertz}
\alias{unitconversion.daHz.to.Hz}
\alias{unitconversion.dekanewton.to.newton}
\alias{unitconversion.daN.to.N}
\alias{unitconversion.dekapascal.to.pascal}
\alias{unitconversion.daPa.to.Pa}
\alias{unitconversion.dekajoule.to.joule}
\alias{unitconversion.daJ.to.J}
\alias{unitconversion.dekawatt.to.watt}
\alias{unitconversion.daW.to.W}
\alias{unitconversion.dekacoulomb.to.coulomb}
\alias{unitconversion.daC.to.C}
\alias{unitconversion.dekavolt.to.volt}
\alias{unitconversion.daV.to.V}
\alias{unitconversion.dekafarad.to.farad}
\alias{unitconversion.daF.to.F}
\alias{unitconversion.dekaohm.to.ohm}
\alias{unitconversion.dekasiemens.to.siemens}
\alias{unitconversion.daS.to.S}
\alias{unitconversion.dekaweber.to.weber}
\alias{unitconversion.daWb.to.Wb}
\alias{unitconversion.dekatesla.to.tesla}
\alias{unitconversion.daT.to.T}
\alias{unitconversion.dekahenry.to.henry}
\alias{unitconversion.daH.to.H}
\alias{unitconversion.dekalumen.to.lumen}
\alias{unitconversion.dalm.to.lm}
\alias{unitconversion.dekalux.to.lux}
\alias{unitconversion.dalx.to.lx}
\alias{unitconversion.dekabecquerel.to.becquerel}
\alias{unitconversion.daBq.to.Bq}
\alias{unitconversion.dekagray.to.gray}
\alias{unitconversion.daGy.to.Gy}
\alias{unitconversion.dekasievert.to.sievert}
\alias{unitconversion.daSv.to.Sv}
\alias{unitconversion.dekakatal.to.katal}
\alias{unitconversion.dakat.to.kat}
\title{Unit Conversion - SI Prefixes - Deka- to Base}
\usage{
unitconversion.siprefix.deka.to.base(x = 1)

unitconversion.dekasecond.to.second(x = 1)

unitconversion.das.to.s(x = 1)

unitconversion.dekameter.to.meter(x = 1)

unitconversion.dam.to.m(x = 1)

unitconversion.dekagram.to.gram(x = 1)

unitconversion.dag.to.g(x = 1)

unitconversion.dekaampere.to.ampere(x = 1)

unitconversion.daA.to.A(x = 1)

unitconversion.dekakelvin.to.kelvin(x = 1)

unitconversion.daK.to.K(x = 1)

unitconversion.dekamole.to.mole(x = 1)

unitconversion.damol.to.mol(x = 1)

unitconversion.dekacandela.to.candela(x = 1)

unitconversion.dacd.to.cd(x = 1)

unitconversion.dekaradian.to.radian(x = 1)

unitconversion.darad.to.rad(x = 1)

unitconversion.dekasteradian.to.steradian(x = 1)

unitconversion.dasr.to.sr(x = 1)

unitconversion.dekahertz.to.hertz(x = 1)

unitconversion.daHz.to.Hz(x = 1)

unitconversion.dekanewton.to.newton(x = 1)

unitconversion.daN.to.N(x = 1)

unitconversion.dekapascal.to.pascal(x = 1)

unitconversion.daPa.to.Pa(x = 1)

unitconversion.dekajoule.to.joule(x = 1)

unitconversion.daJ.to.J(x = 1)

unitconversion.dekawatt.to.watt(x = 1)

unitconversion.daW.to.W(x = 1)

unitconversion.dekacoulomb.to.coulomb(x = 1)

unitconversion.daC.to.C(x = 1)

unitconversion.dekavolt.to.volt(x = 1)

unitconversion.daV.to.V(x = 1)

unitconversion.dekafarad.to.farad(x = 1)

unitconversion.daF.to.F(x = 1)

unitconversion.dekaohm.to.ohm(x = 1)

unitconversion.dekasiemens.to.siemens(x = 1)

unitconversion.daS.to.S(x = 1)

unitconversion.dekaweber.to.weber(x = 1)

unitconversion.daWb.to.Wb(x = 1)

unitconversion.dekatesla.to.tesla(x = 1)

unitconversion.daT.to.T(x = 1)

unitconversion.dekahenry.to.henry(x = 1)

unitconversion.daH.to.H(x = 1)

unitconversion.dekalumen.to.lumen(x = 1)

unitconversion.dalm.to.lm(x = 1)

unitconversion.dekalux.to.lux(x = 1)

unitconversion.dalx.to.lx(x = 1)

unitconversion.dekabecquerel.to.becquerel(x = 1)

unitconversion.daBq.to.Bq(x = 1)

unitconversion.dekagray.to.gray(x = 1)

unitconversion.daGy.to.Gy(x = 1)

unitconversion.dekasievert.to.sievert(x = 1)

unitconversion.daSv.to.Sv(x = 1)

unitconversion.dekakatal.to.katal(x = 1)

unitconversion.dakat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of deka-units}
}
\value{
x, but converted to base units
}
\description{
Deka- tens, 10, or 10^1
}
\details{
Performs a conversion from deka-units to base units (ex. dekagrams to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
