% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.exa.to.base.R
\name{unitconversion.siprefix.exa.to.base}
\alias{unitconversion.siprefix.exa.to.base}
\alias{unitconversion.exasecond.to.second}
\alias{unitconversion.Es.to.s}
\alias{unitconversion.exameter.to.meter}
\alias{unitconversion.Em.to.m}
\alias{unitconversion.exagram.to.gram}
\alias{unitconversion.Eg.to.g}
\alias{unitconversion.exaampere.to.ampere}
\alias{unitconversion.EA.to.A}
\alias{unitconversion.exakelvin.to.kelvin}
\alias{unitconversion.EK.to.K}
\alias{unitconversion.examole.to.mole}
\alias{unitconversion.Emol.to.mol}
\alias{unitconversion.exacandela.to.candela}
\alias{unitconversion.Ecd.to.cd}
\alias{unitconversion.exaradian.to.radian}
\alias{unitconversion.Erad.to.rad}
\alias{unitconversion.exasteradian.to.steradian}
\alias{unitconversion.Esr.to.sr}
\alias{unitconversion.exahertz.to.hertz}
\alias{unitconversion.EHz.to.Hz}
\alias{unitconversion.exanewton.to.newton}
\alias{unitconversion.EN.to.N}
\alias{unitconversion.exapascal.to.pascal}
\alias{unitconversion.EPa.to.Pa}
\alias{unitconversion.exajoule.to.joule}
\alias{unitconversion.EJ.to.J}
\alias{unitconversion.exawatt.to.watt}
\alias{unitconversion.EW.to.W}
\alias{unitconversion.exacoulomb.to.coulomb}
\alias{unitconversion.EC.to.C}
\alias{unitconversion.exavolt.to.volt}
\alias{unitconversion.EV.to.V}
\alias{unitconversion.exafarad.to.farad}
\alias{unitconversion.EF.to.F}
\alias{unitconversion.exaohm.to.ohm}
\alias{unitconversion.exasiemens.to.siemens}
\alias{unitconversion.ES.to.S}
\alias{unitconversion.exaweber.to.weber}
\alias{unitconversion.EWb.to.Wb}
\alias{unitconversion.exatesla.to.tesla}
\alias{unitconversion.ET.to.T}
\alias{unitconversion.exahenry.to.henry}
\alias{unitconversion.EH.to.H}
\alias{unitconversion.exalumen.to.lumen}
\alias{unitconversion.Elm.to.lm}
\alias{unitconversion.exalux.to.lux}
\alias{unitconversion.Elx.to.lx}
\alias{unitconversion.exabecquerel.to.becquerel}
\alias{unitconversion.EBq.to.Bq}
\alias{unitconversion.exagray.to.gray}
\alias{unitconversion.EGy.to.Gy}
\alias{unitconversion.exasievert.to.sievert}
\alias{unitconversion.ESv.to.Sv}
\alias{unitconversion.exakatal.to.katal}
\alias{unitconversion.Ekat.to.kat}
\title{Unit Conversion - SI Prefixes - Exa- to Base}
\usage{
unitconversion.siprefix.exa.to.base(x = 1)

unitconversion.exasecond.to.second(x = 1)

unitconversion.Es.to.s(x = 1)

unitconversion.exameter.to.meter(x = 1)

unitconversion.Em.to.m(x = 1)

unitconversion.exagram.to.gram(x = 1)

unitconversion.Eg.to.g(x = 1)

unitconversion.exaampere.to.ampere(x = 1)

unitconversion.EA.to.A(x = 1)

unitconversion.exakelvin.to.kelvin(x = 1)

unitconversion.EK.to.K(x = 1)

unitconversion.examole.to.mole(x = 1)

unitconversion.Emol.to.mol(x = 1)

unitconversion.exacandela.to.candela(x = 1)

unitconversion.Ecd.to.cd(x = 1)

unitconversion.exaradian.to.radian(x = 1)

unitconversion.Erad.to.rad(x = 1)

unitconversion.exasteradian.to.steradian(x = 1)

unitconversion.Esr.to.sr(x = 1)

unitconversion.exahertz.to.hertz(x = 1)

unitconversion.EHz.to.Hz(x = 1)

unitconversion.exanewton.to.newton(x = 1)

unitconversion.EN.to.N(x = 1)

unitconversion.exapascal.to.pascal(x = 1)

unitconversion.EPa.to.Pa(x = 1)

unitconversion.exajoule.to.joule(x = 1)

unitconversion.EJ.to.J(x = 1)

unitconversion.exawatt.to.watt(x = 1)

unitconversion.EW.to.W(x = 1)

unitconversion.exacoulomb.to.coulomb(x = 1)

unitconversion.EC.to.C(x = 1)

unitconversion.exavolt.to.volt(x = 1)

unitconversion.EV.to.V(x = 1)

unitconversion.exafarad.to.farad(x = 1)

unitconversion.EF.to.F(x = 1)

unitconversion.exaohm.to.ohm(x = 1)

unitconversion.exasiemens.to.siemens(x = 1)

unitconversion.ES.to.S(x = 1)

unitconversion.exaweber.to.weber(x = 1)

unitconversion.EWb.to.Wb(x = 1)

unitconversion.exatesla.to.tesla(x = 1)

unitconversion.ET.to.T(x = 1)

unitconversion.exahenry.to.henry(x = 1)

unitconversion.EH.to.H(x = 1)

unitconversion.exalumen.to.lumen(x = 1)

unitconversion.Elm.to.lm(x = 1)

unitconversion.exalux.to.lux(x = 1)

unitconversion.Elx.to.lx(x = 1)

unitconversion.exabecquerel.to.becquerel(x = 1)

unitconversion.EBq.to.Bq(x = 1)

unitconversion.exagray.to.gray(x = 1)

unitconversion.EGy.to.Gy(x = 1)

unitconversion.exasievert.to.sievert(x = 1)

unitconversion.ESv.to.Sv(x = 1)

unitconversion.exakatal.to.katal(x = 1)

unitconversion.Ekat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of exa-units}
}
\value{
x, but converted to base units
}
\description{
Exa- quintillions, 1 000 000 000 000 000 000, or 10^18
}
\details{
Performs a conversion from exa-units to base units (ex. exagrams to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
