% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.femto.to.base.R
\name{unitconversion.siprefix.femto.to.base}
\alias{unitconversion.siprefix.femto.to.base}
\alias{unitconversion.femtosecond.to.second}
\alias{unitconversion.fs.to.s}
\alias{unitconversion.femtometer.to.meter}
\alias{unitconversion.fm.to.m}
\alias{unitconversion.femtogram.to.gram}
\alias{unitconversion.fg.to.g}
\alias{unitconversion.femtoampere.to.ampere}
\alias{unitconversion.fA.to.A}
\alias{unitconversion.femtokelvin.to.kelvin}
\alias{unitconversion.fK.to.K}
\alias{unitconversion.femtomole.to.mole}
\alias{unitconversion.fmol.to.mol}
\alias{unitconversion.femtocandela.to.candela}
\alias{unitconversion.fcd.to.cd}
\alias{unitconversion.femtoradian.to.radian}
\alias{unitconversion.frad.to.rad}
\alias{unitconversion.femtosteradian.to.steradian}
\alias{unitconversion.fsr.to.sr}
\alias{unitconversion.femtohertz.to.hertz}
\alias{unitconversion.fHz.to.Hz}
\alias{unitconversion.femtonewton.to.newton}
\alias{unitconversion.fN.to.N}
\alias{unitconversion.femtopascal.to.pascal}
\alias{unitconversion.fPa.to.Pa}
\alias{unitconversion.femtojoule.to.joule}
\alias{unitconversion.fJ.to.J}
\alias{unitconversion.femtowatt.to.watt}
\alias{unitconversion.fW.to.W}
\alias{unitconversion.femtocoulomb.to.coulomb}
\alias{unitconversion.fC.to.C}
\alias{unitconversion.femtovolt.to.volt}
\alias{unitconversion.fV.to.V}
\alias{unitconversion.femtofarad.to.farad}
\alias{unitconversion.fF.to.F}
\alias{unitconversion.femtoohm.to.ohm}
\alias{unitconversion.femtosiemens.to.siemens}
\alias{unitconversion.fS.to.S}
\alias{unitconversion.femtoweber.to.weber}
\alias{unitconversion.fWb.to.Wb}
\alias{unitconversion.femtotesla.to.tesla}
\alias{unitconversion.fT.to.T}
\alias{unitconversion.femtohenry.to.henry}
\alias{unitconversion.fH.to.H}
\alias{unitconversion.femtolumen.to.lumen}
\alias{unitconversion.flm.to.lm}
\alias{unitconversion.femtolux.to.lux}
\alias{unitconversion.flx.to.lx}
\alias{unitconversion.femtobecquerel.to.becquerel}
\alias{unitconversion.fBq.to.Bq}
\alias{unitconversion.femtogray.to.gray}
\alias{unitconversion.fGy.to.Gy}
\alias{unitconversion.femtosievert.to.sievert}
\alias{unitconversion.fSv.to.Sv}
\alias{unitconversion.femtokatal.to.katal}
\alias{unitconversion.fkat.to.kat}
\title{Unit Conversion - SI Prefixes - Femto- to Base}
\usage{
unitconversion.siprefix.femto.to.base(x = 1)

unitconversion.femtosecond.to.second(x = 1)

unitconversion.fs.to.s(x = 1)

unitconversion.femtometer.to.meter(x = 1)

unitconversion.fm.to.m(x = 1)

unitconversion.femtogram.to.gram(x = 1)

unitconversion.fg.to.g(x = 1)

unitconversion.femtoampere.to.ampere(x = 1)

unitconversion.fA.to.A(x = 1)

unitconversion.femtokelvin.to.kelvin(x = 1)

unitconversion.fK.to.K(x = 1)

unitconversion.femtomole.to.mole(x = 1)

unitconversion.fmol.to.mol(x = 1)

unitconversion.femtocandela.to.candela(x = 1)

unitconversion.fcd.to.cd(x = 1)

unitconversion.femtoradian.to.radian(x = 1)

unitconversion.frad.to.rad(x = 1)

unitconversion.femtosteradian.to.steradian(x = 1)

unitconversion.fsr.to.sr(x = 1)

unitconversion.femtohertz.to.hertz(x = 1)

unitconversion.fHz.to.Hz(x = 1)

unitconversion.femtonewton.to.newton(x = 1)

unitconversion.fN.to.N(x = 1)

unitconversion.femtopascal.to.pascal(x = 1)

unitconversion.fPa.to.Pa(x = 1)

unitconversion.femtojoule.to.joule(x = 1)

unitconversion.fJ.to.J(x = 1)

unitconversion.femtowatt.to.watt(x = 1)

unitconversion.fW.to.W(x = 1)

unitconversion.femtocoulomb.to.coulomb(x = 1)

unitconversion.fC.to.C(x = 1)

unitconversion.femtovolt.to.volt(x = 1)

unitconversion.fV.to.V(x = 1)

unitconversion.femtofarad.to.farad(x = 1)

unitconversion.fF.to.F(x = 1)

unitconversion.femtoohm.to.ohm(x = 1)

unitconversion.femtosiemens.to.siemens(x = 1)

unitconversion.fS.to.S(x = 1)

unitconversion.femtoweber.to.weber(x = 1)

unitconversion.fWb.to.Wb(x = 1)

unitconversion.femtotesla.to.tesla(x = 1)

unitconversion.fT.to.T(x = 1)

unitconversion.femtohenry.to.henry(x = 1)

unitconversion.fH.to.H(x = 1)

unitconversion.femtolumen.to.lumen(x = 1)

unitconversion.flm.to.lm(x = 1)

unitconversion.femtolux.to.lux(x = 1)

unitconversion.flx.to.lx(x = 1)

unitconversion.femtobecquerel.to.becquerel(x = 1)

unitconversion.fBq.to.Bq(x = 1)

unitconversion.femtogray.to.gray(x = 1)

unitconversion.fGy.to.Gy(x = 1)

unitconversion.femtosievert.to.sievert(x = 1)

unitconversion.fSv.to.Sv(x = 1)

unitconversion.femtokatal.to.katal(x = 1)

unitconversion.fkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of femto-units}
}
\value{
x, but converted to base units
}
\description{
Femto- quadrillionths, 1/1 000 000 000 000 000, or 10^(-15)
}
\details{
Performs a conversion from femto-units to base (ex. femtograms to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
