% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.giga.to.base.R
\name{unitconversion.siprefix.giga.to.base}
\alias{unitconversion.siprefix.giga.to.base}
\alias{unitconversion.gigasecond.to.second}
\alias{unitconversion.Gs.to.s}
\alias{unitconversion.gigameter.to.meter}
\alias{unitconversion.Gm.to.m}
\alias{unitconversion.gigagram.to.gram}
\alias{unitconversion.Gg.to.g}
\alias{unitconversion.gigaampere.to.ampere}
\alias{unitconversion.GA.to.A}
\alias{unitconversion.gigakelvin.to.kelvin}
\alias{unitconversion.GK.to.K}
\alias{unitconversion.gigamole.to.mole}
\alias{unitconversion.Gmol.to.mol}
\alias{unitconversion.gigacandela.to.candela}
\alias{unitconversion.Gcd.to.cd}
\alias{unitconversion.gigaradian.to.radian}
\alias{unitconversion.Grad.to.rad}
\alias{unitconversion.gigasteradian.to.steradian}
\alias{unitconversion.Gsr.to.sr}
\alias{unitconversion.gigahertz.to.hertz}
\alias{unitconversion.GHz.to.Hz}
\alias{unitconversion.giganewton.to.newton}
\alias{unitconversion.GN.to.N}
\alias{unitconversion.gigapascal.to.pascal}
\alias{unitconversion.GPa.to.Pa}
\alias{unitconversion.gigajoule.to.joule}
\alias{unitconversion.GJ.to.J}
\alias{unitconversion.gigawatt.to.watt}
\alias{unitconversion.GW.to.W}
\alias{unitconversion.gigacoulomb.to.coulomb}
\alias{unitconversion.GC.to.C}
\alias{unitconversion.gigavolt.to.volt}
\alias{unitconversion.GV.to.V}
\alias{unitconversion.gigafarad.to.farad}
\alias{unitconversion.GF.to.F}
\alias{unitconversion.gigaohm.to.ohm}
\alias{unitconversion.gigasiemens.to.siemens}
\alias{unitconversion.GS.to.S}
\alias{unitconversion.gigaweber.to.weber}
\alias{unitconversion.GWb.to.Wb}
\alias{unitconversion.gigatesla.to.tesla}
\alias{unitconversion.GT.to.T}
\alias{unitconversion.gigahenry.to.henry}
\alias{unitconversion.GH.to.H}
\alias{unitconversion.gigalumen.to.lumen}
\alias{unitconversion.Glm.to.lm}
\alias{unitconversion.gigalux.to.lux}
\alias{unitconversion.Glx.to.lx}
\alias{unitconversion.gigabecquerel.to.becquerel}
\alias{unitconversion.GBq.to.Bq}
\alias{unitconversion.gigagray.to.gray}
\alias{unitconversion.GGy.to.Gy}
\alias{unitconversion.gigasievert.to.sievert}
\alias{unitconversion.GSv.to.Sv}
\alias{unitconversion.gigakatal.to.katal}
\alias{unitconversion.Gkat.to.kat}
\title{Unit Conversion - SI Prefixes - Giga- to Base}
\usage{
unitconversion.siprefix.giga.to.base(x = 1)

unitconversion.gigasecond.to.second(x = 1)

unitconversion.Gs.to.s(x = 1)

unitconversion.gigameter.to.meter(x = 1)

unitconversion.Gm.to.m(x = 1)

unitconversion.gigagram.to.gram(x = 1)

unitconversion.Gg.to.g(x = 1)

unitconversion.gigaampere.to.ampere(x = 1)

unitconversion.GA.to.A(x = 1)

unitconversion.gigakelvin.to.kelvin(x = 1)

unitconversion.GK.to.K(x = 1)

unitconversion.gigamole.to.mole(x = 1)

unitconversion.Gmol.to.mol(x = 1)

unitconversion.gigacandela.to.candela(x = 1)

unitconversion.Gcd.to.cd(x = 1)

unitconversion.gigaradian.to.radian(x = 1)

unitconversion.Grad.to.rad(x = 1)

unitconversion.gigasteradian.to.steradian(x = 1)

unitconversion.Gsr.to.sr(x = 1)

unitconversion.gigahertz.to.hertz(x = 1)

unitconversion.GHz.to.Hz(x = 1)

unitconversion.giganewton.to.newton(x = 1)

unitconversion.GN.to.N(x = 1)

unitconversion.gigapascal.to.pascal(x = 1)

unitconversion.GPa.to.Pa(x = 1)

unitconversion.gigajoule.to.joule(x = 1)

unitconversion.GJ.to.J(x = 1)

unitconversion.gigawatt.to.watt(x = 1)

unitconversion.GW.to.W(x = 1)

unitconversion.gigacoulomb.to.coulomb(x = 1)

unitconversion.GC.to.C(x = 1)

unitconversion.gigavolt.to.volt(x = 1)

unitconversion.GV.to.V(x = 1)

unitconversion.gigafarad.to.farad(x = 1)

unitconversion.GF.to.F(x = 1)

unitconversion.gigaohm.to.ohm(x = 1)

unitconversion.gigasiemens.to.siemens(x = 1)

unitconversion.GS.to.S(x = 1)

unitconversion.gigaweber.to.weber(x = 1)

unitconversion.GWb.to.Wb(x = 1)

unitconversion.gigatesla.to.tesla(x = 1)

unitconversion.GT.to.T(x = 1)

unitconversion.gigahenry.to.henry(x = 1)

unitconversion.GH.to.H(x = 1)

unitconversion.gigalumen.to.lumen(x = 1)

unitconversion.Glm.to.lm(x = 1)

unitconversion.gigalux.to.lux(x = 1)

unitconversion.Glx.to.lx(x = 1)

unitconversion.gigabecquerel.to.becquerel(x = 1)

unitconversion.GBq.to.Bq(x = 1)

unitconversion.gigagray.to.gray(x = 1)

unitconversion.GGy.to.Gy(x = 1)

unitconversion.gigasievert.to.sievert(x = 1)

unitconversion.GSv.to.Sv(x = 1)

unitconversion.gigakatal.to.katal(x = 1)

unitconversion.Gkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of giga-units}
}
\value{
x, but converted to base units
}
\description{
Giga- billions, 1 000 000 000, or 10^9
}
\details{
Performs a conversion from giga-units to base units (ex. gigagrams to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
