% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.hecto.to.base.R
\name{unitconversion.siprefix.hecto.to.base}
\alias{unitconversion.siprefix.hecto.to.base}
\alias{unitconversion.hectosecond.to.second}
\alias{unitconversion.hs.to.s}
\alias{unitconversion.hectometer.to.meter}
\alias{unitconversion.hm.to.m}
\alias{unitconversion.hectogram.to.gram}
\alias{unitconversion.hg.to.g}
\alias{unitconversion.hectoampere.to.ampere}
\alias{unitconversion.hA.to.A}
\alias{unitconversion.hectokelvin.to.kelvin}
\alias{unitconversion.hK.to.K}
\alias{unitconversion.hectomole.to.mole}
\alias{unitconversion.hmol.to.mol}
\alias{unitconversion.hectocandela.to.candela}
\alias{unitconversion.hcd.to.cd}
\alias{unitconversion.hectoradian.to.radian}
\alias{unitconversion.hrad.to.rad}
\alias{unitconversion.hectosteradian.to.steradian}
\alias{unitconversion.hsr.to.sr}
\alias{unitconversion.hectohertz.to.hertz}
\alias{unitconversion.hHz.to.Hz}
\alias{unitconversion.hectonewton.to.newton}
\alias{unitconversion.hN.to.N}
\alias{unitconversion.hectopascal.to.pascal}
\alias{unitconversion.hPa.to.Pa}
\alias{unitconversion.hectojoule.to.joule}
\alias{unitconversion.hJ.to.J}
\alias{unitconversion.hectowatt.to.watt}
\alias{unitconversion.hW.to.W}
\alias{unitconversion.hectocoulomb.to.coulomb}
\alias{unitconversion.hC.to.C}
\alias{unitconversion.hectovolt.to.volt}
\alias{unitconversion.hV.to.V}
\alias{unitconversion.hectofarad.to.farad}
\alias{unitconversion.hF.to.F}
\alias{unitconversion.hectoohm.to.ohm}
\alias{unitconversion.hectosiemens.to.siemens}
\alias{unitconversion.hS.to.S}
\alias{unitconversion.hectoweber.to.weber}
\alias{unitconversion.hWb.to.Wb}
\alias{unitconversion.hectotesla.to.tesla}
\alias{unitconversion.hT.to.T}
\alias{unitconversion.hectohenry.to.henry}
\alias{unitconversion.hH.to.H}
\alias{unitconversion.hectolumen.to.lumen}
\alias{unitconversion.hlm.to.lm}
\alias{unitconversion.hectolux.to.lux}
\alias{unitconversion.hlx.to.lx}
\alias{unitconversion.hectobecquerel.to.becquerel}
\alias{unitconversion.hBq.to.Bq}
\alias{unitconversion.hectogray.to.gray}
\alias{unitconversion.hGy.to.Gy}
\alias{unitconversion.hectosievert.to.sievert}
\alias{unitconversion.hSv.to.Sv}
\alias{unitconversion.hectokatal.to.katal}
\alias{unitconversion.hkat.to.kat}
\title{Unit Conversion - SI Prefixes - Hecto- to Base}
\usage{
unitconversion.siprefix.hecto.to.base(x = 1)

unitconversion.hectosecond.to.second(x = 1)

unitconversion.hs.to.s(x = 1)

unitconversion.hectometer.to.meter(x = 1)

unitconversion.hm.to.m(x = 1)

unitconversion.hectogram.to.gram(x = 1)

unitconversion.hg.to.g(x = 1)

unitconversion.hectoampere.to.ampere(x = 1)

unitconversion.hA.to.A(x = 1)

unitconversion.hectokelvin.to.kelvin(x = 1)

unitconversion.hK.to.K(x = 1)

unitconversion.hectomole.to.mole(x = 1)

unitconversion.hmol.to.mol(x = 1)

unitconversion.hectocandela.to.candela(x = 1)

unitconversion.hcd.to.cd(x = 1)

unitconversion.hectoradian.to.radian(x = 1)

unitconversion.hrad.to.rad(x = 1)

unitconversion.hectosteradian.to.steradian(x = 1)

unitconversion.hsr.to.sr(x = 1)

unitconversion.hectohertz.to.hertz(x = 1)

unitconversion.hHz.to.Hz(x = 1)

unitconversion.hectonewton.to.newton(x = 1)

unitconversion.hN.to.N(x = 1)

unitconversion.hectopascal.to.pascal(x = 1)

unitconversion.hPa.to.Pa(x = 1)

unitconversion.hectojoule.to.joule(x = 1)

unitconversion.hJ.to.J(x = 1)

unitconversion.hectowatt.to.watt(x = 1)

unitconversion.hW.to.W(x = 1)

unitconversion.hectocoulomb.to.coulomb(x = 1)

unitconversion.hC.to.C(x = 1)

unitconversion.hectovolt.to.volt(x = 1)

unitconversion.hV.to.V(x = 1)

unitconversion.hectofarad.to.farad(x = 1)

unitconversion.hF.to.F(x = 1)

unitconversion.hectoohm.to.ohm(x = 1)

unitconversion.hectosiemens.to.siemens(x = 1)

unitconversion.hS.to.S(x = 1)

unitconversion.hectoweber.to.weber(x = 1)

unitconversion.hWb.to.Wb(x = 1)

unitconversion.hectotesla.to.tesla(x = 1)

unitconversion.hT.to.T(x = 1)

unitconversion.hectohenry.to.henry(x = 1)

unitconversion.hH.to.H(x = 1)

unitconversion.hectolumen.to.lumen(x = 1)

unitconversion.hlm.to.lm(x = 1)

unitconversion.hectolux.to.lux(x = 1)

unitconversion.hlx.to.lx(x = 1)

unitconversion.hectobecquerel.to.becquerel(x = 1)

unitconversion.hBq.to.Bq(x = 1)

unitconversion.hectogray.to.gray(x = 1)

unitconversion.hGy.to.Gy(x = 1)

unitconversion.hectosievert.to.sievert(x = 1)

unitconversion.hSv.to.Sv(x = 1)

unitconversion.hectokatal.to.katal(x = 1)

unitconversion.hkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of hecto-units}
}
\value{
x, but converted to base units
}
\description{
Hecto- hundreds, 100, or 10^2
}
\details{
Performs a conversion from hecto-units to base (ex. hectograms to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
