% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.kilo.to.base.R
\name{unitconversion.siprefix.kilo.to.base}
\alias{unitconversion.siprefix.kilo.to.base}
\alias{unitconversion.kilosecond.to.second}
\alias{unitconversion.ks.to.s}
\alias{unitconversion.kilometer.to.meter}
\alias{unitconversion.km.to.m}
\alias{unitconversion.kilogram.to.gram}
\alias{unitconversion.kg.to.g}
\alias{unitconversion.kiloampere.to.ampere}
\alias{unitconversion.kA.to.A}
\alias{unitconversion.kilokelvin.to.kelvin}
\alias{unitconversion.kK.to.K}
\alias{unitconversion.kilomole.to.mole}
\alias{unitconversion.kmol.to.mol}
\alias{unitconversion.kilocandela.to.candela}
\alias{unitconversion.kcd.to.cd}
\alias{unitconversion.kiloradian.to.radian}
\alias{unitconversion.krad.to.rad}
\alias{unitconversion.kilosteradian.to.steradian}
\alias{unitconversion.ksr.to.sr}
\alias{unitconversion.kilohertz.to.hertz}
\alias{unitconversion.kHz.to.Hz}
\alias{unitconversion.kilonewton.to.newton}
\alias{unitconversion.kN.to.N}
\alias{unitconversion.kilopascal.to.pascal}
\alias{unitconversion.kPa.to.Pa}
\alias{unitconversion.kilojoule.to.joule}
\alias{unitconversion.kJ.to.J}
\alias{unitconversion.kilowatt.to.watt}
\alias{unitconversion.kW.to.W}
\alias{unitconversion.kilocoulomb.to.coulomb}
\alias{unitconversion.kC.to.C}
\alias{unitconversion.kilovolt.to.volt}
\alias{unitconversion.kV.to.V}
\alias{unitconversion.kilofarad.to.farad}
\alias{unitconversion.kF.to.F}
\alias{unitconversion.kiloohm.to.ohm}
\alias{unitconversion.kilosiemens.to.siemens}
\alias{unitconversion.kS.to.S}
\alias{unitconversion.kiloweber.to.weber}
\alias{unitconversion.kWb.to.Wb}
\alias{unitconversion.kilotesla.to.tesla}
\alias{unitconversion.kT.to.T}
\alias{unitconversion.kilohenry.to.henry}
\alias{unitconversion.kH.to.H}
\alias{unitconversion.kilolumen.to.lumen}
\alias{unitconversion.klm.to.lm}
\alias{unitconversion.kilolux.to.lux}
\alias{unitconversion.klx.to.lx}
\alias{unitconversion.kilobecquerel.to.becquerel}
\alias{unitconversion.kBq.to.Bq}
\alias{unitconversion.kilogray.to.gray}
\alias{unitconversion.kGy.to.Gy}
\alias{unitconversion.kilosievert.to.sievert}
\alias{unitconversion.kSv.to.Sv}
\alias{unitconversion.kilokatal.to.katal}
\alias{unitconversion.kkat.to.kat}
\title{Unit Conversion - SI Prefixes - Kilo- to Base}
\usage{
unitconversion.siprefix.kilo.to.base(x = 1)

unitconversion.kilosecond.to.second(x = 1)

unitconversion.ks.to.s(x = 1)

unitconversion.kilometer.to.meter(x = 1)

unitconversion.km.to.m(x = 1)

unitconversion.kilogram.to.gram(x = 1)

unitconversion.kg.to.g(x = 1)

unitconversion.kiloampere.to.ampere(x = 1)

unitconversion.kA.to.A(x = 1)

unitconversion.kilokelvin.to.kelvin(x = 1)

unitconversion.kK.to.K(x = 1)

unitconversion.kilomole.to.mole(x = 1)

unitconversion.kmol.to.mol(x = 1)

unitconversion.kilocandela.to.candela(x = 1)

unitconversion.kcd.to.cd(x = 1)

unitconversion.kiloradian.to.radian(x = 1)

unitconversion.krad.to.rad(x = 1)

unitconversion.kilosteradian.to.steradian(x = 1)

unitconversion.ksr.to.sr(x = 1)

unitconversion.kilohertz.to.hertz(x = 1)

unitconversion.kHz.to.Hz(x = 1)

unitconversion.kilonewton.to.newton(x = 1)

unitconversion.kN.to.N(x = 1)

unitconversion.kilopascal.to.pascal(x = 1)

unitconversion.kPa.to.Pa(x = 1)

unitconversion.kilojoule.to.joule(x = 1)

unitconversion.kJ.to.J(x = 1)

unitconversion.kilowatt.to.watt(x = 1)

unitconversion.kW.to.W(x = 1)

unitconversion.kilocoulomb.to.coulomb(x = 1)

unitconversion.kC.to.C(x = 1)

unitconversion.kilovolt.to.volt(x = 1)

unitconversion.kV.to.V(x = 1)

unitconversion.kilofarad.to.farad(x = 1)

unitconversion.kF.to.F(x = 1)

unitconversion.kiloohm.to.ohm(x = 1)

unitconversion.kilosiemens.to.siemens(x = 1)

unitconversion.kS.to.S(x = 1)

unitconversion.kiloweber.to.weber(x = 1)

unitconversion.kWb.to.Wb(x = 1)

unitconversion.kilotesla.to.tesla(x = 1)

unitconversion.kT.to.T(x = 1)

unitconversion.kilohenry.to.henry(x = 1)

unitconversion.kH.to.H(x = 1)

unitconversion.kilolumen.to.lumen(x = 1)

unitconversion.klm.to.lm(x = 1)

unitconversion.kilolux.to.lux(x = 1)

unitconversion.klx.to.lx(x = 1)

unitconversion.kilobecquerel.to.becquerel(x = 1)

unitconversion.kBq.to.Bq(x = 1)

unitconversion.kilogray.to.gray(x = 1)

unitconversion.kGy.to.Gy(x = 1)

unitconversion.kilosievert.to.sievert(x = 1)

unitconversion.kSv.to.Sv(x = 1)

unitconversion.kilokatal.to.katal(x = 1)

unitconversion.kkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of kilo-units}
}
\value{
x, but converted to base units
}
\description{
Kilo- thousands, 1000, or 10^3
}
\details{
Performs a conversion from kilo-units to base units (ex. kilograms to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
