% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.mega.to.base.R
\name{unitconversion.siprefix.mega.to.base}
\alias{unitconversion.siprefix.mega.to.base}
\alias{unitconversion.megasecond.to.second}
\alias{unitconversion.Ms.to.s}
\alias{unitconversion.megameter.to.meter}
\alias{unitconversion.Mm.to.m}
\alias{unitconversion.megagram.to.gram}
\alias{unitconversion.Mg.to.g}
\alias{unitconversion.megaampere.to.ampere}
\alias{unitconversion.MA.to.A}
\alias{unitconversion.megakelvin.to.kelvin}
\alias{unitconversion.MK.to.K}
\alias{unitconversion.megamole.to.mole}
\alias{unitconversion.Mmol.to.mol}
\alias{unitconversion.megacandela.to.candela}
\alias{unitconversion.Mcd.to.cd}
\alias{unitconversion.megaradian.to.radian}
\alias{unitconversion.Mrad.to.rad}
\alias{unitconversion.megasteradian.to.steradian}
\alias{unitconversion.Msr.to.sr}
\alias{unitconversion.megahertz.to.hertz}
\alias{unitconversion.MHz.to.Hz}
\alias{unitconversion.meganewton.to.newton}
\alias{unitconversion.MN.to.N}
\alias{unitconversion.megapascal.to.pascal}
\alias{unitconversion.MPa.to.Pa}
\alias{unitconversion.megajoule.to.joule}
\alias{unitconversion.MJ.to.J}
\alias{unitconversion.megawatt.to.watt}
\alias{unitconversion.MW.to.W}
\alias{unitconversion.megacoulomb.to.coulomb}
\alias{unitconversion.MC.to.C}
\alias{unitconversion.megavolt.to.volt}
\alias{unitconversion.MV.to.V}
\alias{unitconversion.megafarad.to.farad}
\alias{unitconversion.MF.to.F}
\alias{unitconversion.megaohm.to.ohm}
\alias{unitconversion.megasiemens.to.siemens}
\alias{unitconversion.MS.to.S}
\alias{unitconversion.megaweber.to.weber}
\alias{unitconversion.MWb.to.Wb}
\alias{unitconversion.megatesla.to.tesla}
\alias{unitconversion.MT.to.T}
\alias{unitconversion.megahenry.to.henry}
\alias{unitconversion.MH.to.H}
\alias{unitconversion.megalumen.to.lumen}
\alias{unitconversion.Mlm.to.lm}
\alias{unitconversion.megalux.to.lux}
\alias{unitconversion.Mlx.to.lx}
\alias{unitconversion.megabecquerel.to.becquerel}
\alias{unitconversion.MBq.to.Bq}
\alias{unitconversion.megagray.to.gray}
\alias{unitconversion.MGy.to.Gy}
\alias{unitconversion.megasievert.to.sievert}
\alias{unitconversion.MSv.to.Sv}
\alias{unitconversion.megakatal.to.katal}
\alias{unitconversion.Mkat.to.kat}
\title{Unit Conversion - SI Prefixes - Mega- to Base}
\usage{
unitconversion.siprefix.mega.to.base(x = 1)

unitconversion.megasecond.to.second(x = 1)

unitconversion.Ms.to.s(x = 1)

unitconversion.megameter.to.meter(x = 1)

unitconversion.Mm.to.m(x = 1)

unitconversion.megagram.to.gram(x = 1)

unitconversion.Mg.to.g(x = 1)

unitconversion.megaampere.to.ampere(x = 1)

unitconversion.MA.to.A(x = 1)

unitconversion.megakelvin.to.kelvin(x = 1)

unitconversion.MK.to.K(x = 1)

unitconversion.megamole.to.mole(x = 1)

unitconversion.Mmol.to.mol(x = 1)

unitconversion.megacandela.to.candela(x = 1)

unitconversion.Mcd.to.cd(x = 1)

unitconversion.megaradian.to.radian(x = 1)

unitconversion.Mrad.to.rad(x = 1)

unitconversion.megasteradian.to.steradian(x = 1)

unitconversion.Msr.to.sr(x = 1)

unitconversion.megahertz.to.hertz(x = 1)

unitconversion.MHz.to.Hz(x = 1)

unitconversion.meganewton.to.newton(x = 1)

unitconversion.MN.to.N(x = 1)

unitconversion.megapascal.to.pascal(x = 1)

unitconversion.MPa.to.Pa(x = 1)

unitconversion.megajoule.to.joule(x = 1)

unitconversion.MJ.to.J(x = 1)

unitconversion.megawatt.to.watt(x = 1)

unitconversion.MW.to.W(x = 1)

unitconversion.megacoulomb.to.coulomb(x = 1)

unitconversion.MC.to.C(x = 1)

unitconversion.megavolt.to.volt(x = 1)

unitconversion.MV.to.V(x = 1)

unitconversion.megafarad.to.farad(x = 1)

unitconversion.MF.to.F(x = 1)

unitconversion.megaohm.to.ohm(x = 1)

unitconversion.megasiemens.to.siemens(x = 1)

unitconversion.MS.to.S(x = 1)

unitconversion.megaweber.to.weber(x = 1)

unitconversion.MWb.to.Wb(x = 1)

unitconversion.megatesla.to.tesla(x = 1)

unitconversion.MT.to.T(x = 1)

unitconversion.megahenry.to.henry(x = 1)

unitconversion.MH.to.H(x = 1)

unitconversion.megalumen.to.lumen(x = 1)

unitconversion.Mlm.to.lm(x = 1)

unitconversion.megalux.to.lux(x = 1)

unitconversion.Mlx.to.lx(x = 1)

unitconversion.megabecquerel.to.becquerel(x = 1)

unitconversion.MBq.to.Bq(x = 1)

unitconversion.megagray.to.gray(x = 1)

unitconversion.MGy.to.Gy(x = 1)

unitconversion.megasievert.to.sievert(x = 1)

unitconversion.MSv.to.Sv(x = 1)

unitconversion.megakatal.to.katal(x = 1)

unitconversion.Mkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of mega-units}
}
\value{
x, but converted to base units
}
\description{
Mega- millions, 1 000 000, or 10^6
}
\details{
Performs a conversion from mega-units to base units (ex. megagrams to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
