% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.micro.to.base.R
\name{unitconversion.siprefix.micro.to.base}
\alias{unitconversion.siprefix.micro.to.base}
\alias{unitconversion.microsecond.to.second}
\alias{unitconversion.us.to.s}
\alias{unitconversion.micrometer.to.meter}
\alias{unitconversion.um.to.m}
\alias{unitconversion.microgram.to.gram}
\alias{unitconversion.ug.to.g}
\alias{unitconversion.microampere.to.ampere}
\alias{unitconversion.uA.to.A}
\alias{unitconversion.microkelvin.to.kelvin}
\alias{unitconversion.uK.to.K}
\alias{unitconversion.micromole.to.mole}
\alias{unitconversion.umol.to.mol}
\alias{unitconversion.microcandela.to.candela}
\alias{unitconversion.ucd.to.cd}
\alias{unitconversion.microradian.to.radian}
\alias{unitconversion.urad.to.rad}
\alias{unitconversion.microsteradian.to.steradian}
\alias{unitconversion.usr.to.sr}
\alias{unitconversion.microhertz.to.hertz}
\alias{unitconversion.uHz.to.Hz}
\alias{unitconversion.micronewton.to.newton}
\alias{unitconversion.uN.to.N}
\alias{unitconversion.micropascal.to.pascal}
\alias{unitconversion.uPa.to.Pa}
\alias{unitconversion.microjoule.to.joule}
\alias{unitconversion.uJ.to.J}
\alias{unitconversion.microwatt.to.watt}
\alias{unitconversion.uW.to.W}
\alias{unitconversion.microcoulomb.to.coulomb}
\alias{unitconversion.uC.to.C}
\alias{unitconversion.microvolt.to.volt}
\alias{unitconversion.uV.to.V}
\alias{unitconversion.microfarad.to.farad}
\alias{unitconversion.uF.to.F}
\alias{unitconversion.microohm.to.ohm}
\alias{unitconversion.microsiemens.to.siemens}
\alias{unitconversion.uS.to.S}
\alias{unitconversion.microweber.to.weber}
\alias{unitconversion.uWb.to.Wb}
\alias{unitconversion.microtesla.to.tesla}
\alias{unitconversion.uT.to.T}
\alias{unitconversion.microhenry.to.henry}
\alias{unitconversion.uH.to.H}
\alias{unitconversion.microlumen.to.lumen}
\alias{unitconversion.ulm.to.lm}
\alias{unitconversion.microlux.to.lux}
\alias{unitconversion.ulx.to.lx}
\alias{unitconversion.microbecquerel.to.becquerel}
\alias{unitconversion.uBq.to.Bq}
\alias{unitconversion.microgray.to.gray}
\alias{unitconversion.uGy.to.Gy}
\alias{unitconversion.microsievert.to.sievert}
\alias{unitconversion.uSv.to.Sv}
\alias{unitconversion.microkatal.to.katal}
\alias{unitconversion.ukat.to.kat}
\title{Unit Conversion - SI Prefixes - Micro- to Base}
\usage{
unitconversion.siprefix.micro.to.base(x = 1)

unitconversion.microsecond.to.second(x = 1)

unitconversion.us.to.s(x = 1)

unitconversion.micrometer.to.meter(x = 1)

unitconversion.um.to.m(x = 1)

unitconversion.microgram.to.gram(x = 1)

unitconversion.ug.to.g(x = 1)

unitconversion.microampere.to.ampere(x = 1)

unitconversion.uA.to.A(x = 1)

unitconversion.microkelvin.to.kelvin(x = 1)

unitconversion.uK.to.K(x = 1)

unitconversion.micromole.to.mole(x = 1)

unitconversion.umol.to.mol(x = 1)

unitconversion.microcandela.to.candela(x = 1)

unitconversion.ucd.to.cd(x = 1)

unitconversion.microradian.to.radian(x = 1)

unitconversion.urad.to.rad(x = 1)

unitconversion.microsteradian.to.steradian(x = 1)

unitconversion.usr.to.sr(x = 1)

unitconversion.microhertz.to.hertz(x = 1)

unitconversion.uHz.to.Hz(x = 1)

unitconversion.micronewton.to.newton(x = 1)

unitconversion.uN.to.N(x = 1)

unitconversion.micropascal.to.pascal(x = 1)

unitconversion.uPa.to.Pa(x = 1)

unitconversion.microjoule.to.joule(x = 1)

unitconversion.uJ.to.J(x = 1)

unitconversion.microwatt.to.watt(x = 1)

unitconversion.uW.to.W(x = 1)

unitconversion.microcoulomb.to.coulomb(x = 1)

unitconversion.uC.to.C(x = 1)

unitconversion.microvolt.to.volt(x = 1)

unitconversion.uV.to.V(x = 1)

unitconversion.microfarad.to.farad(x = 1)

unitconversion.uF.to.F(x = 1)

unitconversion.microohm.to.ohm(x = 1)

unitconversion.microsiemens.to.siemens(x = 1)

unitconversion.uS.to.S(x = 1)

unitconversion.microweber.to.weber(x = 1)

unitconversion.uWb.to.Wb(x = 1)

unitconversion.microtesla.to.tesla(x = 1)

unitconversion.uT.to.T(x = 1)

unitconversion.microhenry.to.henry(x = 1)

unitconversion.uH.to.H(x = 1)

unitconversion.microlumen.to.lumen(x = 1)

unitconversion.ulm.to.lm(x = 1)

unitconversion.microlux.to.lux(x = 1)

unitconversion.ulx.to.lx(x = 1)

unitconversion.microbecquerel.to.becquerel(x = 1)

unitconversion.uBq.to.Bq(x = 1)

unitconversion.microgray.to.gray(x = 1)

unitconversion.uGy.to.Gy(x = 1)

unitconversion.microsievert.to.sievert(x = 1)

unitconversion.uSv.to.Sv(x = 1)

unitconversion.microkatal.to.katal(x = 1)

unitconversion.ukat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of micro-units}
}
\value{
x, but converted to base units
}
\description{
Micro- millionths, 1/1 000 000, or 10^(-6)
}
\details{
Performs a conversion from micro-units to base units (ex. micrograms to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
