% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.milli.to.base.R
\name{unitconversion.siprefix.milli.to.base}
\alias{unitconversion.siprefix.milli.to.base}
\alias{unitconversion.millisecond.to.second}
\alias{unitconversion.ms.to.s}
\alias{unitconversion.millimeter.to.meter}
\alias{unitconversion.mm.to.m}
\alias{unitconversion.milligram.to.gram}
\alias{unitconversion.mg.to.g}
\alias{unitconversion.milliampere.to.ampere}
\alias{unitconversion.mA.to.A}
\alias{unitconversion.millikelvin.to.kelvin}
\alias{unitconversion.mK.to.K}
\alias{unitconversion.millimole.to.mole}
\alias{unitconversion.mmol.to.mol}
\alias{unitconversion.millicandela.to.candela}
\alias{unitconversion.mcd.to.cd}
\alias{unitconversion.milliradian.to.radian}
\alias{unitconversion.mrad.to.rad}
\alias{unitconversion.millisteradian.to.steradian}
\alias{unitconversion.msr.to.sr}
\alias{unitconversion.millihertz.to.hertz}
\alias{unitconversion.mHz.to.Hz}
\alias{unitconversion.millinewton.to.newton}
\alias{unitconversion.mN.to.N}
\alias{unitconversion.millipascal.to.pascal}
\alias{unitconversion.mPa.to.Pa}
\alias{unitconversion.millijoule.to.joule}
\alias{unitconversion.mJ.to.J}
\alias{unitconversion.milliwatt.to.watt}
\alias{unitconversion.mW.to.W}
\alias{unitconversion.millicoulomb.to.coulomb}
\alias{unitconversion.mC.to.C}
\alias{unitconversion.millivolt.to.volt}
\alias{unitconversion.mV.to.V}
\alias{unitconversion.millifarad.to.farad}
\alias{unitconversion.mF.to.F}
\alias{unitconversion.milliohm.to.ohm}
\alias{unitconversion.millisiemens.to.siemens}
\alias{unitconversion.mS.to.S}
\alias{unitconversion.milliweber.to.weber}
\alias{unitconversion.mWb.to.Wb}
\alias{unitconversion.millitesla.to.tesla}
\alias{unitconversion.mT.to.T}
\alias{unitconversion.millihenry.to.henry}
\alias{unitconversion.mH.to.H}
\alias{unitconversion.millilumen.to.lumen}
\alias{unitconversion.mlm.to.lm}
\alias{unitconversion.millilux.to.lux}
\alias{unitconversion.mlx.to.lx}
\alias{unitconversion.millibecquerel.to.becquerel}
\alias{unitconversion.mBq.to.Bq}
\alias{unitconversion.milligray.to.gray}
\alias{unitconversion.mGy.to.Gy}
\alias{unitconversion.millisievert.to.sievert}
\alias{unitconversion.mSv.to.Sv}
\alias{unitconversion.millikatal.to.katal}
\alias{unitconversion.mkat.to.kat}
\title{Unit Conversion - SI Prefixes - Milli- to Base}
\usage{
unitconversion.siprefix.milli.to.base(x = 1)

unitconversion.millisecond.to.second(x = 1)

unitconversion.ms.to.s(x = 1)

unitconversion.millimeter.to.meter(x = 1)

unitconversion.mm.to.m(x = 1)

unitconversion.milligram.to.gram(x = 1)

unitconversion.mg.to.g(x = 1)

unitconversion.milliampere.to.ampere(x = 1)

unitconversion.mA.to.A(x = 1)

unitconversion.millikelvin.to.kelvin(x = 1)

unitconversion.mK.to.K(x = 1)

unitconversion.millimole.to.mole(x = 1)

unitconversion.mmol.to.mol(x = 1)

unitconversion.millicandela.to.candela(x = 1)

unitconversion.mcd.to.cd(x = 1)

unitconversion.milliradian.to.radian(x = 1)

unitconversion.mrad.to.rad(x = 1)

unitconversion.millisteradian.to.steradian(x = 1)

unitconversion.msr.to.sr(x = 1)

unitconversion.millihertz.to.hertz(x = 1)

unitconversion.mHz.to.Hz(x = 1)

unitconversion.millinewton.to.newton(x = 1)

unitconversion.mN.to.N(x = 1)

unitconversion.millipascal.to.pascal(x = 1)

unitconversion.mPa.to.Pa(x = 1)

unitconversion.millijoule.to.joule(x = 1)

unitconversion.mJ.to.J(x = 1)

unitconversion.milliwatt.to.watt(x = 1)

unitconversion.mW.to.W(x = 1)

unitconversion.millicoulomb.to.coulomb(x = 1)

unitconversion.mC.to.C(x = 1)

unitconversion.millivolt.to.volt(x = 1)

unitconversion.mV.to.V(x = 1)

unitconversion.millifarad.to.farad(x = 1)

unitconversion.mF.to.F(x = 1)

unitconversion.milliohm.to.ohm(x = 1)

unitconversion.millisiemens.to.siemens(x = 1)

unitconversion.mS.to.S(x = 1)

unitconversion.milliweber.to.weber(x = 1)

unitconversion.mWb.to.Wb(x = 1)

unitconversion.millitesla.to.tesla(x = 1)

unitconversion.mT.to.T(x = 1)

unitconversion.millihenry.to.henry(x = 1)

unitconversion.mH.to.H(x = 1)

unitconversion.millilumen.to.lumen(x = 1)

unitconversion.mlm.to.lm(x = 1)

unitconversion.millilux.to.lux(x = 1)

unitconversion.mlx.to.lx(x = 1)

unitconversion.millibecquerel.to.becquerel(x = 1)

unitconversion.mBq.to.Bq(x = 1)

unitconversion.milligray.to.gray(x = 1)

unitconversion.mGy.to.Gy(x = 1)

unitconversion.millisievert.to.sievert(x = 1)

unitconversion.mSv.to.Sv(x = 1)

unitconversion.millikatal.to.katal(x = 1)

unitconversion.mkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of milli-units}
}
\value{
x, but converted to base units
}
\description{
Milli- thousandths, 1/1000, or 10^(-3)
}
\details{
Performs a conversion from milli-units to base units (ex. milligrams to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
