% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.nano.to.base.R
\name{unitconversion.siprefix.nano.to.base}
\alias{unitconversion.siprefix.nano.to.base}
\alias{unitconversion.nanosecond.to.second}
\alias{unitconversion.ns.to.s}
\alias{unitconversion.nanometer.to.meter}
\alias{unitconversion.nm.to.m}
\alias{unitconversion.nanogram.to.gram}
\alias{unitconversion.ng.to.g}
\alias{unitconversion.nanoampere.to.ampere}
\alias{unitconversion.nA.to.A}
\alias{unitconversion.nanokelvin.to.kelvin}
\alias{unitconversion.nK.to.K}
\alias{unitconversion.nanomole.to.mole}
\alias{unitconversion.nmol.to.mol}
\alias{unitconversion.nanocandela.to.candela}
\alias{unitconversion.ncd.to.cd}
\alias{unitconversion.nanoradian.to.radian}
\alias{unitconversion.nrad.to.rad}
\alias{unitconversion.nanosteradian.to.steradian}
\alias{unitconversion.nsr.to.sr}
\alias{unitconversion.nanohertz.to.hertz}
\alias{unitconversion.nHz.to.Hz}
\alias{unitconversion.nanonewton.to.newton}
\alias{unitconversion.nN.to.N}
\alias{unitconversion.nanopascal.to.pascal}
\alias{unitconversion.nPa.to.Pa}
\alias{unitconversion.nanojoule.to.joule}
\alias{unitconversion.nJ.to.J}
\alias{unitconversion.nanowatt.to.watt}
\alias{unitconversion.nW.to.W}
\alias{unitconversion.nanocoulomb.to.coulomb}
\alias{unitconversion.nC.to.C}
\alias{unitconversion.nanovolt.to.volt}
\alias{unitconversion.nV.to.V}
\alias{unitconversion.nanofarad.to.farad}
\alias{unitconversion.nF.to.F}
\alias{unitconversion.nanoohm.to.ohm}
\alias{unitconversion.nanosiemens.to.siemens}
\alias{unitconversion.nS.to.S}
\alias{unitconversion.nanoweber.to.weber}
\alias{unitconversion.nWb.to.Wb}
\alias{unitconversion.nanotesla.to.tesla}
\alias{unitconversion.nT.to.T}
\alias{unitconversion.nanohenry.to.henry}
\alias{unitconversion.nH.to.H}
\alias{unitconversion.nanolumen.to.lumen}
\alias{unitconversion.nlm.to.lm}
\alias{unitconversion.nanolux.to.lux}
\alias{unitconversion.nlx.to.lx}
\alias{unitconversion.nanobecquerel.to.becquerel}
\alias{unitconversion.nBq.to.Bq}
\alias{unitconversion.nanogray.to.gray}
\alias{unitconversion.nGy.to.Gy}
\alias{unitconversion.nanosievert.to.sievert}
\alias{unitconversion.nSv.to.Sv}
\alias{unitconversion.nanokatal.to.katal}
\alias{unitconversion.nkat.to.kat}
\title{Unit Conversion - SI Prefixes - Nano- to Base}
\usage{
unitconversion.siprefix.nano.to.base(x = 1)

unitconversion.nanosecond.to.second(x = 1)

unitconversion.ns.to.s(x = 1)

unitconversion.nanometer.to.meter(x = 1)

unitconversion.nm.to.m(x = 1)

unitconversion.nanogram.to.gram(x = 1)

unitconversion.ng.to.g(x = 1)

unitconversion.nanoampere.to.ampere(x = 1)

unitconversion.nA.to.A(x = 1)

unitconversion.nanokelvin.to.kelvin(x = 1)

unitconversion.nK.to.K(x = 1)

unitconversion.nanomole.to.mole(x = 1)

unitconversion.nmol.to.mol(x = 1)

unitconversion.nanocandela.to.candela(x = 1)

unitconversion.ncd.to.cd(x = 1)

unitconversion.nanoradian.to.radian(x = 1)

unitconversion.nrad.to.rad(x = 1)

unitconversion.nanosteradian.to.steradian(x = 1)

unitconversion.nsr.to.sr(x = 1)

unitconversion.nanohertz.to.hertz(x = 1)

unitconversion.nHz.to.Hz(x = 1)

unitconversion.nanonewton.to.newton(x = 1)

unitconversion.nN.to.N(x = 1)

unitconversion.nanopascal.to.pascal(x = 1)

unitconversion.nPa.to.Pa(x = 1)

unitconversion.nanojoule.to.joule(x = 1)

unitconversion.nJ.to.J(x = 1)

unitconversion.nanowatt.to.watt(x = 1)

unitconversion.nW.to.W(x = 1)

unitconversion.nanocoulomb.to.coulomb(x = 1)

unitconversion.nC.to.C(x = 1)

unitconversion.nanovolt.to.volt(x = 1)

unitconversion.nV.to.V(x = 1)

unitconversion.nanofarad.to.farad(x = 1)

unitconversion.nF.to.F(x = 1)

unitconversion.nanoohm.to.ohm(x = 1)

unitconversion.nanosiemens.to.siemens(x = 1)

unitconversion.nS.to.S(x = 1)

unitconversion.nanoweber.to.weber(x = 1)

unitconversion.nWb.to.Wb(x = 1)

unitconversion.nanotesla.to.tesla(x = 1)

unitconversion.nT.to.T(x = 1)

unitconversion.nanohenry.to.henry(x = 1)

unitconversion.nH.to.H(x = 1)

unitconversion.nanolumen.to.lumen(x = 1)

unitconversion.nlm.to.lm(x = 1)

unitconversion.nanolux.to.lux(x = 1)

unitconversion.nlx.to.lx(x = 1)

unitconversion.nanobecquerel.to.becquerel(x = 1)

unitconversion.nBq.to.Bq(x = 1)

unitconversion.nanogray.to.gray(x = 1)

unitconversion.nGy.to.Gy(x = 1)

unitconversion.nanosievert.to.sievert(x = 1)

unitconversion.nSv.to.Sv(x = 1)

unitconversion.nanokatal.to.katal(x = 1)

unitconversion.nkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of nano-units}
}
\value{
x, but converted to base units
}
\description{
Nano- billionths, 1/1 000 000 000, or 10^(-9)
}
\details{
Performs a conversion from nano-units to base units (ex. nanograms to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
