% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.peta.to.base.R
\name{unitconversion.siprefix.peta.to.base}
\alias{unitconversion.siprefix.peta.to.base}
\alias{unitconversion.petasecond.to.second}
\alias{unitconversion.Ps.to.s}
\alias{unitconversion.petameter.to.meter}
\alias{unitconversion.Pm.to.m}
\alias{unitconversion.petagram.to.gram}
\alias{unitconversion.Pg.to.g}
\alias{unitconversion.petaampere.to.ampere}
\alias{unitconversion.PA.to.A}
\alias{unitconversion.petakelvin.to.kelvin}
\alias{unitconversion.PK.to.K}
\alias{unitconversion.petamole.to.mole}
\alias{unitconversion.Pmol.to.mol}
\alias{unitconversion.petacandela.to.candela}
\alias{unitconversion.Pcd.to.cd}
\alias{unitconversion.petaradian.to.radian}
\alias{unitconversion.Prad.to.rad}
\alias{unitconversion.petasteradian.to.steradian}
\alias{unitconversion.Psr.to.sr}
\alias{unitconversion.petahertz.to.hertz}
\alias{unitconversion.PHz.to.Hz}
\alias{unitconversion.petanewton.to.newton}
\alias{unitconversion.PN.to.N}
\alias{unitconversion.petapascal.to.pascal}
\alias{unitconversion.PPa.to.Pa}
\alias{unitconversion.petajoule.to.joule}
\alias{unitconversion.PJ.to.J}
\alias{unitconversion.petawatt.to.watt}
\alias{unitconversion.PW.to.W}
\alias{unitconversion.petacoulomb.to.coulomb}
\alias{unitconversion.PC.to.C}
\alias{unitconversion.petavolt.to.volt}
\alias{unitconversion.PV.to.V}
\alias{unitconversion.petafarad.to.farad}
\alias{unitconversion.PF.to.F}
\alias{unitconversion.petaohm.to.ohm}
\alias{unitconversion.petasiemens.to.siemens}
\alias{unitconversion.PS.to.S}
\alias{unitconversion.petaweber.to.weber}
\alias{unitconversion.PWb.to.Wb}
\alias{unitconversion.petatesla.to.tesla}
\alias{unitconversion.PT.to.T}
\alias{unitconversion.petahenry.to.henry}
\alias{unitconversion.PH.to.H}
\alias{unitconversion.petalumen.to.lumen}
\alias{unitconversion.Plm.to.lm}
\alias{unitconversion.petalux.to.lux}
\alias{unitconversion.Plx.to.lx}
\alias{unitconversion.petabecquerel.to.becquerel}
\alias{unitconversion.PBq.to.Bq}
\alias{unitconversion.petagray.to.gray}
\alias{unitconversion.PGy.to.Gy}
\alias{unitconversion.petasievert.to.sievert}
\alias{unitconversion.PSv.to.Sv}
\alias{unitconversion.petakatal.to.katal}
\alias{unitconversion.Pkat.to.kat}
\title{Unit Conversion - SI Prefixes - Peta- to Base}
\usage{
unitconversion.siprefix.peta.to.base(x = 1)

unitconversion.petasecond.to.second(x = 1)

unitconversion.Ps.to.s(x = 1)

unitconversion.petameter.to.meter(x = 1)

unitconversion.Pm.to.m(x = 1)

unitconversion.petagram.to.gram(x = 1)

unitconversion.Pg.to.g(x = 1)

unitconversion.petaampere.to.ampere(x = 1)

unitconversion.PA.to.A(x = 1)

unitconversion.petakelvin.to.kelvin(x = 1)

unitconversion.PK.to.K(x = 1)

unitconversion.petamole.to.mole(x = 1)

unitconversion.Pmol.to.mol(x = 1)

unitconversion.petacandela.to.candela(x = 1)

unitconversion.Pcd.to.cd(x = 1)

unitconversion.petaradian.to.radian(x = 1)

unitconversion.Prad.to.rad(x = 1)

unitconversion.petasteradian.to.steradian(x = 1)

unitconversion.Psr.to.sr(x = 1)

unitconversion.petahertz.to.hertz(x = 1)

unitconversion.PHz.to.Hz(x = 1)

unitconversion.petanewton.to.newton(x = 1)

unitconversion.PN.to.N(x = 1)

unitconversion.petapascal.to.pascal(x = 1)

unitconversion.PPa.to.Pa(x = 1)

unitconversion.petajoule.to.joule(x = 1)

unitconversion.PJ.to.J(x = 1)

unitconversion.petawatt.to.watt(x = 1)

unitconversion.PW.to.W(x = 1)

unitconversion.petacoulomb.to.coulomb(x = 1)

unitconversion.PC.to.C(x = 1)

unitconversion.petavolt.to.volt(x = 1)

unitconversion.PV.to.V(x = 1)

unitconversion.petafarad.to.farad(x = 1)

unitconversion.PF.to.F(x = 1)

unitconversion.petaohm.to.ohm(x = 1)

unitconversion.petasiemens.to.siemens(x = 1)

unitconversion.PS.to.S(x = 1)

unitconversion.petaweber.to.weber(x = 1)

unitconversion.PWb.to.Wb(x = 1)

unitconversion.petatesla.to.tesla(x = 1)

unitconversion.PT.to.T(x = 1)

unitconversion.petahenry.to.henry(x = 1)

unitconversion.PH.to.H(x = 1)

unitconversion.petalumen.to.lumen(x = 1)

unitconversion.Plm.to.lm(x = 1)

unitconversion.petalux.to.lux(x = 1)

unitconversion.Plx.to.lx(x = 1)

unitconversion.petabecquerel.to.becquerel(x = 1)

unitconversion.PBq.to.Bq(x = 1)

unitconversion.petagray.to.gray(x = 1)

unitconversion.PGy.to.Gy(x = 1)

unitconversion.petasievert.to.sievert(x = 1)

unitconversion.PSv.to.Sv(x = 1)

unitconversion.petakatal.to.katal(x = 1)

unitconversion.Pkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of peta-units}
}
\value{
x, but converted to base units
}
\description{
Peta- quadrillions, 1 000 000 000 000 000, or 10^15
}
\details{
Performs a conversion from peta-units to base units (ex. petagrams to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
