% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.pico.to.base.R
\name{unitconversion.siprefix.pico.to.base}
\alias{unitconversion.siprefix.pico.to.base}
\alias{unitconversion.picosecond.to.second}
\alias{unitconversion.ps.to.s}
\alias{unitconversion.picometer.to.meter}
\alias{unitconversion.pm.to.m}
\alias{unitconversion.picogram.to.gram}
\alias{unitconversion.pg.to.g}
\alias{unitconversion.picoampere.to.ampere}
\alias{unitconversion.pA.to.A}
\alias{unitconversion.picokelvin.to.kelvin}
\alias{unitconversion.pK.to.K}
\alias{unitconversion.picomole.to.mole}
\alias{unitconversion.pmol.to.mol}
\alias{unitconversion.picocandela.to.candela}
\alias{unitconversion.pcd.to.cd}
\alias{unitconversion.picoradian.to.radian}
\alias{unitconversion.prad.to.rad}
\alias{unitconversion.picosteradian.to.steradian}
\alias{unitconversion.psr.to.sr}
\alias{unitconversion.picohertz.to.hertz}
\alias{unitconversion.pHz.to.Hz}
\alias{unitconversion.piconewton.to.newton}
\alias{unitconversion.pN.to.N}
\alias{unitconversion.picopascal.to.pascal}
\alias{unitconversion.pPa.to.Pa}
\alias{unitconversion.picojoule.to.joule}
\alias{unitconversion.pJ.to.J}
\alias{unitconversion.picowatt.to.watt}
\alias{unitconversion.pW.to.W}
\alias{unitconversion.picocoulomb.to.coulomb}
\alias{unitconversion.pC.to.C}
\alias{unitconversion.picovolt.to.volt}
\alias{unitconversion.pV.to.V}
\alias{unitconversion.picofarad.to.farad}
\alias{unitconversion.pF.to.F}
\alias{unitconversion.picoohm.to.ohm}
\alias{unitconversion.picosiemens.to.siemens}
\alias{unitconversion.pS.to.S}
\alias{unitconversion.picoweber.to.weber}
\alias{unitconversion.pWb.to.Wb}
\alias{unitconversion.picotesla.to.tesla}
\alias{unitconversion.pT.to.T}
\alias{unitconversion.picohenry.to.henry}
\alias{unitconversion.pH.to.H}
\alias{unitconversion.picolumen.to.lumen}
\alias{unitconversion.plm.to.lm}
\alias{unitconversion.picolux.to.lux}
\alias{unitconversion.plx.to.lx}
\alias{unitconversion.picobecquerel.to.becquerel}
\alias{unitconversion.pBq.to.Bq}
\alias{unitconversion.picogray.to.gray}
\alias{unitconversion.pGy.to.Gy}
\alias{unitconversion.picosievert.to.sievert}
\alias{unitconversion.pSv.to.Sv}
\alias{unitconversion.picokatal.to.katal}
\alias{unitconversion.pkat.to.kat}
\title{Unit Conversion - SI Prefixes - Pico- to Base}
\usage{
unitconversion.siprefix.pico.to.base(x = 1)

unitconversion.picosecond.to.second(x = 1)

unitconversion.ps.to.s(x = 1)

unitconversion.picometer.to.meter(x = 1)

unitconversion.pm.to.m(x = 1)

unitconversion.picogram.to.gram(x = 1)

unitconversion.pg.to.g(x = 1)

unitconversion.picoampere.to.ampere(x = 1)

unitconversion.pA.to.A(x = 1)

unitconversion.picokelvin.to.kelvin(x = 1)

unitconversion.pK.to.K(x = 1)

unitconversion.picomole.to.mole(x = 1)

unitconversion.pmol.to.mol(x = 1)

unitconversion.picocandela.to.candela(x = 1)

unitconversion.pcd.to.cd(x = 1)

unitconversion.picoradian.to.radian(x = 1)

unitconversion.prad.to.rad(x = 1)

unitconversion.picosteradian.to.steradian(x = 1)

unitconversion.psr.to.sr(x = 1)

unitconversion.picohertz.to.hertz(x = 1)

unitconversion.pHz.to.Hz(x = 1)

unitconversion.piconewton.to.newton(x = 1)

unitconversion.pN.to.N(x = 1)

unitconversion.picopascal.to.pascal(x = 1)

unitconversion.pPa.to.Pa(x = 1)

unitconversion.picojoule.to.joule(x = 1)

unitconversion.pJ.to.J(x = 1)

unitconversion.picowatt.to.watt(x = 1)

unitconversion.pW.to.W(x = 1)

unitconversion.picocoulomb.to.coulomb(x = 1)

unitconversion.pC.to.C(x = 1)

unitconversion.picovolt.to.volt(x = 1)

unitconversion.pV.to.V(x = 1)

unitconversion.picofarad.to.farad(x = 1)

unitconversion.pF.to.F(x = 1)

unitconversion.picoohm.to.ohm(x = 1)

unitconversion.picosiemens.to.siemens(x = 1)

unitconversion.pS.to.S(x = 1)

unitconversion.picoweber.to.weber(x = 1)

unitconversion.pWb.to.Wb(x = 1)

unitconversion.picotesla.to.tesla(x = 1)

unitconversion.pT.to.T(x = 1)

unitconversion.picohenry.to.henry(x = 1)

unitconversion.pH.to.H(x = 1)

unitconversion.picolumen.to.lumen(x = 1)

unitconversion.plm.to.lm(x = 1)

unitconversion.picolux.to.lux(x = 1)

unitconversion.plx.to.lx(x = 1)

unitconversion.picobecquerel.to.becquerel(x = 1)

unitconversion.pBq.to.Bq(x = 1)

unitconversion.picogray.to.gray(x = 1)

unitconversion.pGy.to.Gy(x = 1)

unitconversion.picosievert.to.sievert(x = 1)

unitconversion.pSv.to.Sv(x = 1)

unitconversion.picokatal.to.katal(x = 1)

unitconversion.pkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pico-units}
}
\value{
x, but converted to base units
}
\description{
Pico- trillionths, 1/1 000 000 000 000, or 10^(-12)
}
\details{
Performs a conversion from pico-units to base units (ex. picograms to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
