% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.tera.to.base.R
\name{unitconversion.siprefix.tera.to.base}
\alias{unitconversion.siprefix.tera.to.base}
\alias{unitconversion.terasecond.to.second}
\alias{unitconversion.Ts.to.s}
\alias{unitconversion.terameter.to.meter}
\alias{unitconversion.Tm.to.m}
\alias{unitconversion.teragram.to.gram}
\alias{unitconversion.Tg.to.g}
\alias{unitconversion.teraampere.to.ampere}
\alias{unitconversion.TA.to.A}
\alias{unitconversion.terakelvin.to.kelvin}
\alias{unitconversion.TK.to.K}
\alias{unitconversion.teramole.to.mole}
\alias{unitconversion.Tmol.to.mol}
\alias{unitconversion.teracandela.to.candela}
\alias{unitconversion.Tcd.to.cd}
\alias{unitconversion.teraradian.to.radian}
\alias{unitconversion.Trad.to.rad}
\alias{unitconversion.terasteradian.to.steradian}
\alias{unitconversion.Tsr.to.sr}
\alias{unitconversion.terahertz.to.hertz}
\alias{unitconversion.THz.to.Hz}
\alias{unitconversion.teranewton.to.newton}
\alias{unitconversion.TN.to.N}
\alias{unitconversion.terapascal.to.pascal}
\alias{unitconversion.TPa.to.Pa}
\alias{unitconversion.terajoule.to.joule}
\alias{unitconversion.TJ.to.J}
\alias{unitconversion.terawatt.to.watt}
\alias{unitconversion.TW.to.W}
\alias{unitconversion.teracoulomb.to.coulomb}
\alias{unitconversion.TC.to.C}
\alias{unitconversion.teravolt.to.volt}
\alias{unitconversion.TV.to.V}
\alias{unitconversion.terafarad.to.farad}
\alias{unitconversion.TF.to.F}
\alias{unitconversion.teraohm.to.ohm}
\alias{unitconversion.terasiemens.to.siemens}
\alias{unitconversion.TS.to.S}
\alias{unitconversion.teraweber.to.weber}
\alias{unitconversion.TWb.to.Wb}
\alias{unitconversion.teratesla.to.tesla}
\alias{unitconversion.TT.to.T}
\alias{unitconversion.terahenry.to.henry}
\alias{unitconversion.TH.to.H}
\alias{unitconversion.teralumen.to.lumen}
\alias{unitconversion.Tlm.to.lm}
\alias{unitconversion.teralux.to.lux}
\alias{unitconversion.Tlx.to.lx}
\alias{unitconversion.terabecquerel.to.becquerel}
\alias{unitconversion.TBq.to.Bq}
\alias{unitconversion.teragray.to.gray}
\alias{unitconversion.TGy.to.Gy}
\alias{unitconversion.terasievert.to.sievert}
\alias{unitconversion.TSv.to.Sv}
\alias{unitconversion.terakatal.to.katal}
\alias{unitconversion.Tkat.to.kat}
\title{Unit Conversion - SI Prefixes - Tera- to Base}
\usage{
unitconversion.siprefix.tera.to.base(x = 1)

unitconversion.terasecond.to.second(x = 1)

unitconversion.Ts.to.s(x = 1)

unitconversion.terameter.to.meter(x = 1)

unitconversion.Tm.to.m(x = 1)

unitconversion.teragram.to.gram(x = 1)

unitconversion.Tg.to.g(x = 1)

unitconversion.teraampere.to.ampere(x = 1)

unitconversion.TA.to.A(x = 1)

unitconversion.terakelvin.to.kelvin(x = 1)

unitconversion.TK.to.K(x = 1)

unitconversion.teramole.to.mole(x = 1)

unitconversion.Tmol.to.mol(x = 1)

unitconversion.teracandela.to.candela(x = 1)

unitconversion.Tcd.to.cd(x = 1)

unitconversion.teraradian.to.radian(x = 1)

unitconversion.Trad.to.rad(x = 1)

unitconversion.terasteradian.to.steradian(x = 1)

unitconversion.Tsr.to.sr(x = 1)

unitconversion.terahertz.to.hertz(x = 1)

unitconversion.THz.to.Hz(x = 1)

unitconversion.teranewton.to.newton(x = 1)

unitconversion.TN.to.N(x = 1)

unitconversion.terapascal.to.pascal(x = 1)

unitconversion.TPa.to.Pa(x = 1)

unitconversion.terajoule.to.joule(x = 1)

unitconversion.TJ.to.J(x = 1)

unitconversion.terawatt.to.watt(x = 1)

unitconversion.TW.to.W(x = 1)

unitconversion.teracoulomb.to.coulomb(x = 1)

unitconversion.TC.to.C(x = 1)

unitconversion.teravolt.to.volt(x = 1)

unitconversion.TV.to.V(x = 1)

unitconversion.terafarad.to.farad(x = 1)

unitconversion.TF.to.F(x = 1)

unitconversion.teraohm.to.ohm(x = 1)

unitconversion.terasiemens.to.siemens(x = 1)

unitconversion.TS.to.S(x = 1)

unitconversion.teraweber.to.weber(x = 1)

unitconversion.TWb.to.Wb(x = 1)

unitconversion.teratesla.to.tesla(x = 1)

unitconversion.TT.to.T(x = 1)

unitconversion.terahenry.to.henry(x = 1)

unitconversion.TH.to.H(x = 1)

unitconversion.teralumen.to.lumen(x = 1)

unitconversion.Tlm.to.lm(x = 1)

unitconversion.teralux.to.lux(x = 1)

unitconversion.Tlx.to.lx(x = 1)

unitconversion.terabecquerel.to.becquerel(x = 1)

unitconversion.TBq.to.Bq(x = 1)

unitconversion.teragray.to.gray(x = 1)

unitconversion.TGy.to.Gy(x = 1)

unitconversion.terasievert.to.sievert(x = 1)

unitconversion.TSv.to.Sv(x = 1)

unitconversion.terakatal.to.katal(x = 1)

unitconversion.Tkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of tera-units}
}
\value{
x, but converted to base units
}
\description{
Tera- trillions, 1 000 000 000 000, or 10^12
}
\details{
Performs a conversion from tera-units to base units (ex. teragrams to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
