% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.yocto.to.base.R
\name{unitconversion.siprefix.yocto.to.base}
\alias{unitconversion.siprefix.yocto.to.base}
\alias{unitconversion.yoctosecond.to.second}
\alias{unitconversion.ys.to.s}
\alias{unitconversion.yoctometer.to.meter}
\alias{unitconversion.ym.to.m}
\alias{unitconversion.yoctogram.to.gram}
\alias{unitconversion.yg.to.g}
\alias{unitconversion.yoctoampere.to.ampere}
\alias{unitconversion.yA.to.A}
\alias{unitconversion.yoctokelvin.to.kelvin}
\alias{unitconversion.yK.to.K}
\alias{unitconversion.yoctomole.to.mole}
\alias{unitconversion.ymol.to.mol}
\alias{unitconversion.yoctocandela.to.candela}
\alias{unitconversion.ycd.to.cd}
\alias{unitconversion.yoctoradian.to.radian}
\alias{unitconversion.yrad.to.rad}
\alias{unitconversion.yoctosteradian.to.steradian}
\alias{unitconversion.ysr.to.sr}
\alias{unitconversion.yoctohertz.to.hertz}
\alias{unitconversion.yHz.to.Hz}
\alias{unitconversion.yoctonewton.to.newton}
\alias{unitconversion.yN.to.N}
\alias{unitconversion.yoctopascal.to.pascal}
\alias{unitconversion.yPa.to.Pa}
\alias{unitconversion.yoctojoule.to.joule}
\alias{unitconversion.yJ.to.J}
\alias{unitconversion.yoctowatt.to.watt}
\alias{unitconversion.yW.to.W}
\alias{unitconversion.yoctocoulomb.to.coulomb}
\alias{unitconversion.yC.to.C}
\alias{unitconversion.yoctovolt.to.volt}
\alias{unitconversion.yV.to.V}
\alias{unitconversion.yoctofarad.to.farad}
\alias{unitconversion.yF.to.F}
\alias{unitconversion.yoctoohm.to.ohm}
\alias{unitconversion.yoctosiemens.to.siemens}
\alias{unitconversion.yS.to.S}
\alias{unitconversion.yoctoweber.to.weber}
\alias{unitconversion.yWb.to.Wb}
\alias{unitconversion.yoctotesla.to.tesla}
\alias{unitconversion.yT.to.T}
\alias{unitconversion.yoctohenry.to.henry}
\alias{unitconversion.yH.to.H}
\alias{unitconversion.yoctolumen.to.lumen}
\alias{unitconversion.ylm.to.lm}
\alias{unitconversion.yoctolux.to.lux}
\alias{unitconversion.ylx.to.lx}
\alias{unitconversion.yoctobecquerel.to.becquerel}
\alias{unitconversion.yBq.to.Bq}
\alias{unitconversion.yoctogray.to.gray}
\alias{unitconversion.yGy.to.Gy}
\alias{unitconversion.yoctosievert.to.sievert}
\alias{unitconversion.ySv.to.Sv}
\alias{unitconversion.yoctokatal.to.katal}
\alias{unitconversion.ykat.to.kat}
\title{Unit Conversion - SI Prefixes - Yocto- to Base}
\usage{
unitconversion.siprefix.yocto.to.base(x = 1)

unitconversion.yoctosecond.to.second(x = 1)

unitconversion.ys.to.s(x = 1)

unitconversion.yoctometer.to.meter(x = 1)

unitconversion.ym.to.m(x = 1)

unitconversion.yoctogram.to.gram(x = 1)

unitconversion.yg.to.g(x = 1)

unitconversion.yoctoampere.to.ampere(x = 1)

unitconversion.yA.to.A(x = 1)

unitconversion.yoctokelvin.to.kelvin(x = 1)

unitconversion.yK.to.K(x = 1)

unitconversion.yoctomole.to.mole(x = 1)

unitconversion.ymol.to.mol(x = 1)

unitconversion.yoctocandela.to.candela(x = 1)

unitconversion.ycd.to.cd(x = 1)

unitconversion.yoctoradian.to.radian(x = 1)

unitconversion.yrad.to.rad(x = 1)

unitconversion.yoctosteradian.to.steradian(x = 1)

unitconversion.ysr.to.sr(x = 1)

unitconversion.yoctohertz.to.hertz(x = 1)

unitconversion.yHz.to.Hz(x = 1)

unitconversion.yoctonewton.to.newton(x = 1)

unitconversion.yN.to.N(x = 1)

unitconversion.yoctopascal.to.pascal(x = 1)

unitconversion.yPa.to.Pa(x = 1)

unitconversion.yoctojoule.to.joule(x = 1)

unitconversion.yJ.to.J(x = 1)

unitconversion.yoctowatt.to.watt(x = 1)

unitconversion.yW.to.W(x = 1)

unitconversion.yoctocoulomb.to.coulomb(x = 1)

unitconversion.yC.to.C(x = 1)

unitconversion.yoctovolt.to.volt(x = 1)

unitconversion.yV.to.V(x = 1)

unitconversion.yoctofarad.to.farad(x = 1)

unitconversion.yF.to.F(x = 1)

unitconversion.yoctoohm.to.ohm(x = 1)

unitconversion.yoctosiemens.to.siemens(x = 1)

unitconversion.yS.to.S(x = 1)

unitconversion.yoctoweber.to.weber(x = 1)

unitconversion.yWb.to.Wb(x = 1)

unitconversion.yoctotesla.to.tesla(x = 1)

unitconversion.yT.to.T(x = 1)

unitconversion.yoctohenry.to.henry(x = 1)

unitconversion.yH.to.H(x = 1)

unitconversion.yoctolumen.to.lumen(x = 1)

unitconversion.ylm.to.lm(x = 1)

unitconversion.yoctolux.to.lux(x = 1)

unitconversion.ylx.to.lx(x = 1)

unitconversion.yoctobecquerel.to.becquerel(x = 1)

unitconversion.yBq.to.Bq(x = 1)

unitconversion.yoctogray.to.gray(x = 1)

unitconversion.yGy.to.Gy(x = 1)

unitconversion.yoctosievert.to.sievert(x = 1)

unitconversion.ySv.to.Sv(x = 1)

unitconversion.yoctokatal.to.katal(x = 1)

unitconversion.ykat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of yocto-units}
}
\value{
x, but converted to base units
}
\description{
Yocto- means 1/1 000 000 000 000 000 000 000 000 or 10^(-24)
}
\details{
Performs a conversion from yocto-units to base units (ex. yoctograms to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
