% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.yotta.to.base.R
\name{unitconversion.siprefix.yotta.to.base}
\alias{unitconversion.siprefix.yotta.to.base}
\alias{unitconversion.yottasecond.to.second}
\alias{unitconversion.Ys.to.s}
\alias{unitconversion.yottameter.to.meter}
\alias{unitconversion.Ym.to.m}
\alias{unitconversion.yottagram.to.gram}
\alias{unitconversion.Yg.to.g}
\alias{unitconversion.yottaampere.to.ampere}
\alias{unitconversion.YA.to.A}
\alias{unitconversion.yottakelvin.to.kelvin}
\alias{unitconversion.YK.to.K}
\alias{unitconversion.yottamole.to.mole}
\alias{unitconversion.Ymol.to.mol}
\alias{unitconversion.yottacandela.to.candela}
\alias{unitconversion.Ycd.to.cd}
\alias{unitconversion.yottaradian.to.radian}
\alias{unitconversion.Yrad.to.rad}
\alias{unitconversion.yottasteradian.to.steradian}
\alias{unitconversion.Ysr.to.sr}
\alias{unitconversion.yottahertz.to.hertz}
\alias{unitconversion.YHz.to.Hz}
\alias{unitconversion.yottanewton.to.newton}
\alias{unitconversion.YN.to.N}
\alias{unitconversion.yottapascal.to.pascal}
\alias{unitconversion.YPa.to.Pa}
\alias{unitconversion.yottajoule.to.joule}
\alias{unitconversion.YJ.to.J}
\alias{unitconversion.yottawatt.to.watt}
\alias{unitconversion.YW.to.W}
\alias{unitconversion.yottacoulomb.to.coulomb}
\alias{unitconversion.YC.to.C}
\alias{unitconversion.yottavolt.to.volt}
\alias{unitconversion.YV.to.V}
\alias{unitconversion.yottafarad.to.farad}
\alias{unitconversion.YF.to.F}
\alias{unitconversion.yottaohm.to.ohm}
\alias{unitconversion.yottasiemens.to.siemens}
\alias{unitconversion.YS.to.S}
\alias{unitconversion.yottaweber.to.weber}
\alias{unitconversion.YWb.to.Wb}
\alias{unitconversion.yottatesla.to.tesla}
\alias{unitconversion.YT.to.T}
\alias{unitconversion.yottahenry.to.henry}
\alias{unitconversion.YH.to.H}
\alias{unitconversion.yottalumen.to.lumen}
\alias{unitconversion.Ylm.to.lm}
\alias{unitconversion.yottalux.to.lux}
\alias{unitconversion.Ylx.to.lx}
\alias{unitconversion.yottabecquerel.to.becquerel}
\alias{unitconversion.YBq.to.Bq}
\alias{unitconversion.yottagray.to.gray}
\alias{unitconversion.YGy.to.Gy}
\alias{unitconversion.yottasievert.to.sievert}
\alias{unitconversion.YSv.to.Sv}
\alias{unitconversion.yottakatal.to.katal}
\alias{unitconversion.Ykat.to.kat}
\title{Unit Conversion - SI Prefixes - Yotta- to Base}
\usage{
unitconversion.siprefix.yotta.to.base(x = 1)

unitconversion.yottasecond.to.second(x = 1)

unitconversion.Ys.to.s(x = 1)

unitconversion.yottameter.to.meter(x = 1)

unitconversion.Ym.to.m(x = 1)

unitconversion.yottagram.to.gram(x = 1)

unitconversion.Yg.to.g(x = 1)

unitconversion.yottaampere.to.ampere(x = 1)

unitconversion.YA.to.A(x = 1)

unitconversion.yottakelvin.to.kelvin(x = 1)

unitconversion.YK.to.K(x = 1)

unitconversion.yottamole.to.mole(x = 1)

unitconversion.Ymol.to.mol(x = 1)

unitconversion.yottacandela.to.candela(x = 1)

unitconversion.Ycd.to.cd(x = 1)

unitconversion.yottaradian.to.radian(x = 1)

unitconversion.Yrad.to.rad(x = 1)

unitconversion.yottasteradian.to.steradian(x = 1)

unitconversion.Ysr.to.sr(x = 1)

unitconversion.yottahertz.to.hertz(x = 1)

unitconversion.YHz.to.Hz(x = 1)

unitconversion.yottanewton.to.newton(x = 1)

unitconversion.YN.to.N(x = 1)

unitconversion.yottapascal.to.pascal(x = 1)

unitconversion.YPa.to.Pa(x = 1)

unitconversion.yottajoule.to.joule(x = 1)

unitconversion.YJ.to.J(x = 1)

unitconversion.yottawatt.to.watt(x = 1)

unitconversion.YW.to.W(x = 1)

unitconversion.yottacoulomb.to.coulomb(x = 1)

unitconversion.YC.to.C(x = 1)

unitconversion.yottavolt.to.volt(x = 1)

unitconversion.YV.to.V(x = 1)

unitconversion.yottafarad.to.farad(x = 1)

unitconversion.YF.to.F(x = 1)

unitconversion.yottaohm.to.ohm(x = 1)

unitconversion.yottasiemens.to.siemens(x = 1)

unitconversion.YS.to.S(x = 1)

unitconversion.yottaweber.to.weber(x = 1)

unitconversion.YWb.to.Wb(x = 1)

unitconversion.yottatesla.to.tesla(x = 1)

unitconversion.YT.to.T(x = 1)

unitconversion.yottahenry.to.henry(x = 1)

unitconversion.YH.to.H(x = 1)

unitconversion.yottalumen.to.lumen(x = 1)

unitconversion.Ylm.to.lm(x = 1)

unitconversion.yottalux.to.lux(x = 1)

unitconversion.Ylx.to.lx(x = 1)

unitconversion.yottabecquerel.to.becquerel(x = 1)

unitconversion.YBq.to.Bq(x = 1)

unitconversion.yottagray.to.gray(x = 1)

unitconversion.YGy.to.Gy(x = 1)

unitconversion.yottasievert.to.sievert(x = 1)

unitconversion.YSv.to.Sv(x = 1)

unitconversion.yottakatal.to.katal(x = 1)

unitconversion.Ykat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of yotta-units}
}
\value{
x, but converted to base units
}
\description{
Yotta- means 1 000 000 000 000 000 000 000 000 or 10^24
}
\details{
Performs a conversion from yotta-units to base units (ex. yottagrams to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
