% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.zepto.to.base.R
\name{unitconversion.siprefix.zepto.to.base}
\alias{unitconversion.siprefix.zepto.to.base}
\alias{unitconversion.zeptosecond.to.second}
\alias{unitconversion.zs.to.s}
\alias{unitconversion.zeptometer.to.meter}
\alias{unitconversion.zm.to.m}
\alias{unitconversion.zeptogram.to.gram}
\alias{unitconversion.zg.to.g}
\alias{unitconversion.zeptoampere.to.ampere}
\alias{unitconversion.zA.to.A}
\alias{unitconversion.zeptokelvin.to.kelvin}
\alias{unitconversion.zK.to.K}
\alias{unitconversion.zeptomole.to.mole}
\alias{unitconversion.zmol.to.mol}
\alias{unitconversion.zeptocandela.to.candela}
\alias{unitconversion.zcd.to.cd}
\alias{unitconversion.zeptoradian.to.radian}
\alias{unitconversion.zrad.to.rad}
\alias{unitconversion.zeptosteradian.to.steradian}
\alias{unitconversion.zsr.to.sr}
\alias{unitconversion.zeptohertz.to.hertz}
\alias{unitconversion.zHz.to.Hz}
\alias{unitconversion.zeptonewton.to.newton}
\alias{unitconversion.zN.to.N}
\alias{unitconversion.zeptopascal.to.pascal}
\alias{unitconversion.zPa.to.Pa}
\alias{unitconversion.zeptojoule.to.joule}
\alias{unitconversion.zJ.to.J}
\alias{unitconversion.zeptowatt.to.watt}
\alias{unitconversion.zW.to.W}
\alias{unitconversion.zeptocoulomb.to.coulomb}
\alias{unitconversion.zC.to.C}
\alias{unitconversion.zeptovolt.to.volt}
\alias{unitconversion.zV.to.V}
\alias{unitconversion.zeptofarad.to.farad}
\alias{unitconversion.zF.to.F}
\alias{unitconversion.zeptoohm.to.ohm}
\alias{unitconversion.zeptosiemens.to.siemens}
\alias{unitconversion.zS.to.S}
\alias{unitconversion.zeptoweber.to.weber}
\alias{unitconversion.zWb.to.Wb}
\alias{unitconversion.zeptotesla.to.tesla}
\alias{unitconversion.zT.to.T}
\alias{unitconversion.zeptohenry.to.henry}
\alias{unitconversion.zH.to.H}
\alias{unitconversion.zeptolumen.to.lumen}
\alias{unitconversion.zlm.to.lm}
\alias{unitconversion.zeptolux.to.lux}
\alias{unitconversion.zlx.to.lx}
\alias{unitconversion.zeptobecquerel.to.becquerel}
\alias{unitconversion.zBq.to.Bq}
\alias{unitconversion.zeptogray.to.gray}
\alias{unitconversion.zGy.to.Gy}
\alias{unitconversion.zeptosievert.to.sievert}
\alias{unitconversion.zSv.to.Sv}
\alias{unitconversion.zeptokatal.to.katal}
\alias{unitconversion.zkat.to.kat}
\title{Unit Conversion - SI Prefixes - Zepto- to Base}
\usage{
unitconversion.siprefix.zepto.to.base(x = 1)

unitconversion.zeptosecond.to.second(x = 1)

unitconversion.zs.to.s(x = 1)

unitconversion.zeptometer.to.meter(x = 1)

unitconversion.zm.to.m(x = 1)

unitconversion.zeptogram.to.gram(x = 1)

unitconversion.zg.to.g(x = 1)

unitconversion.zeptoampere.to.ampere(x = 1)

unitconversion.zA.to.A(x = 1)

unitconversion.zeptokelvin.to.kelvin(x = 1)

unitconversion.zK.to.K(x = 1)

unitconversion.zeptomole.to.mole(x = 1)

unitconversion.zmol.to.mol(x = 1)

unitconversion.zeptocandela.to.candela(x = 1)

unitconversion.zcd.to.cd(x = 1)

unitconversion.zeptoradian.to.radian(x = 1)

unitconversion.zrad.to.rad(x = 1)

unitconversion.zeptosteradian.to.steradian(x = 1)

unitconversion.zsr.to.sr(x = 1)

unitconversion.zeptohertz.to.hertz(x = 1)

unitconversion.zHz.to.Hz(x = 1)

unitconversion.zeptonewton.to.newton(x = 1)

unitconversion.zN.to.N(x = 1)

unitconversion.zeptopascal.to.pascal(x = 1)

unitconversion.zPa.to.Pa(x = 1)

unitconversion.zeptojoule.to.joule(x = 1)

unitconversion.zJ.to.J(x = 1)

unitconversion.zeptowatt.to.watt(x = 1)

unitconversion.zW.to.W(x = 1)

unitconversion.zeptocoulomb.to.coulomb(x = 1)

unitconversion.zC.to.C(x = 1)

unitconversion.zeptovolt.to.volt(x = 1)

unitconversion.zV.to.V(x = 1)

unitconversion.zeptofarad.to.farad(x = 1)

unitconversion.zF.to.F(x = 1)

unitconversion.zeptoohm.to.ohm(x = 1)

unitconversion.zeptosiemens.to.siemens(x = 1)

unitconversion.zS.to.S(x = 1)

unitconversion.zeptoweber.to.weber(x = 1)

unitconversion.zWb.to.Wb(x = 1)

unitconversion.zeptotesla.to.tesla(x = 1)

unitconversion.zT.to.T(x = 1)

unitconversion.zeptohenry.to.henry(x = 1)

unitconversion.zH.to.H(x = 1)

unitconversion.zeptolumen.to.lumen(x = 1)

unitconversion.zlm.to.lm(x = 1)

unitconversion.zeptolux.to.lux(x = 1)

unitconversion.zlx.to.lx(x = 1)

unitconversion.zeptobecquerel.to.becquerel(x = 1)

unitconversion.zBq.to.Bq(x = 1)

unitconversion.zeptogray.to.gray(x = 1)

unitconversion.zGy.to.Gy(x = 1)

unitconversion.zeptosievert.to.sievert(x = 1)

unitconversion.zSv.to.Sv(x = 1)

unitconversion.zeptokatal.to.katal(x = 1)

unitconversion.zkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of zepto-units}
}
\value{
x, but converted to base units
}
\description{
Zepto- means 1/1 000 000 000 000 000 000 000 or 10^(-21)
}
\details{
Performs a conversion from zepto-units to base units (ex. zeptograms to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
