% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.siprefix.zetta.to.base.R
\name{unitconversion.siprefix.zetta.to.base}
\alias{unitconversion.siprefix.zetta.to.base}
\alias{unitconversion.zettasecond.to.second}
\alias{unitconversion.Zs.to.s}
\alias{unitconversion.zettameter.to.meter}
\alias{unitconversion.Zm.to.m}
\alias{unitconversion.zettagram.to.gram}
\alias{unitconversion.Zg.to.g}
\alias{unitconversion.zettaampere.to.ampere}
\alias{unitconversion.ZA.to.A}
\alias{unitconversion.zettakelvin.to.kelvin}
\alias{unitconversion.ZK.to.K}
\alias{unitconversion.zettamole.to.mole}
\alias{unitconversion.Zmol.to.mol}
\alias{unitconversion.zettacandela.to.candela}
\alias{unitconversion.Zcd.to.cd}
\alias{unitconversion.zettaradian.to.radian}
\alias{unitconversion.Zrad.to.rad}
\alias{unitconversion.zettasteradian.to.steradian}
\alias{unitconversion.Zsr.to.sr}
\alias{unitconversion.zettahertz.to.hertz}
\alias{unitconversion.ZHz.to.Hz}
\alias{unitconversion.zettanewton.to.newton}
\alias{unitconversion.ZN.to.N}
\alias{unitconversion.zettapascal.to.pascal}
\alias{unitconversion.ZPa.to.Pa}
\alias{unitconversion.zettajoule.to.joule}
\alias{unitconversion.ZJ.to.J}
\alias{unitconversion.zettawatt.to.watt}
\alias{unitconversion.ZW.to.W}
\alias{unitconversion.zettacoulomb.to.coulomb}
\alias{unitconversion.ZC.to.C}
\alias{unitconversion.zettavolt.to.volt}
\alias{unitconversion.ZV.to.V}
\alias{unitconversion.zettafarad.to.farad}
\alias{unitconversion.ZF.to.F}
\alias{unitconversion.zettaohm.to.ohm}
\alias{unitconversion.zettasiemens.to.siemens}
\alias{unitconversion.ZS.to.S}
\alias{unitconversion.zettaweber.to.weber}
\alias{unitconversion.ZWb.to.Wb}
\alias{unitconversion.zettatesla.to.tesla}
\alias{unitconversion.ZT.to.T}
\alias{unitconversion.zettahenry.to.henry}
\alias{unitconversion.ZH.to.H}
\alias{unitconversion.zettalumen.to.lumen}
\alias{unitconversion.Zlm.to.lm}
\alias{unitconversion.zettalux.to.lux}
\alias{unitconversion.Zlx.to.lx}
\alias{unitconversion.zettabecquerel.to.becquerel}
\alias{unitconversion.ZBq.to.Bq}
\alias{unitconversion.zettagray.to.gray}
\alias{unitconversion.ZGy.to.Gy}
\alias{unitconversion.zettasievert.to.sievert}
\alias{unitconversion.ZSv.to.Sv}
\alias{unitconversion.zettakatal.to.katal}
\alias{unitconversion.Zkat.to.kat}
\title{Unit Conversion - SI Prefixes - Zetta- to Base}
\usage{
unitconversion.siprefix.zetta.to.base(x = 1)

unitconversion.zettasecond.to.second(x = 1)

unitconversion.Zs.to.s(x = 1)

unitconversion.zettameter.to.meter(x = 1)

unitconversion.Zm.to.m(x = 1)

unitconversion.zettagram.to.gram(x = 1)

unitconversion.Zg.to.g(x = 1)

unitconversion.zettaampere.to.ampere(x = 1)

unitconversion.ZA.to.A(x = 1)

unitconversion.zettakelvin.to.kelvin(x = 1)

unitconversion.ZK.to.K(x = 1)

unitconversion.zettamole.to.mole(x = 1)

unitconversion.Zmol.to.mol(x = 1)

unitconversion.zettacandela.to.candela(x = 1)

unitconversion.Zcd.to.cd(x = 1)

unitconversion.zettaradian.to.radian(x = 1)

unitconversion.Zrad.to.rad(x = 1)

unitconversion.zettasteradian.to.steradian(x = 1)

unitconversion.Zsr.to.sr(x = 1)

unitconversion.zettahertz.to.hertz(x = 1)

unitconversion.ZHz.to.Hz(x = 1)

unitconversion.zettanewton.to.newton(x = 1)

unitconversion.ZN.to.N(x = 1)

unitconversion.zettapascal.to.pascal(x = 1)

unitconversion.ZPa.to.Pa(x = 1)

unitconversion.zettajoule.to.joule(x = 1)

unitconversion.ZJ.to.J(x = 1)

unitconversion.zettawatt.to.watt(x = 1)

unitconversion.ZW.to.W(x = 1)

unitconversion.zettacoulomb.to.coulomb(x = 1)

unitconversion.ZC.to.C(x = 1)

unitconversion.zettavolt.to.volt(x = 1)

unitconversion.ZV.to.V(x = 1)

unitconversion.zettafarad.to.farad(x = 1)

unitconversion.ZF.to.F(x = 1)

unitconversion.zettaohm.to.ohm(x = 1)

unitconversion.zettasiemens.to.siemens(x = 1)

unitconversion.ZS.to.S(x = 1)

unitconversion.zettaweber.to.weber(x = 1)

unitconversion.ZWb.to.Wb(x = 1)

unitconversion.zettatesla.to.tesla(x = 1)

unitconversion.ZT.to.T(x = 1)

unitconversion.zettahenry.to.henry(x = 1)

unitconversion.ZH.to.H(x = 1)

unitconversion.zettalumen.to.lumen(x = 1)

unitconversion.Zlm.to.lm(x = 1)

unitconversion.zettalux.to.lux(x = 1)

unitconversion.Zlx.to.lx(x = 1)

unitconversion.zettabecquerel.to.becquerel(x = 1)

unitconversion.ZBq.to.Bq(x = 1)

unitconversion.zettagray.to.gray(x = 1)

unitconversion.ZGy.to.Gy(x = 1)

unitconversion.zettasievert.to.sievert(x = 1)

unitconversion.ZSv.to.Sv(x = 1)

unitconversion.zettakatal.to.katal(x = 1)

unitconversion.Zkat.to.kat(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of zetta-units}
}
\value{
x, but converted to base units
}
\description{
Zetta- means 1 000 000 000 000 000 000 000 or 10^21
}
\details{
Performs a conversion from zetta-units to base units (ex. zettagrams to grams).
}
\references{
NIST. Metric (SI) Prefixes. 2022. Accessed 4/7/2022. 
https://www.nist.gov/pml/weights-and-measures/metric-si-prefixes
}
