% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.square.foot.to.square.inch.R
\name{unitconversion.square.foot.to.square.inch}
\alias{unitconversion.square.foot.to.square.inch}
\alias{unitconversion.sq.ft.to.sq.in}
\title{Unit Conversion - Area - Square Foot to Square Inch}
\usage{
unitconversion.square.foot.to.square.inch(x = 1)

unitconversion.sq.ft.to.sq.in(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of square feet}
}
\value{
x, but converted to square inch
}
\description{
Performs a conversion of areas from square feet to square inches.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
