% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.square.foot.to.square.rod.R
\name{unitconversion.square.foot.to.square.rod}
\alias{unitconversion.square.foot.to.square.rod}
\alias{unitconversion.sq.ft.to.sq.rd}
\alias{unitconversion.square.foot.to.square.pole}
\alias{unitconversion.square.foot.to.square.perch}
\title{Unit Conversion - Area - Square Foot to Square Rod}
\usage{
unitconversion.square.foot.to.square.rod(x = 1)

unitconversion.sq.ft.to.sq.rd(x = 1)

unitconversion.square.foot.to.square.pole(x = 1)

unitconversion.square.foot.to.square.perch(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of square feet}
}
\value{
x, but converted to square yards
}
\description{
Performs a conversion of areas from square feet (U.S. Survey) to square rods.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
