% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.square.inch.to.square.foot.R
\name{unitconversion.square.inch.to.square.foot}
\alias{unitconversion.square.inch.to.square.foot}
\alias{unitconversion.sq.in.to.sq.ft}
\title{Unit Conversion - Area - Square Inch to Square Foot}
\usage{
unitconversion.square.inch.to.square.foot(x = 1)

unitconversion.sq.in.to.sq.ft(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of square inches}
}
\value{
x, but converted to square feet
}
\description{
Performs a conversion of areas from square inches to square feet.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
