% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.square.meter.to.hectare.R
\name{unitconversion.square.meter.to.hectare}
\alias{unitconversion.square.meter.to.hectare}
\alias{unitconversion.sq.m.to.ha}
\title{Unit Conversion - Area - Square Meter to Hectare}
\usage{
unitconversion.square.meter.to.hectare(x = 1)

unitconversion.sq.m.to.ha(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of square meters}
}
\value{
x, but converted to hectares
}
\description{
Performs a conversion of areas from square meters to hectares.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
