% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.square.mile.to.township.R
\name{unitconversion.square.mile.to.township}
\alias{unitconversion.square.mile.to.township}
\title{Unit Conversion - Area - Square Mile to Township}
\usage{
unitconversion.square.mile.to.township(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of square miles}
}
\value{
x, but converted to townships
}
\description{
Performs a conversion of areas from square miles (U.S. Survey) to townships.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
