% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.stone.to.pound.R
\name{unitconversion.stone.to.pound}
\alias{unitconversion.stone.to.pound}
\alias{unitconversion.st.to.lb}
\title{Unit Conversion - Mass/Weight - Stone to U.S. Pound}
\usage{
unitconversion.stone.to.pound(x = 1)

unitconversion.st.to.lb(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of stones}
}
\value{
x, but converted to pounds
}
\description{
Performs a conversion of weights from stones to pounds.
}
