% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.survey.foot.to.foot.R
\name{unitconversion.survey.foot.to.foot}
\alias{unitconversion.survey.foot.to.foot}
\title{Unit Conversion - Length - U.S. Foot to Foot}
\usage{
unitconversion.survey.foot.to.foot(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of survey}
}
\value{
x, but converted to international feet
}
\description{
Performs a conversion of lengths from U.S. survey feet to feet (international).
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
