% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.therm.us.to.joule.R
\name{unitconversion.therm.us.to.joule}
\alias{unitconversion.therm.us.to.joule}
\title{Unit Conversion - Energy - U.S. Therm to Joule}
\usage{
unitconversion.therm.us.to.joule(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of therms}
}
\value{
x, but converted to joules
}
\description{
Performs a conversion of energy from U.S. therms to joules.
}
\references{
Thompson, Ambler and Taylor, Barry. NIST Special Publication 811 - Guide for the use of the International System of Units (SI). 2008.  
https://nvlpubs.nist.gov/nistpubs/Legacy/SP/nistspecialpublication811e2008.pdf
}
