% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.township.to.square.mile.R
\name{unitconversion.township.to.square.mile}
\alias{unitconversion.township.to.square.mile}
\title{Unit Conversion - Area - Township to Square Mile}
\usage{
unitconversion.township.to.square.mile(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of townships}
}
\value{
x, but converted to square miles
}
\description{
Performs a conversion of areas from townships to square miles (U.S. Survey).
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
