% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.troy.ounce.to.grain.R
\name{unitconversion.troy.ounce.to.grain}
\alias{unitconversion.troy.ounce.to.grain}
\alias{unitconversion.ozt.to.gr}
\title{Unit Conversion - Mass/Weight - Troy Ounce to Grain}
\usage{
unitconversion.troy.ounce.to.grain(x = 1)

unitconversion.ozt.to.gr(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of troy ounces}
}
\value{
x, but converted to grains
}
\description{
Performs a conversion of weights from troy ounces to grains.
}
