% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.troy.ounce.to.ounce.R
\name{unitconversion.troy.ounce.to.ounce}
\alias{unitconversion.troy.ounce.to.ounce}
\alias{unitconversion.ozt.to.oz}
\title{Unit Conversion - Mass/Weight - Troy Ounce to U.S. Ounce}
\usage{
unitconversion.troy.ounce.to.ounce(x = 1)

unitconversion.ozt.to.oz(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of troy ounces}
}
\value{
x, but converted to ounces (avoirdupois)
}
\description{
Performs a conversion of weights from troy ounces to ounces (avoirdupois).
}
