% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.troy.ounce.to.pennyweight.R
\name{unitconversion.troy.ounce.to.pennyweight}
\alias{unitconversion.troy.ounce.to.pennyweight}
\alias{unitconversion.ozt.to.dwt}
\title{Unit Conversion - Mass/Weight - Troy Ounce to Troy Pennyweight}
\usage{
unitconversion.troy.ounce.to.pennyweight(x = 1)

unitconversion.ozt.to.dwt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of troy ounces}
}
\value{
x, but converted to troy pennyweights
}
\description{
Performs a conversion of weights from troy ounces to troy pennyweights.
}
