% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.troy.ounce.to.troy.pound.R
\name{unitconversion.troy.ounce.to.troy.pound}
\alias{unitconversion.troy.ounce.to.troy.pound}
\alias{unitconversion.ozt.to.lbt}
\title{Unit Conversion - Mass/Weight - Troy Ounce to Troy Pound}
\usage{
unitconversion.troy.ounce.to.troy.pound(x = 1)

unitconversion.ozt.to.lbt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of troy ounces}
}
\value{
x, but converted to troy pounds
}
\description{
Performs a conversion of weights from troy ounces to troy pounds.
}
