% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.troy.pound.to.grain.R
\name{unitconversion.troy.pound.to.grain}
\alias{unitconversion.troy.pound.to.grain}
\alias{unitconversion.lbt.to.gr}
\title{Unit Conversion - Mass/Weight - Troy Pound to Grain}
\usage{
unitconversion.troy.pound.to.grain(x = 1)

unitconversion.lbt.to.gr(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pounds}
}
\value{
x, but converted to grains
}
\description{
Performs a conversion of weights from pounds (Troy) to grains.
}
