% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.troy.pound.to.troy.ounce.R
\name{unitconversion.troy.pound.to.troy.ounce}
\alias{unitconversion.troy.pound.to.troy.ounce}
\alias{unitconversion.lbt.to.ozt}
\title{Unit Conversion - Mass/Weight - Troy Pound to Troy Ounce}
\usage{
unitconversion.troy.pound.to.troy.ounce(x = 1)

unitconversion.lbt.to.ozt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of troy pounds}
}
\value{
x, but converted to troy ounces
}
\description{
Performs a conversion of weights from troy pounds to troy ounces.
}
