% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.yard.to.meter.R
\name{unitconversion.yard.to.meter}
\alias{unitconversion.yard.to.meter}
\alias{unitconversion.yd.to.m}
\title{Unit Conversion - Length - Yard to Meter}
\usage{
unitconversion.yard.to.meter(x = 1)

unitconversion.yd.to.m(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of yards}
}
\value{
x, but converted to meters
}
\description{
Performs a conversion of lengths from yards to meters.
}
\references{
NIST. Approximate Conversions from U.S. Customary Measures to Metric. 2022. Accessed 3/16/2022. 
https://www.nist.gov/pml/weights-and-measures/approximate-conversions-us-customary-measures-metric
}
