% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.extract.named.value.R
\name{utility.extract.named.value}
\alias{utility.extract.named.value}
\title{Extract Value By Name}
\usage{
utility.extract.named.value(obj = c(a = 1, b = 2, c = 3), name = "b")
}
\arguments{
\item{obj}{Vector - values}

\item{name}{Character - name of value to extract}
}
\value{
x, but converted to meters
}
\description{
Extracts a value based on name from a named vector.
}
