% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.htest.alternative.R
\name{validate.htest.alternative}
\alias{validate.htest.alternative}
\title{Validation - alternative Parameter for Hypothesis Tests}
\usage{
validate.htest.alternative(
  alternative,
  allowed = c("two.sided", "less", "greater")
)
}
\arguments{
\item{alternative}{Scalar/character - The value to test}

\item{allowed}{Vector/character - Possible values}
}
\value{
no return value
}
\description{
Verifies that the alternative hypothesis argument matches a set of correct values.
If nonconforming, then stop() is called
}
