% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt.sums.R
\name{xt.sums}
\alias{xt.sums}
\title{Matrix - Row and Column Sums}
\usage{
xt.sums(x)
}
\arguments{
\item{x}{Matrix - matrix to calculate row sums, column sums, and total sum.}
}
\value{
x with additional row and column for marginal sums.
}
\description{
Calculate row sums, column sums, and total sum.
}
