% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.test.onesample.R, R/z.test.onesample.simple.R
\name{z.test.onesample}
\alias{z.test.onesample}
\alias{z.test.onesample.simple}
\title{One Sample Z Test for Mean}
\usage{
z.test.onesample(
  x,
  known.population.variance,
  null.hypothesis.mean,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  finite.population.N = NA
)

z.test.onesample.simple(
  sample.mean,
  known.population.variance = 1,
  sample.size,
  null.hypothesis.mean = 0,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  finite.population.N = NA
)
}
\arguments{
\item{x}{Vector - sample values to be used for mean calculation.}

\item{known.population.variance}{Scalar - population variance.}

\item{null.hypothesis.mean}{Scalar - null hypothesis mean value, sample.mean is tested against this value.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}

\item{finite.population.N}{Scalar - the population size if finite population adjustment needs to be made.}

\item{sample.mean}{Scalar - sample mean to be tested.}

\item{sample.size}{Scalar - sample size.}
}
\value{
The results of the statistical test.
}
\description{
Calculate one sample z test for mean.
}
