\name{lolcat-package}
\alias{lolcat-package}
\alias{lolcat}
\docType{package}
\title{
\packageTitle{lolcat}
}
\description{
\packageDescription{lolcat}
}
\details{

The public release of the lolcat statistical package. See \url{http://mikemstech.blogspot.com/2016/03/the-lolcat-statistical-package-public.html}
}
\author{
\packageAuthor{lolcat}

Maintainer: \packageMaintainer{lolcat}
}
