% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.test.twosample.exact.simple.R
\name{proportion.test.twosample.exact.simple}
\alias{proportion.test.twosample.exact.simple}
\title{Two Sample Proportion Test (Exact)}
\usage{
proportion.test.twosample.exact.simple(
  sample.proportion.g1 = NA,
  sample.size.g1,
  np.g1 = sample.proportion.g1 * sample.size.g1,
  sample.proportion.g2 = NA,
  sample.size.g2,
  np.g2 = sample.proportion.g2 * sample.size.g2,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{sample.proportion.g1}{Scalar/numeric - Group 1 - sample proportion between 0 and 1}

\item{sample.size.g1}{Scalar/numeric - Group 1 - sample size.}

\item{np.g1}{Scalar/numeric - sample size multipled by sample proportion for group 1. Required if sample.proportion not specified.}

\item{sample.proportion.g2}{Scalar/numeric - Group 2 - sample proportion between 0 and 1}

\item{sample.size.g2}{Scalar/numeric - Group 2 - sample size.}

\item{np.g2}{Scalar/numeric - sample size multipled by sample proportion for group 2. Required if sample.proportion not specified.}

\item{alternative}{The alternative hypothesis to use for the test computation.}

\item{conf.level}{The confidence level for this test, between 0 and 1.}
}
\value{
Hypothesis test result showing results of test.
}
\description{
Calculates a two-sample proportion test to determine if a samples  
come from different populations. Fisher Exact test is used for test computation.
}
