% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.constant.calculation.D4.uppercase.R,
%   R/spc.constant.calculation.d4.lowercase.R
\name{spc.constant.calculation.D4}
\alias{spc.constant.calculation.D4}
\alias{spc.constant.calculation.d4}
\title{Control Chart Constant d4 / D4}
\usage{
spc.constant.calculation.D4(sample.size, n.sigma = 3)

spc.constant.calculation.d4(sample.size)
}
\arguments{
\item{sample.size}{A scalar or vector of sample sizes}

\item{n.sigma}{How many standard errors to use (default 3 for most control charts)}
}
\value{
A scalar or vector with computed constants.
}
\description{
Calculate D4 control chart constant from Wheeler's SPC books. These are in the same help file 
because of a quirk of how R generates help file names, but they are different constants with different uses.
}
