% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.dry.pint.to.dry.quart.R
\name{unitconversion.dry.pint.to.dry.quart}
\alias{unitconversion.dry.pint.to.dry.quart}
\alias{unitconversion.dry.pt.to.dry.qt}
\title{Unit Conversion - Dry Volume - Dry Pint to Dry Quart}
\usage{
unitconversion.dry.pint.to.dry.quart(x = 1)

unitconversion.dry.pt.to.dry.qt(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pints}
}
\value{
x, but converted to quarts
}
\description{
Performs a conversion of volumes from dry pints to dry quarts.
}
