% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.dry.quart.to.peck.R
\name{unitconversion.dry.quart.to.peck}
\alias{unitconversion.dry.quart.to.peck}
\alias{unitconversion.dry.qt.to.pk}
\title{Unit Conversion - Dry Volume - Dry Quart to Peck}
\usage{
unitconversion.dry.quart.to.peck(x = 1)

unitconversion.dry.qt.to.pk(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of quarts}
}
\value{
x, but converted to pecks
}
\description{
Performs a conversion of volumes from dry quarts to pecks.
}
