% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.mile.to.kilometer.R
\name{unitconversion.mile.to.kilometer}
\alias{unitconversion.mile.to.kilometer}
\alias{unitconversion.mi.to.km}
\title{Unit Conversion - Length - Mile to Kilometer}
\usage{
unitconversion.mile.to.kilometer(x = 1)

unitconversion.mi.to.km(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of miles}
}
\value{
x, but converted to kilometers
}
\description{
Performs a conversion of lengths from miles to kilometers.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
