% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.peck.to.bushel.R
\name{unitconversion.peck.to.bushel}
\alias{unitconversion.peck.to.bushel}
\alias{unitconversion.pk.to.bu}
\title{Unit Conversion - Dry Volume - Peck to Bushel}
\usage{
unitconversion.peck.to.bushel(x = 1)

unitconversion.pk.to.bu(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pecks}
}
\value{
x, but converted to bushels
}
\description{
Performs a conversion of volumes from pecks to bushels.
}
