% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.pound.to.stone.R
\name{unitconversion.pound.to.stone}
\alias{unitconversion.pound.to.stone}
\alias{unitconversion.lb.to.st}
\title{Unit Conversion - Mass/Weight - U.S. Pound to Stone}
\usage{
unitconversion.pound.to.stone(x = 1)

unitconversion.lb.to.st(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of pounds}
}
\value{
x, but converted to stone
}
\description{
Performs a conversion of weights from pounds (avoirdupois) to stones.
}
\references{
NIST. Handbook 133 - Checking the Net Contents of Packaged Goods - Appendix E -  General Tables of Units of Measurement. 2020.  
https://www.nist.gov/system/files/documents/2019/12/03/00-20-h133-apdxE_final-17.pdf
}
