% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconversion.quart.to.gallon.R
\name{unitconversion.quart.to.gallon}
\alias{unitconversion.quart.to.gallon}
\alias{unitconversion.qt.to.gal}
\title{Unit Conversion - Liquid Volume - Quart to Gallon}
\usage{
unitconversion.quart.to.gallon(x = 1)

unitconversion.qt.to.gal(x = 1)
}
\arguments{
\item{x}{Vector - Values in units of quarts}
}
\value{
x, but converted to gallons
}
\description{
Performs a conversion of volumes from quarts to gallons.
}
